+++ .gitignore
@@ -1,4 +1,5 @@
 COSX
+dist
 .DS_Store
