--- README.md
@@ -1,5 +1,0 @@
-# New-Tongwentang
-Firefox Add-on, converts web page between Simplified and Traditional Chinese.
-
-# License
-Released under [MIT/X11 License](https://opensource.org/licenses/mit-license.php)
