--- lib/tongwen/tongwen_table_ps2t.js
@@ -1,7052 +1,0 @@
-//TongWen.addS2TTable(
-var default_PS2T_table = {
-"0余":"0餘",
-"0天":"0天",
-"0天后":"0天後",
-"0寸":"0吋",
-"0海":"0海",
-"0海裡":"0海里",
-"0裡":"0里",
-"1天":"1天",
-"1天后":"1天後",
-"1寸":"1吋",
-"1海":"1海",
-"1海裡":"1海里",
-"1裡":"1里",
-"2天":"2天",
-"2天后":"2天後",
-"2寸":"2吋",
-"2海":"2海",
-"2海裡":"2海里",
-"2裡":"2里",
-"3天":"3天",
-"3天后":"3天後",
-"3寸":"3吋",
-"3海":"3海",
-"3海裡":"3海里",
-"3裡":"3里",
-"4天":"4天",
-"4天后":"4天後",
-"4寸":"4吋",
-"4海":"4海",
-"4海裡":"4海里",
-"4裡":"4里",
-"5天":"5天",
-"5天后":"5天後",
-"5寸":"5吋",
-"5海":"5海",
-"5海裡":"5海里",
-"5裡":"5里",
-"6天":"6天",
-"6天后":"6天後",
-"6寸":"6吋",
-"6海":"6海",
-"6海裡":"6海里",
-"6裡":"6里",
-"7天":"7天",
-"7天后":"7天後",
-"7寸":"7吋",
-"7海":"7海",
-"7海裡":"7海里",
-"7裡":"7里",
-"8天":"8天",
-"8天后":"8天後",
-"8寸":"8吋",
-"8海":"8海",
-"8海裡":"8海里",
-"8裡":"8里",
-"9天":"9天",
-"9天后":"9天後",
-"9寸":"9吋",
-"9海":"9海",
-"9海裡":"9海里",
-"9裡":"9里",
-"A』":"A'",
-"B』":"B'",
-"C』":"C'",
-"D』":"D'",
-"E』":"E'",
-"F』":"F'",
-"G』":"G'",
-"H』":"H'",
-"I』":"I'",
-"J』":"J'",
-"K』":"K'",
-"L』":"L'",
-"M』":"M'",
-"N』":"N'",
-"O』":"O'",
-"P』":"P'",
-"Q』":"Q'",
-"R』":"R'",
-"S』":"S'",
-"T』":"T'",
-"U』":"U'",
-"V』":"V'",
-"W』":"W'",
-"X』":"X'",
-"Y』":"Y'",
-"Z』":"Z'",
-"a』":"a'",
-"b』":"b'",
-"c』":"c'",
-"d』":"d'",
-"e』":"e'",
-"f』":"f'",
-"g』":"g'",
-"h』":"h'",
-"i』":"i'",
-"j』":"j'",
-"k』":"k'",
-"l』":"l'",
-"m』":"m'",
-"n』":"n'",
-"o』":"o'",
-"p』":"p'",
-"q』":"q'",
-"r』":"r'",
-"s』":"s'",
-"t』":"t'",
-"u』":"u'",
-"v』":"v'",
-"w』":"w'",
-"x』":"x'",
-"y』":"y'",
-"z』":"z'",
-"○余":"○餘",
-"○海":"○海",
-"○海裡":"○海里",
-"○裡":"○里",
-"「恩":"「嗯",
-"「恩！":"「嗯！",
-"「當":"「當",
-"「當」":"「噹」",
-"「當！":"「噹！",
-"一並":"一併",
-"一並不":"一並不",
-"一並且":"一並且",
-"一並非":"一並非",
-"一伙":"一夥",
-"一個":"一個",
-"一出":"一出",
-"一出京劇":"一齣京劇",
-"一出劇":"一齣劇",
-"一出劇院":"一出劇院",
-"一出好戲":"一齣好戲",
-"一出戲":"一齣戲",
-"一出話劇":"一齣話劇",
-"一出電影":"一齣電影",
-"一制":"一制",
-"一制作":"一製作",
-"一制程":"一製程",
-"一制藥":"一製藥",
-"一只":"一隻",
-"一只只":"一隻隻",
-"一只只夠":"一隻只夠",
-"一只只有":"一隻只有",
-"一周":"一週",
-"一大":"一大",
-"一大伙":"一大夥",
-"一天":"一天",
-"一天后":"一天後",
-"一干":"一干",
-"一干二淨":"一乾二淨",
-"一干就":"一幹就",
-"一干而盡":"一乾而盡",
-"一斗":"一斗",
-"一斗不":"一鬥不",
-"一斗再斗":"一鬥再鬥",
-"一松":"一鬆",
-"一松林":"一松林",
-"一松樹":"一松樹",
-"一松鼠":"一松鼠",
-"一樹":"一樹",
-"一樹百獲":"一樹百穫",
-"一毫":"一毫",
-"一毫一發":"一毫一髮",
-"一沖":"一沖",
-"一沖一撞":"一衝一撞",
-"一海":"一海",
-"一海裡":"一海里",
-"一涂":"一塗",
-"一游":"一遊",
-"一物":"一物",
-"一物克一":"一物剋一",
-"一發":"一發",
-"一發之差":"一髮之差",
-"一發之間":"一髮之間",
-"一發千鈞":"一髮千鈞",
-"一直":"一直",
-"一直干":"一直幹",
-"一直發女":"一直髮女",
-"一箭":"一箭",
-"一箭雙雕":"一箭雙鵰",
-"一般":"一般",
-"一般只":"一般只",
-"一裡":"一里",
-"一裡余":"一里餘",
-"一贊":"一讚",
-"一鍋":"一鍋",
-"一鍋面":"一鍋麵",
-"一面":"一面",
-"一點":"一點",
-"一點丑":"一點醜",
-"丁當":"叮噹",
-"丁鈴":"叮鈴",
-"七只":"七隻",
-"七天":"七天",
-"七天后":"七天後",
-"七海":"七海",
-"七海裡":"七海里",
-"七海裡面":"七海裡面",
-"七裡":"七里",
-"七裡余":"七里餘",
-"七面":"七面",
-"丈余":"丈餘",
-"三個":"三個",
-"三只":"三隻",
-"三周":"三週",
-"三嘆":"三歎",
-"三國":"三國",
-"三國志":"三國志",
-"三天":"三天",
-"三海":"三海",
-"三海裡":"三海里",
-"三裡":"三里",
-"三裡余":"三里餘",
-"三豐":"三丰",
-"上准":"上準",
-"上周":"上週",
-"上布":"上佈",
-"上折":"上摺",
-"上梁":"上樑",
-"上梁山":"上梁山",
-"上沖":"上衝",
-"上沖刷":"上沖刷",
-"上沖水":"上沖水",
-"上沖洗":"上沖洗",
-"上沖涼":"上沖涼",
-"上沖澡":"上沖澡",
-"上簽":"上籤",
-"下台":"下台",
-"下咽":"下嚥",
-"下擺":"下襬",
-"下梁":"下樑",
-"下沖":"下衝",
-"下簽":"下籤",
-"下裡":"下里",
-"不丑":"不醜",
-"不並":"不併",
-"不余":"不餘",
-"不修":"不修",
-"不修篇幅":"不修邊幅",
-"不值":"不值",
-"不值一曬":"不值一哂",
-"不大":"不大",
-"不大象":"不大像",
-"不寒":"不寒",
-"不寒而栗":"不寒而慄",
-"不干":"不干",
-"不干不淨":"不乾不淨",
-"不干他":"不干他",
-"不干休":"不干休",
-"不干你":"不干你",
-"不干她":"不干她",
-"不干我":"不干我",
-"不干擾":"不干擾",
-"不干政":"不干政",
-"不干杯":"不乾杯",
-"不干涉":"不干涉",
-"不干淨":"不乾淨",
-"不干系":"不干係",
-"不干脆":"不乾脆",
-"不干著急":"不乾著急",
-"不干裂":"不乾裂",
-"不干預":"不干預",
-"不愈":"不癒",
-"不斗":"不鬥",
-"不斷":"不斷",
-"不斷發展":"不斷發展",
-"不方":"不方",
-"不方便":"不方便",
-"不松":"不松",
-"不松下":"不鬆下",
-"不涂":"不塗",
-"不當":"不當",
-"不當准":"不當準",
-"不系":"不繫",
-"不能":"不能",
-"不能干":"不能干",
-"不能象":"不能象",
-"不能象征":"不能象徵",
-"不舍":"不捨",
-"不要":"不要",
-"不要干":"不要干",
-"不要象":"不要象",
-"不象":"不像",
-"不贊":"不讚",
-"不贊同":"不贊同",
-"不采":"不採",
-"丑一":"丑一",
-"丑一點":"醜一點",
-"丑不":"丑不",
-"丑不丑":"醜不醜",
-"丑丑":"醜醜",
-"丑丫":"丑丫",
-"丑丫頭":"醜丫頭",
-"丑丫鬟":"醜丫鬟",
-"丑也":"丑也",
-"丑也丑":"醜也醜",
-"丑事":"醜事",
-"丑人":"醜人",
-"丑似":"醜似",
-"丑俊":"醜俊",
-"丑八":"丑八",
-"丑八怪":"醜八怪",
-"丑八戒":"醜八戒",
-"丑劇":"醜劇",
-"丑化":"醜化",
-"丑名":"醜名",
-"丑女":"醜女",
-"丑如":"醜如",
-"丑婆":"醜婆",
-"丑婦":"醜婦",
-"丑媳":"丑媳",
-"丑媳婦":"醜媳婦",
-"丑家":"丑家",
-"丑家伙":"醜傢伙",
-"丑小":"丑小",
-"丑小子":"醜小子",
-"丑小鴨":"醜小鴨",
-"丑少":"丑少",
-"丑少年":"醜少年",
-"丑巴":"丑巴",
-"丑巴怪":"醜巴怪",
-"丑得":"醜得",
-"丑得象":"醜得像",
-"丑怪":"醜怪",
-"丑惡":"醜惡",
-"丑態":"醜態",
-"丑東":"丑東",
-"丑東西":"醜東西",
-"丑模":"丑模",
-"丑模樣":"醜模樣",
-"丑樣":"醜樣",
-"丑死":"醜死",
-"丑漢":"醜漢",
-"丑煞":"丑煞",
-"丑煞人":"醜煞人",
-"丑男":"醜男",
-"丑的":"醜的",
-"丑老":"醜老",
-"丑老頭":"醜老頭",
-"丑聞":"醜聞",
-"丑聲":"醜聲",
-"丑臉":"醜臉",
-"丑行":"醜行",
-"丑詆":"醜詆",
-"丑話":"醜話",
-"丑語":"醜語",
-"丑貌":"醜貌",
-"丑陋":"醜陋",
-"丑面":"醜面",
-"丑類":"醜類",
-"丑鬼":"醜鬼",
-"且斗":"且斗",
-"且斗且":"且鬥且",
-"世界":"世界",
-"世界杯":"世界盃",
-"丟丑":"丟醜",
-"丟了":"丟了",
-"丟了丑":"丟了醜",
-"並不":"並不",
-"並不並":"併不併",
-"並入":"併入",
-"並力":"併力",
-"並包":"并包",
-"並合":"併合",
-"並吞":"併吞",
-"並圖":"併圖",
-"並攏":"併攏",
-"並案":"併案",
-"並疊":"併疊",
-"並發":"並發",
-"並發症":"併發症",
-"並砌":"併砌",
-"並科":"併科",
-"並肩":"並肩",
-"並肩子":"併肩子",
-"並肩干":"並肩幹",
-"並負":"併負",
-"並負責":"並負責",
-"並購":"併購",
-"並購買":"並購買",
-"並贓":"併贓",
-"並贓拿賊":"併贓拿賊",
-"並贓治罪":"併贓治罪",
-"並采":"並採",
-"並除":"併除",
-"並隴":"併攏",
-"中侖":"中崙",
-"中岳":"中嶽",
-"中布":"中佈",
-"中簽":"中籤",
-"主干":"主幹",
-"久斗":"久鬥",
-"之余":"之餘",
-"之准":"之準",
-"之欲":"之慾",
-"之游":"之遊",
-"九只":"九隻",
-"九嘆":"九歎",
-"九天":"九天",
-"九天后":"九天後",
-"九海":"九海",
-"九海裡":"九海里",
-"九裡":"九里",
-"九裡余":"九里餘",
-"九面":"九面",
-"也沒":"也沒",
-"也沒干":"也沒幹",
-"也舍":"也舍",
-"也舍下":"也捨下",
-"也舍棄":"也捨棄",
-"也象":"也像",
-"亂斗":"亂鬥",
-"亂沖":"亂衝",
-"亂涂":"亂塗",
-"了准":"了准",
-"了准備":"了準備",
-"了准時":"了準時",
-"了准有":"了準有",
-"了准的":"了準的",
-"了准確":"了準確",
-"了如":"了如",
-"了如指掌":"瞭如指掌",
-"了望":"瞭望",
-"了然":"瞭然",
-"了然於胸":"了然於胸",
-"了若":"了若",
-"了若指掌":"瞭若指掌",
-"了解":"瞭解",
-"事也":"事也",
-"事也干":"事也幹",
-"事可":"事可",
-"事可干":"事可幹",
-"事好":"事好",
-"事好干":"事好幹",
-"事干":"事幹",
-"事干擾":"事干擾",
-"事干涉":"事干涉",
-"事干淨":"事乾淨",
-"事干系":"事幹係",
-"事干脆":"事乾脆",
-"事干預":"事干預",
-"事情":"事情",
-"事情可干":"事情可幹",
-"事情好干":"事情好幹",
-"事情干":"事情幹",
-"事跡":"事蹟",
-"事都":"事都",
-"事都干":"事都幹",
-"事關":"事關",
-"二個":"二個",
-"二只":"二隻",
-"二天":"二天",
-"二天后":"二天後",
-"二手":"二手",
-"二手煙":"二手菸",
-"二撇":"二撇",
-"二撇胡":"二撇鬍",
-"二海":"二海",
-"二海裡":"二海里",
-"二裡":"二里",
-"二面":"二面",
-"互斗":"互鬥",
-"五只":"五隻",
-"五周":"五週",
-"五天":"五天",
-"五岳":"五嶽",
-"五海":"五海",
-"五海裡":"五海里",
-"五裡":"五里",
-"五裡余":"五里餘",
-"五谷":"五穀",
-"五髒":"五臟",
-"些丑":"些醜",
-"些干":"些干",
-"些干頭":"些幹頭",
-"些象":"些像",
-"亞裡":"亞里",
-"交托":"交託",
-"交游":"交遊",
-"亦舍":"亦舍",
-"亦舍下":"亦捨下",
-"亦雲":"亦云",
-"京沈":"京瀋",
-"亮丑":"亮醜",
-"人丑":"人醜",
-"人像":"人像",
-"人像征":"人象徵",
-"人干":"人幹",
-"人欲":"人欲",
-"人欲橫流":"人慾橫流",
-"人游":"人遊",
-"人物":"人物",
-"人物志":"人物誌",
-"人象":"人像",
-"人贊":"人讚",
-"人雲":"人云",
-"人雲亦雲":"人云亦云",
-"人雲游":"人云遊",
-"人雲集":"人雲集",
-"什幺":"什麼",
-"什幺女人":"什麼女人",
-"什錦":"什錦",
-"什錦面":"什錦麵",
-"什麼":"什麼",
-"介系":"介系",
-"介系詞":"介係詞",
-"介胄":"介冑",
-"仍斗":"仍鬥",
-"仍象":"仍像",
-"他不":"他不",
-"他不干渴":"他不乾渴",
-"他丑":"他醜",
-"他干":"他幹",
-"他干一杯":"他乾一杯",
-"他干兒子":"他乾兒子",
-"他干女兒":"他乾女兒",
-"他干擾":"他干擾",
-"他干涉":"他干涉",
-"他干淨":"他乾淨",
-"他干爹":"他乾爹",
-"他干脆":"他乾脆",
-"他干預":"他干預",
-"他白":"他白",
-"他白干":"他白幹",
-"他象":"他像",
-"他象征戰":"他像征戰",
-"付托":"付託",
-"仙後":"仙后",
-"仙游":"仙遊",
-"仙范":"仙範",
-"代征":"代徵",
-"代理":"代理",
-"令媛":"令嬡",
-"仰沖":"仰衝",
-"仿佛":"彷彿",
-"仿制":"仿製",
-"仿徉":"彷徉",
-"仿徨":"徬徨",
-"仿效":"倣傚",
-"休克":"休克",
-"伙伴":"夥伴",
-"伙同":"夥同",
-"伙夫":"伙伕",
-"伙眾":"夥眾",
-"伙計":"夥計",
-"估不":"估不",
-"估不准":"估不準",
-"伴伙":"伴夥",
-"伴游":"伴遊",
-"伶敏":"靈敏",
-"伸卷":"伸捲",
-"伺奉":"侍奉",
-"似松":"似松",
-"似松實緊":"似鬆實緊",
-"但雲":"但云",
-"低蕩":"低盪",
-"低迥":"低迴",
-"住扎":"住紮",
-"佔人":"占人",
-"佔候":"占候",
-"佔卜":"占卜",
-"佔卦":"占卦",
-"佔士":"佔士",
-"佔士邦":"占士邦",
-"佔夢":"占夢",
-"佔射":"占射",
-"佔拜":"占拜",
-"佔斷":"占斷",
-"佔星":"占星",
-"佔有":"佔有",
-"佔有欲":"佔有慾",
-"佔親":"占親",
-"佔課":"占課",
-"佔身":"占身",
-"佔風":"佔風",
-"佔風使帆":"占風使帆",
-"佔驗":"占驗",
-"佔鳳":"占鳳",
-"余丈":"餘丈",
-"余下":"餘下",
-"余事":"餘事",
-"余人":"餘人",
-"余位":"餘位",
-"余低":"餘低",
-"余僇":"餘僇",
-"余僧":"餘僧",
-"余光":"餘光",
-"余光中":"余光中",
-"余刃":"餘刃",
-"余利":"餘利",
-"余剩":"餘剩",
-"余力":"餘力",
-"余勁":"餘勁",
-"余勇":"餘勇",
-"余勢":"餘勢",
-"余匪":"餘匪",
-"余匹":"餘匹",
-"余只":"餘隻",
-"余名":"餘名",
-"余味":"餘味",
-"余地":"餘地",
-"余天":"餘天",
-"余姚":"餘姚",
-"余威":"餘威",
-"余子":"餘子",
-"余子碌碌":"餘子碌碌",
-"余字":"餘字",
-"余存":"餘存",
-"余孬":"餘孬",
-"余孽":"餘孽",
-"余寸":"餘寸",
-"余尺":"餘尺",
-"余年":"餘年",
-"余座":"餘座",
-"余廕":"餘廕",
-"余弦":"餘弦",
-"余怒":"餘怒",
-"余恨":"餘恨",
-"余息":"餘息",
-"余悸":"餘悸",
-"余情":"餘情",
-"余慶":"餘慶",
-"余憤":"餘憤",
-"余數":"餘數",
-"余日":"餘日",
-"余暇":"餘暇",
-"余暉":"餘暉",
-"余月":"餘月",
-"余杭":"餘杭",
-"余枝":"餘枝",
-"余桶":"餘桶",
-"余次":"餘次",
-"余款":"餘款",
-"余歲":"餘歲",
-"余殃":"餘殃",
-"余毒":"餘毒",
-"余氣":"餘氣",
-"余水":"餘水",
-"余波":"餘波",
-"余海":"余海",
-"余海裡":"餘海里",
-"余溫":"餘溫",
-"余漏":"餘漏",
-"余瀝":"餘瀝",
-"余烈":"餘烈",
-"余熱":"餘熱",
-"余燼":"餘燼",
-"余物":"餘物",
-"余生":"餘生",
-"余畝":"餘畝",
-"余眾":"餘眾",
-"余禍":"餘禍",
-"余糧":"餘糧",
-"余線":"餘線",
-"余缺":"餘缺",
-"余羹":"餘羹",
-"余者":"餘者",
-"余脈":"餘脈",
-"余興":"餘興",
-"余蔭":"餘蔭",
-"余裕":"餘裕",
-"余裡":"餘裡",
-"余責":"餘責",
-"余賊":"餘賊",
-"余路":"餘路",
-"余載":"餘載",
-"余輝":"餘輝",
-"余辜":"餘辜",
-"余部":"餘部",
-"余錢":"餘錢",
-"余閒":"餘閒",
-"余震":"餘震",
-"余音":"餘音",
-"余韻":"餘韻",
-"余額":"餘額",
-"余風":"餘風",
-"余飯":"餘飯",
-"余香":"餘香",
-"余騎":"餘騎",
-"余點":"餘點",
-"余黨":"餘黨",
-"佛歷":"佛曆",
-"佛象":"佛像",
-"作准":"作準",
-"作奸":"作奸",
-"作奸犯科":"作姦犯科",
-"作崇":"作祟",
-"作廢":"作廢",
-"作廢後":"作廢後",
-"作的":"作的",
-"作的准":"作的準",
-"你丑":"你醜",
-"你干":"你幹",
-"你干一杯":"你乾一杯",
-"你干兒子":"你乾兒子",
-"你干女兒":"你乾女兒",
-"你干擾":"你干擾",
-"你干涉":"你干涉",
-"你干爹":"你乾爹",
-"你干脆":"你乾脆",
-"你干那杯":"你乾那杯",
-"你干預":"你干預",
-"你白":"你白",
-"你白干":"你白幹",
-"你象":"你像",
-"佩環":"珮環",
-"佩飾":"珮飾",
-"佳肴":"佳餚",
-"佳裡":"佳里",
-"使心":"使心",
-"使心作幸":"使心作倖",
-"來干":"來幹",
-"來干擾":"來干擾",
-"來干旱":"來乾旱",
-"來干杯":"來乾杯",
-"來干枯":"來乾枯",
-"來干涉":"來干涉",
-"來干淨":"來乾淨",
-"來干爽":"來乾爽",
-"來干脆":"來乾脆",
-"來干裂":"來乾裂",
-"來干預":"來干預",
-"來范":"來範",
-"來象":"來像",
-"例范":"例範",
-"侏羅":"侏儸",
-"侖背":"崙背",
-"供制":"供製",
-"依准":"依準",
-"依托":"依託",
-"侵並":"侵併",
-"侵范":"侵範",
-"便吃":"便吃",
-"便吃干":"便吃乾",
-"便松":"便鬆",
-"便象":"便像",
-"俊丑":"俊醜",
-"保不":"保不",
-"保不准":"保不準",
-"保准":"保準",
-"信托":"信託",
-"信緘":"信箋",
-"信而":"信而",
-"信而有征":"信而有徵",
-"修浚":"修濬",
-"修胡":"修胡",
-"修胡刀":"修鬍刀",
-"俯沖":"俯衝",
-"俱象":"俱像",
-"個象":"個像",
-"倍余":"倍餘",
-"倏松":"倏鬆",
-"們干":"們幹",
-"們干兒子":"們干兒子",
-"們干女兒":"們干女兒",
-"們干擾":"們干擾",
-"們干涉":"們干涉",
-"們干淨":"們干淨",
-"們干爹":"們干爹",
-"們干脆":"們干脆",
-"們干預":"們干預",
-"倒僕":"倒仆",
-"倒卷":"倒捲",
-"倒象":"倒像",
-"倒黴":"倒霉",
-"倛丑":"倛醜",
-"借以":"藉以",
-"借口":"藉口",
-"借托":"借託",
-"借故":"藉故",
-"借機":"藉機",
-"借著":"藉著",
-"借詞":"藉詞",
-"倦游":"倦遊",
-"值得":"值得",
-"值得一干":"值得一幹",
-"值得干":"值得幹",
-"假發":"假髮",
-"做了":"做了",
-"做出":"做出",
-"做出好戲":"做齣好戲",
-"做到":"做到",
-"做到象":"做到像",
-"停制":"停製",
-"停征":"停徵",
-"側沖":"側衝",
-"偶象":"偶像",
-"偷梁":"偷樑",
-"傅裡":"傅裡",
-"傅裡葉":"傅里葉",
-"傍徨":"徬徨",
-"傑倫":"杰倫",
-"傒幸":"傒倖",
-"備注":"備註",
-"傭金":"佣金",
-"傭錢":"佣錢",
-"傳布":"傳佈",
-"傷愈":"傷癒",
-"傷難":"傷難",
-"傷難愈":"傷難癒",
-"傾刻":"頃刻",
-"傾復":"傾覆",
-"僂羅":"僂儸",
-"僅余":"僅餘",
-"像征":"象徵",
-"像贊":"像贊",
-"像贊他":"像讚他",
-"像贊你":"像讚你",
-"像贊嘆":"像讚嘆",
-"像贊她":"像讚她",
-"像贊我":"像讚我",
-"像贊揚":"像讚揚",
-"像贊美":"像讚美",
-"像贊許":"像讚許",
-"像贊譽":"像讚譽",
-"像贊賞":"像讚賞",
-"像贊頌":"像讚頌",
-"僕倒":"仆倒",
-"僕地":"仆地",
-"僕然":"仆然",
-"僕街":"仆街",
-"僥天":"僥天",
-"僥天之幸":"僥天之悻",
-"僥幸":"僥倖",
-"僥悻":"僥倖",
-"僵僕":"僵仆",
-"僵屍":"殭屍",
-"僵繩":"韁繩",
-"僵蠶":"殭蠶",
-"儀范":"儀範",
-"億余":"億餘",
-"儆效":"儆傚",
-"儉朴":"儉樸",
-"儌幸":"儌倖",
-"儒子":"儒子",
-"儒子可教":"孺子可教",
-"優游":"優遊",
-"兀術":"兀朮",
-"元凶":"元兇",
-"元勳":"元勛",
-"充飢":"充飢",
-"充飢畫餅":"充饑畫餅",
-"先奸":"先姦",
-"先干":"先干",
-"先干為敬":"先乾為敬",
-"光念":"光唸",
-"光致":"光致",
-"光致致":"光緻緻",
-"克克":"剋剋",
-"克剜":"剋剜",
-"克夫":"剋夫",
-"克扣":"剋扣",
-"克日":"剋日",
-"克星":"剋星",
-"克期":"剋期",
-"克死":"剋死",
-"克苦":"刻苦",
-"克薄":"剋薄",
-"克裡":"克里",
-"克裡斯多":"克里斯多",
-"克裡斯托":"克里斯托",
-"免征":"免徵",
-"兜卷":"兜捲",
-"入伙":"入夥",
-"內函":"內涵",
-"內制":"內製",
-"內哄":"內鬨",
-"內斗":"內鬥",
-"內松":"內松",
-"內松外緊":"內鬆外緊",
-"內緊":"內緊",
-"內緊外松":"內緊外鬆",
-"內裡":"內裡",
-"內髒":"內臟",
-"全台":"全台",
-"全日":"全日",
-"全日制":"全日制",
-"全裡":"全里",
-"全象":"全像",
-"兩個":"兩個",
-"兩個中":"兩個中",
-"兩制":"兩制",
-"兩只":"兩隻",
-"兩周":"兩週",
-"兩天":"兩天",
-"兩撇":"兩撇",
-"兩撇胡":"兩撇鬍",
-"兩海":"兩海",
-"兩海裡":"兩海里",
-"兩裡":"兩里",
-"兩裡余":"兩里餘",
-"八只":"八隻",
-"八周":"八週",
-"八天":"八天",
-"八天后":"八天後",
-"八字":"八字",
-"八字胡":"八字鬍",
-"八海":"八海",
-"八海裡":"八海里",
-"八裡":"八里",
-"八裡余":"八里餘",
-"公余":"公餘",
-"公厘":"公釐",
-"公布":"公佈",
-"公干":"公幹",
-"公歷":"公曆",
-"公裡":"公里",
-"六只":"六隻",
-"六周":"六週",
-"六天":"六天",
-"六天后":"六天後",
-"六欲":"六慾",
-"六沖":"六衝",
-"六海":"六海",
-"六海裡":"六海里",
-"六裡":"六里",
-"六裡余":"六里餘",
-"六須":"六須",
-"六須鯰":"六鬚鯰",
-"共制":"共製",
-"共御":"共禦",
-"共斗":"共鬥",
-"兵簽":"兵籤",
-"其丑":"其醜",
-"其余":"其餘",
-"典范":"典範",
-"兼並":"兼併",
-"再丑":"再醜",
-"再制":"再製",
-"再卷":"再捲",
-"再干":"再幹",
-"再干一杯":"再乾一杯",
-"再斗":"再鬥",
-"再斗一斗":"再鬥一鬥",
-"再斗斗":"再鬥鬥",
-"再松":"再鬆",
-"再沖":"再衝",
-"再象":"再像",
-"冗余":"冗餘",
-"冬台":"冬颱",
-"冶游":"冶遊",
-"准不":"准不",
-"准不准備":"準不準備",
-"准不准確":"準不準確",
-"准不及":"準不及",
-"准不可":"準不可",
-"准不得":"準不得",
-"准不成":"準不成",
-"准不是":"準不是",
-"准不會":"準不會",
-"准不能":"準不能",
-"准且":"准且",
-"准且勁":"準且勁",
-"准些":"準些",
-"准以":"准以",
-"准以為":"準以為",
-"准倒":"准倒",
-"准倒黴":"準倒霉",
-"准備":"準備",
-"准兒":"準兒",
-"准准":"準準",
-"准到":"準到",
-"准則":"準則",
-"准博":"准博",
-"准博士":"準博士",
-"准叫":"準叫",
-"准叫你":"準叫你",
-"准可":"準可",
-"准司":"准司",
-"准司法":"準司法",
-"准吃":"准吃",
-"准吃虧":"準吃虧",
-"准噶":"准噶",
-"准噶爾":"準噶爾",
-"准外":"准外",
-"准外母":"準外母",
-"准外父":"準外父",
-"准夠":"準夠",
-"准女":"准女",
-"准女婿":"準女婿",
-"准媽":"准媽",
-"准媽媽":"準媽媽",
-"准定":"準定",
-"准平":"准平",
-"准平原":"準平原",
-"准度":"準度",
-"准得":"準得",
-"准成":"準成",
-"准把":"準把",
-"准據":"準據",
-"准擬":"準擬",
-"准新":"准新",
-"准新娘":"準新娘",
-"准新郎":"準新郎",
-"准星":"準星",
-"准是":"準是",
-"准時":"準時",
-"准會":"準會",
-"准有":"準有",
-"准極":"準極",
-"准死":"準死",
-"准決":"準決",
-"准決斗":"准決鬥",
-"准決賽":"準決賽",
-"准沒":"準沒",
-"准爸":"准爸",
-"准爸爸":"準爸爸",
-"准直":"準直",
-"准知":"準知",
-"准確":"準確",
-"准線":"準線",
-"准繩":"準繩",
-"准而":"准而",
-"准而快":"準而快",
-"准能":"準能",
-"准被":"準被",
-"准要":"準要",
-"准話":"準話",
-"准譜":"准譜",
-"准譜兒":"準譜兒",
-"准錯":"准錯",
-"准錯不了":"準錯不了",
-"准頭":"準頭",
-"准點":"準點",
-"准！":"準！",
-"凜列":"凜洌",
-"凜栗":"凜慄",
-"凝干":"凝乾",
-"凱裡":"凱里",
-"凶嫌":"兇嫌",
-"凶徒":"兇徒",
-"凶懼":"兇懼",
-"凶案":"兇案",
-"凶殺":"兇殺",
-"凶犯":"兇犯",
-"凶猛":"兇猛",
-"出丑":"出醜",
-"出了":"出了",
-"出了丑":"出了醜",
-"出個":"出個",
-"出個丑":"出個醜",
-"出槌":"出鎚",
-"出游":"出遊",
-"出的":"出的",
-"出的丑":"出的醜",
-"出糴":"出糶",
-"出過":"出過",
-"出過丑":"出過醜",
-"出電":"出電",
-"出電影":"齣電影",
-"函容":"涵容",
-"函復":"函覆",
-"函意":"涵意",
-"函數":"函數",
-"函洞":"涵洞",
-"函濡":"涵濡",
-"函煦":"涵煦",
-"函管":"涵管",
-"函義":"涵義",
-"函蓄":"涵蓄",
-"函蓋":"涵蓋",
-"函蘊":"涵蘊",
-"函養":"涵養",
-"刀削":"刀削",
-"刀削面":"刀削麵",
-"分伙":"分夥",
-"分別":"分別",
-"分布":"分佈",
-"分象":"分像",
-"分道":"分道",
-"分道揚鏢":"分道揚鑣",
-"分量":"份量",
-"分鐘":"分鐘",
-"刑余":"刑餘",
-"初愈":"初癒",
-"初斗":"初鬥",
-"初斗然":"初陡然",
-"別干":"別幹",
-"別干擾":"別干擾",
-"別干杯":"別乾杯",
-"別干涉":"別干涉",
-"別干淨":"別乾淨",
-"別干燥":"別乾燥",
-"別干爽":"別乾爽",
-"別干裂":"別乾裂",
-"別干預":"別干預",
-"別扭":"彆扭",
-"別扭來":"別扭來",
-"別拗":"彆拗",
-"別斗":"別鬥",
-"別氣":"彆氣",
-"別腳":"彆腳",
-"別致":"別緻",
-"刨冰":"鉋冰",
-"刨刀":"鉋刀",
-"刨削":"鉋削",
-"刨子":"鉋子",
-"刨成":"鉋成",
-"刨花":"鉋花",
-"利欲":"利慾",
-"刮倒":"颳倒",
-"刮大":"刮大",
-"刮大風":"颳大風",
-"刮胡":"刮鬍",
-"刮胡刀":"刮鬍刀",
-"刮胡須":"刮鬍鬚",
-"刮起":"颳起",
-"刮須":"刮鬚",
-"刮風":"颳風",
-"制丹":"製丹",
-"制件":"製件",
-"制作":"製作",
-"制做":"製做",
-"制備":"製備",
-"制冰":"製冰",
-"制冷":"製冷",
-"制出":"製出",
-"制劑":"製劑",
-"制品":"製品",
-"制器":"製器",
-"制圖":"製圖",
-"制成":"製成",
-"制播":"製播",
-"制材":"製材",
-"制毒":"製毒",
-"制法":"製法",
-"制為":"製為",
-"制煉":"製煉",
-"制片":"製片",
-"制版":"製版",
-"制程":"製程",
-"制糖":"製糖",
-"制紙":"製紙",
-"制藥":"製藥",
-"制衣":"製衣",
-"制表":"製表",
-"制造":"製造",
-"制醋":"製醋",
-"制面":"製麵",
-"制面具":"製面具",
-"制革":"製革",
-"制鞋":"製鞋",
-"制餡":"製餡",
-"制鹽":"製鹽",
-"刺不":"刺不",
-"刺不准":"刺不準",
-"刺准":"刺準",
-"刻象":"刻像",
-"刻象征服":"刻像征服",
-"剃發":"剃髮",
-"剃須":"剃鬚",
-"則象":"則像",
-"則象征戰":"則像征戰",
-"則象征服":"則像征服",
-"則采":"則採",
-"剉刀":"銼刀",
-"剉工":"銼工",
-"剉平":"銼平",
-"削發":"削髮",
-"剌刀":"刺刀",
-"前僕":"前僕",
-"前僕後":"前仆後",
-"前沖":"前衝",
-"剩余":"剩餘",
-"剩肴":"剩餚",
-"剪彩":"剪綵",
-"剪發":"剪髮",
-"剪除":"翦除",
-"副丑":"副丑",
-"副丑相":"副醜相",
-"割舍":"割捨",
-"創制":"創製",
-"劃不":"劃不",
-"劃不來":"劃不來",
-"劃來":"划來",
-"劃去":"划去",
-"劃單":"劃單",
-"劃單人艇":"划單人艇",
-"劃子":"划子",
-"劃拳":"划拳",
-"劃槳":"划槳",
-"劃步":"划步",
-"劃水":"划水",
-"劃獨":"划獨",
-"劃算":"划算",
-"劃舟":"划舟",
-"劃船":"划船",
-"劃艇":"划艇",
-"劃雙":"劃雙",
-"劃雙人":"划雙人",
-"劃龍":"划龍",
-"劇斗":"劇鬥",
-"劈裡":"劈里",
-"力干":"力幹",
-"力干擾":"力干擾",
-"力干政":"力干政",
-"力干涉":"力干涉",
-"力干預":"力干預",
-"力拼":"力拚",
-"力斗":"力鬥",
-"功勳":"功勛",
-"功虧":"功虧",
-"功虧一匱":"功虧一簣",
-"加上":"加上",
-"加卷":"加捲",
-"加涂":"加塗",
-"加裡":"加里",
-"助理":"助理",
-"劫余":"劫餘",
-"劭雍":"邵雍",
-"劾系":"劾繫",
-"勁斗":"勁鬥",
-"勇斗":"勇鬥",
-"動蕩":"動盪",
-"勘亂":"戡亂",
-"勝余":"勝餘",
-"勞力":"勞力",
-"勞力士表":"勞力士錶",
-"勞騷":"牢騷",
-"勤朴":"勤樸",
-"勳爵":"勛爵",
-"勳章":"勛章",
-"勳級":"勛級",
-"勾干":"勾幹",
-"包准":"包準",
-"包函":"包涵",
-"包卷":"包捲",
-"包扎":"包紮",
-"包谷":"包穀",
-"北回":"北迴",
-"北岳":"北嶽",
-"北極":"北極",
-"北極象":"北極象",
-"北沖":"北衝",
-"北游":"北遊",
-"北裡":"北里",
-"匣缽":"匣砵",
-"匪干":"匪幹",
-"匯刊":"彙刊",
-"匯整":"彙整",
-"匯編":"彙編",
-"匯總":"彙總",
-"匯纂":"彙纂",
-"匯萃":"薈萃",
-"匯輯":"彙輯",
-"匯集":"彙集",
-"區別":"區別",
-"十余":"十餘",
-"十余只":"十餘隻",
-"十余裡":"十餘里",
-"十分":"十分",
-"十分丑":"十分醜",
-"十分准":"十分準",
-"十分干":"十分乾",
-"十分的准":"十分的準",
-"十卷":"十捲",
-"十只":"十隻",
-"十周":"十週",
-"十天":"十天",
-"十海":"十海",
-"十海裡":"十海里",
-"十裡":"十里",
-"十裡余":"十里餘",
-"十面":"十面",
-"千余":"千餘",
-"千余只":"千餘隻",
-"千余裡":"千餘里",
-"千只":"千隻",
-"千層":"千層",
-"千層面":"千層麵",
-"千海":"千海",
-"千海裡":"千海里",
-"千裡":"千里",
-"千鈞":"千鈞",
-"千鈞一發":"千鈞一髮",
-"千鎚":"千錘",
-"卅余":"卅餘",
-"升官":"陞官",
-"升平":"昇平",
-"升汞":"昇汞",
-"升泰":"昇泰",
-"升華":"昇華",
-"升遷":"陞遷",
-"升陽":"昇陽",
-"半只":"半隻",
-"半干":"半乾",
-"卒悉":"窣窸",
-"卒發":"捽髮",
-"卒草":"捽草",
-"卓約":"卓約",
-"卓約風姿":"綽約風姿",
-"南回":"南迴",
-"南岳":"南嶽",
-"南沖":"南衝",
-"南游":"南遊",
-"南裡":"南里",
-"博斗":"博鬥",
-"卡裡":"卡里",
-"卯上":"卯上",
-"卯上干":"卯上幹",
-"印制":"印製",
-"印象":"印象",
-"印象是":"印象是",
-"即像":"即像",
-"即像征服":"即像征服",
-"即像限制":"即像限制",
-"即象":"即像",
-"卷一":"卷一",
-"卷一卷":"捲一捲",
-"卷上":"捲上",
-"卷不":"卷不",
-"卷不起":"捲不起",
-"卷了":"捲了",
-"卷伏":"捲伏",
-"卷伴":"捲伴",
-"卷住":"捲住",
-"卷來":"捲來",
-"卷個":"捲個",
-"卷兒":"捲兒",
-"卷入":"捲入",
-"卷出":"捲出",
-"卷到":"捲到",
-"卷動":"捲動",
-"卷去":"捲去",
-"卷向":"捲向",
-"卷回":"捲回",
-"卷圖":"捲圖",
-"卷土":"捲土",
-"卷土重來":"捲土重來",
-"卷好":"捲好",
-"卷射":"捲射",
-"卷尺":"捲尺",
-"卷往":"捲往",
-"卷得":"捲得",
-"卷心":"捲心",
-"卷心菜":"捲心菜",
-"卷成":"捲成",
-"卷拂":"捲拂",
-"卷掃":"捲掃",
-"卷攏":"捲攏",
-"卷旋":"捲旋",
-"卷曲":"捲曲",
-"卷款":"捲款",
-"卷毛":"捲毛",
-"卷浪":"捲浪",
-"卷湧":"捲湧",
-"卷滾":"捲滾",
-"卷煙":"捲菸",
-"卷發":"捲髮",
-"卷碎":"捲碎",
-"卷筒":"捲筒",
-"卷簾":"捲簾",
-"卷紙":"捲紙",
-"卷縮":"捲縮",
-"卷繞":"捲繞",
-"卷纏":"捲纏",
-"卷罩":"捲罩",
-"卷翹":"捲翹",
-"卷臥":"捲臥",
-"卷至":"捲至",
-"卷舌":"捲舌",
-"卷舞":"捲舞",
-"卷菸":"捲菸",
-"卷落":"卷落",
-"卷落葉":"捲落葉",
-"卷著":"捲著",
-"卷蕩":"捲蕩",
-"卷衣":"卷衣",
-"卷衣袖":"捲衣袖",
-"卷袖":"捲袖",
-"卷裹":"捲裹",
-"卷襲":"捲襲",
-"卷走":"捲走",
-"卷起":"捲起",
-"卷身":"捲身",
-"卷軸":"捲軸",
-"卷逃":"捲逃",
-"卷進":"捲進",
-"卷過":"捲過",
-"卷鋪":"卷鋪",
-"卷鋪蓋":"捲鋪蓋",
-"卷鋼":"捲鋼",
-"卷開":"捲開",
-"卷須":"捲鬚",
-"卷風":"捲風",
-"卷飛":"捲飛",
-"卻象":"卻像",
-"厄立":"厄立",
-"厄立特裡":"厄立特里",
-"厘出":"釐出",
-"厘定":"釐定",
-"厘正":"釐正",
-"厘清":"釐清",
-"厘米":"釐米",
-"厘訂":"釐訂",
-"去卷":"去捲",
-"去干":"去幹",
-"去干一杯":"去乾一杯",
-"去干了杯":"去乾了杯",
-"去干洗":"去乾洗",
-"去干淨":"去乾淨",
-"去干耗":"去乾耗",
-"去干著急":"去乾著急",
-"又丑":"又醜",
-"又咸":"又鹹",
-"又干":"又幹",
-"又干又":"又乾又",
-"又干又硬":"又乾又硬",
-"又干擾":"又干擾",
-"又干政":"又干政",
-"又干涉":"又干涉",
-"又干淨":"又乾淨",
-"又干預":"又干預",
-"又斗":"又鬥",
-"又松":"又鬆",
-"又沖":"又沖",
-"又沖掉":"又衝掉",
-"又象":"又像",
-"又雲":"又云",
-"反卷":"反捲",
-"反反":"反反",
-"反反復復":"反反覆覆",
-"反復":"反覆",
-"反朴":"反朴",
-"反朴歸真":"反璞歸真",
-"反沖":"反衝",
-"取余":"取餘",
-"取准":"取準",
-"取舍":"取捨",
-"取諦":"取締",
-"受人":"受人",
-"受人之托":"受人之託",
-"受托":"受託",
-"受雇":"受僱",
-"口佔":"口占",
-"口干":"口乾",
-"口干一":"口幹一",
-"口干什":"口幹什",
-"口干員":"口幹員",
-"口干掉":"口幹掉",
-"口干活":"口幹活",
-"口干著":"口幹著",
-"口干這":"口幹這",
-"口干那":"口幹那",
-"古朴":"古樸",
-"古法":"古法",
-"古法自制":"古法自製",
-"古跡":"古蹟",
-"古雲":"古云",
-"另辟":"另闢",
-"只余":"只餘",
-"只兔":"隻兔",
-"只大":"只大",
-"只大老鼠":"隻大老鼠",
-"只字":"隻字",
-"只巨":"只巨",
-"只巨獸":"隻巨獸",
-"只影":"隻影",
-"只怪":"只怪",
-"只怪物":"隻怪物",
-"只怪獸":"隻怪獸",
-"只手":"隻手",
-"只招":"只招",
-"只招子":"隻招子",
-"只母":"只母",
-"只母老虎":"隻母老虎",
-"只涂":"只塗",
-"只牛":"隻牛",
-"只犬":"隻犬",
-"只狐":"只狐",
-"只狐狸":"隻狐狸",
-"只狗":"隻狗",
-"只狼":"隻狼",
-"只猛":"只猛",
-"只猛獸":"隻猛獸",
-"只猴":"隻猴",
-"只獅":"隻獅",
-"只畜":"只畜",
-"只畜生":"隻畜生",
-"只眼":"隻眼",
-"只羊":"隻羊",
-"只老":"只老",
-"只老狐狸":"隻老狐狸",
-"只老虎":"隻老虎",
-"只老鼠":"隻老鼠",
-"只聲":"只聲",
-"只聲不出":"隻聲不出",
-"只虎":"隻虎",
-"只蛋":"隻蛋",
-"只言":"只言",
-"只言片字":"隻言片字",
-"只言片語":"隻言片語",
-"只豬":"隻豬",
-"只貓":"隻貓",
-"只身":"隻身",
-"只身分":"隻身份",
-"只野":"只野",
-"只野獸":"隻野獸",
-"只雞":"隻雞",
-"只馬":"隻馬",
-"只騾":"隻騾",
-"只驢":"隻驢",
-"只鬼":"隻鬼",
-"只鳥":"隻鳥",
-"只鳳":"隻鳳",
-"只鴨":"隻鴨",
-"只鵝":"隻鵝",
-"只鷹":"隻鷹",
-"只鹿":"隻鹿",
-"叫於":"叫于",
-"叮叮":"叮叮",
-"叮叮當當":"叮叮噹噹",
-"叮當":"叮噹",
-"可以":"可以",
-"可以干擾":"可以干擾",
-"可以干涉":"可以干涉",
-"可以干預":"可以干預",
-"可以象":"可以像",
-"可余":"可餘",
-"可干":"可干",
-"可干拭":"可乾拭",
-"可干擦":"可乾擦",
-"可干洗":"可乾洗",
-"可干濕":"可乾濕",
-"可干燥":"可乾燥",
-"可干衣":"可乾衣",
-"可干製":"可乾製",
-"可干飲":"可乾飲",
-"可緊":"可緊",
-"可緊可松":"可緊可鬆",
-"可象":"可像",
-"可象征戰":"可像征戰",
-"可贊":"可讚",
-"可采":"可採",
-"台凳":"檯凳",
-"台制":"台製",
-"台子":"檯子",
-"台布":"檯布",
-"台布下":"台佈下",
-"台布告":"台佈告",
-"台布滿":"台佈滿",
-"台布置":"台佈置",
-"台布署":"台佈署",
-"台布道":"台佈道",
-"台布防":"台佈防",
-"台干":"台幹",
-"台歷":"檯曆",
-"台灣":"台灣",
-"台灣制":"台灣製",
-"台灣制定":"台灣制定",
-"台灣制度":"台灣制度",
-"台灣制訂":"台灣制訂",
-"台燈":"檯燈",
-"台球":"檯球",
-"台鐘":"檯鐘",
-"台面":"檯面",
-"台風":"颱風",
-"史跡":"史蹟",
-"右沖":"右衝",
-"叼擾":"叨擾",
-"叼難":"刁難",
-"吃不":"吃不",
-"吃不准":"吃不準",
-"吃准":"吃準",
-"吃完":"吃完",
-"吃完面":"吃完麵",
-"吃干":"吃乾",
-"吃干了":"吃乾了",
-"吃茶":"喫茶",
-"吃著":"吃著",
-"吃過":"吃過",
-"吃過面":"吃過麵",
-"吃面":"吃麵",
-"各干":"各干",
-"各干各":"各幹各",
-"合並":"合併",
-"合並不":"合並不",
-"合並且":"合並且",
-"合並非":"合並非",
-"合伙":"合夥",
-"合府":"闔府",
-"合斗":"合鬥",
-"合眼":"闔眼",
-"合第":"闔第",
-"吉佔":"吉占",
-"吉裡":"吉里",
-"吊信":"弔信",
-"吊取":"弔取",
-"吊古":"弔古",
-"吊唁":"弔唁",
-"吊問":"弔問",
-"吊喉":"弔喉",
-"吊喪":"弔喪",
-"吊場":"弔場",
-"吊奠":"弔奠",
-"吊孝":"弔孝",
-"吊客":"弔客",
-"吊影":"弔影",
-"吊念":"弔念",
-"吊慰":"弔慰",
-"吊文":"弔文",
-"吊旗":"弔旗",
-"吊書":"弔書",
-"吊死":"吊死",
-"吊死問疾":"弔死問疾",
-"吊民":"弔民",
-"吊民伐罪":"弔民伐罪",
-"吊祭":"弔祭",
-"吊系":"吊繫",
-"吊考":"弔考",
-"吊膀":"吊膀",
-"吊膀子":"弔膀子",
-"吊詞":"弔詞",
-"吊詭":"弔詭",
-"吊賀":"弔賀",
-"吊錢":"弔錢",
-"同仇":"同仇",
-"同仇敵慨":"同仇敵愾",
-"同伙":"同夥",
-"同游":"同遊",
-"名字":"名字",
-"名字是於":"名字是于",
-"名跡":"名蹟",
-"后衛":"後衛",
-"吐干":"吐乾",
-"向准":"向準",
-"向午":"晌午",
-"向外":"向外",
-"向外沖":"向外衝",
-"向導":"嚮導",
-"向導游":"嚮導遊",
-"向往":"嚮往",
-"向後":"向後",
-"向後沖":"向後衝",
-"向應":"嚮應",
-"君子":"君子",
-"君子好求":"君子好逑",
-"吞並":"吞併",
-"吞咽":"吞嚥",
-"吟游":"吟遊",
-"吧台":"吧檯",
-"吮干":"吮乾",
-"吵斗":"吵鬥",
-"吸干":"吸乾",
-"吸得":"吸得",
-"吸得干干":"吸得乾乾",
-"吸煙":"吸菸",
-"吹卷":"吹捲",
-"吹干":"吹乾",
-"吹發":"吹髮",
-"吹胡":"吹鬍",
-"吹須":"吹鬚",
-"呂後":"呂后",
-"呆致":"呆致",
-"呆致致":"呆緻緻",
-"告僥":"告饒",
-"告別":"告別",
-"周一":"週一",
-"周一岳":"周一嶽",
-"周三":"週三",
-"周二":"週二",
-"周五":"週五",
-"周休":"週休",
-"周六":"週六",
-"周刊":"週刊",
-"周四":"週四",
-"周報":"週報",
-"周年":"週年",
-"周息":"週息",
-"周數":"週數",
-"周日":"週日",
-"周會":"週會",
-"周期":"週期",
-"周末":"週末",
-"周歲":"週歲",
-"周歷":"週曆",
-"周波":"週波",
-"周游":"周遊",
-"周潤":"周潤",
-"周潤發":"周潤發",
-"周濟":"賙濟",
-"周而":"周而",
-"周而復始":"週而復始",
-"周薪":"週薪",
-"周記":"週記",
-"周遭":"週遭",
-"呷干":"呷乾",
-"呼卷":"呼捲",
-"咀咽":"咀嚥",
-"咎征":"咎徵",
-"咕老":"咕老",
-"咕老肉":"咕咾肉",
-"咨商":"諮商",
-"咨詢":"諮詢",
-"咨諏":"諮諏",
-"咨謀":"諮謀",
-"咨議":"諮議",
-"咬姜":"咬姜",
-"咬姜呷醋":"咬薑呷醋",
-"咬字":"咬字",
-"咬字不准":"咬字不準",
-"咸味":"鹹味",
-"咸咸":"鹹鹹",
-"咸土":"鹹土",
-"咸度":"鹹度",
-"咸得":"鹹得",
-"咸水":"鹹水",
-"咸海":"鹹海",
-"咸淡":"鹹淡",
-"咸湖":"鹹湖",
-"咸湯":"鹹湯",
-"咸濕":"鹹濕",
-"咸的":"鹹的",
-"咸肉":"鹹肉",
-"咸菜":"鹹菜",
-"咸蛋":"鹹蛋",
-"咸豬":"鹹豬",
-"咸類":"鹹類",
-"咸魚":"鹹魚",
-"咸鴨":"鹹鴨",
-"咸鹵":"鹹鹵",
-"咽一":"咽一",
-"咽一口":"嚥一口",
-"咽下":"嚥下",
-"咽不":"咽不",
-"咽不下":"嚥不下",
-"咽不了":"嚥不了",
-"咽了":"嚥了",
-"咽住":"嚥住",
-"咽入":"嚥入",
-"咽口":"咽口",
-"咽口水":"嚥口水",
-"咽唾":"嚥唾",
-"咽嚼":"嚥嚼",
-"咽回":"嚥回",
-"咽得":"咽得",
-"咽得下":"嚥得下",
-"咽氣":"嚥氣",
-"咽苦":"咽苦",
-"咽苦吞甘":"嚥苦吞甘",
-"咽著":"嚥著",
-"咽食":"嚥食",
-"哀吊":"哀弔",
-"品匯":"品彙",
-"品嘗":"品嚐",
-"哆羅":"哆羅",
-"哆羅哆嗦":"哆囉哆嗦",
-"哈羅":"哈囉",
-"哈裡":"哈里",
-"員干":"員干",
-"員干將":"員幹將",
-"哥裡":"哥里",
-"哪裡":"哪裡",
-"哪象":"哪像",
-"哲雲":"哲云",
-"唁吊":"唁弔",
-"唇干":"唇乾",
-"唇燥":"唇燥",
-"唇燥舌干":"唇燥舌乾",
-"唱游":"唱遊",
-"唾面":"唾面",
-"唾面自干":"唾面自乾",
-"商周":"商周",
-"商歷":"商曆",
-"商鋪":"商舖",
-"啟發":"啟發",
-"啷當":"啷噹",
-"喂動":"喂動",
-"喂動物":"餵動物",
-"喂它":"餵牠",
-"喂羊":"餵羊",
-"喂豬":"餵豬",
-"喂馬":"餵馬",
-"喂魚":"餵魚",
-"善斗":"善鬥",
-"喉干":"喉乾",
-"喉干舌燥":"喉乾舌燥",
-"喉頭":"喉頭",
-"喉頭發干":"喉頭發乾",
-"喑啞":"瘖啞",
-"喝一":"喝一",
-"喝一壇":"喝一罈",
-"喝三":"喝三",
-"喝三壇":"喝三罈",
-"喝兩":"喝兩",
-"喝兩壇":"喝兩罈",
-"喝干":"喝乾",
-"喝斗":"喝斗",
-"喝斗聲":"喝鬥聲",
-"喝白":"喝白",
-"喝白干":"喝白乾",
-"喧染":"渲染",
-"單單":"單單",
-"單於":"單于",
-"嗟籲":"嗟吁",
-"嘆服":"歎服",
-"嘆為":"歎為",
-"嘆籲":"嘆吁",
-"嘆羨":"歎羨",
-"嘆賞":"歎賞",
-"嘍羅":"嘍囉",
-"嘩眾":"嘩眾",
-"嘩眾取寵":"譁眾取寵",
-"嘮叼":"嘮叨",
-"嘴干":"嘴乾",
-"嘹亮":"嘹喨",
-"噓唏":"歔欷",
-"噪動":"譟動",
-"噪詐":"譟詐",
-"噫籲":"噫吁",
-"噴涂":"噴塗",
-"嚕嚕":"嚕嚕",
-"嚕嚕蘇蘇":"嚕嚕囌囌",
-"嚕蘇":"嚕囌",
-"嚴複":"嚴復",
-"囉裡":"囉里",
-"囑托":"囑託",
-"四個":"四個",
-"四只":"四隻",
-"四天":"四天",
-"四海":"四海",
-"四海裡":"四海里",
-"四舍":"四捨",
-"四裡":"四里",
-"四裡余":"四里餘",
-"回傭":"回佣",
-"回光":"回光",
-"回光反照":"迴光反照",
-"回光返照":"迴光返照",
-"回卷":"回捲",
-"回圈":"迴圈",
-"回報":"回報",
-"回廊":"迴廊",
-"回復":"回覆",
-"回文":"回文",
-"回文詩":"迴文詩",
-"回旋":"迴旋",
-"回歷":"回曆",
-"回沖":"回衝",
-"回游":"迴游",
-"回然":"迥然",
-"回環":"迴環",
-"回紋":"迴紋",
-"回繞":"迴繞",
-"回翔":"迴翔",
-"回腸":"迴腸",
-"回腸蕩氣":"迴腸蕩氣",
-"回蕩":"迴蕩",
-"回誦":"迴誦",
-"回避":"迴避",
-"回鑾":"迴鑾",
-"回陽":"回陽",
-"回陽蕩氣":"迴陽蕩氣",
-"回響":"迴響",
-"回響當":"迴響當",
-"回風":"迴風",
-"因奸":"因奸",
-"因奸成孕":"因姦成孕",
-"困乏":"睏乏",
-"困倦":"睏倦",
-"困意":"睏意",
-"困獸":"困獸",
-"困獸之斗":"困獸之鬥",
-"困獸斗":"困獸鬥",
-"困獸猶斗":"困獸猶鬥",
-"囹圍":"囹圉",
-"固定":"固定",
-"固定制":"固定制",
-"圈梁":"圈樑",
-"國制":"國製",
-"國制定":"國制定",
-"國制度":"國制度",
-"國制訂":"國制訂",
-"國家":"國家",
-"國小":"國小",
-"國小只":"國小只",
-"國歷":"國曆",
-"國歷代":"國歷代",
-"國歷任":"國歷任",
-"國歷史":"國歷史",
-"國歷屆":"國歷屆",
-"國歷年":"國歷年",
-"國歷盡":"國歷盡",
-"國歷經":"國歷經",
-"國歷險":"國歷險",
-"園游":"園游",
-"園游會":"園遊會",
-"圓椎":"圓錐",
-"圖象":"圖像",
-"團伙":"團夥",
-"團子":"糰子",
-"團粉":"糰粉",
-"土制":"土製",
-"地一":"地一",
-"地一卷":"地一捲",
-"地卷":"地捲",
-"地干":"地干",
-"地干吧":"地幹吧",
-"地干杯":"地乾杯 ",
-"地干燥":"地乾燥",
-"地干裂":"地乾裂",
-"地志":"地誌",
-"地沖":"地衝",
-"均沾":"均霑",
-"坎頂":"崁頂",
-"坐著":"坐著",
-"坤范":"坤範",
-"坦護":"袒護",
-"垂發":"垂髮",
-"垂范":"垂範",
-"型狀":"形狀",
-"型范":"型範",
-"埃裡":"埃裡",
-"埃裡溫":"埃里溫",
-"埔裡":"埔里",
-"基准":"基準",
-"基干":"基幹",
-"報廢":"報廢",
-"報廢後":"報廢後",
-"報道":"報導",
-"塑象":"塑像",
-"塔裡":"塔裡",
-"塔裡木":"塔里木",
-"墓志":"墓誌",
-"墜胎":"墮胎",
-"增辟":"增闢",
-"墨沈":"墨沈",
-"墨沈未干":"墨瀋未乾",
-"墾辟":"墾闢",
-"壇佳":"壇佳",
-"壇佳釀":"罈佳釀",
-"壇壇":"壇壇",
-"壇壇罐罐":"罈罈罐罐",
-"壇女":"壇女",
-"壇女兒紅":"罈女兒紅",
-"壇好":"壇好",
-"壇好酒":"罈好酒",
-"壇子":"罈子",
-"壇燒":"壇燒",
-"壇燒刀子":"罈燒刀子",
-"壇白":"壇白",
-"壇白干":"罈白干",
-"壇美":"壇美",
-"壇美酒":"罈美酒",
-"壇老":"壇老",
-"壇老酒":"罈老酒",
-"壇酒":"罈酒",
-"壇醋":"罈醋",
-"壇陳":"壇陳",
-"壇陳年":"罈陳年",
-"壇騞":"罈騞",
-"壇高":"壇高",
-"壇高粱":"罈高粱",
-"壑然":"豁然",
-"壓咸":"壓鹹",
-"壓面":"壓面",
-"壓面棍":"壓麵棍",
-"壞簽":"壞籤",
-"壺范":"壺範",
-"壽面":"壽麵",
-"夏歷":"夏曆",
-"外制":"外製",
-"外強":"外強",
-"外強中干":"外強中乾",
-"外御":"外御",
-"外御強敵":"外禦強敵",
-"外松":"外松",
-"外松內緊":"外鬆內緊",
-"外涂":"外塗",
-"多丑":"多醜",
-"多余":"多餘",
-"多出":"多出",
-"多出電影":"多齣電影",
-"多少":"多少",
-"多少只":"多少隻",
-"多干":"多干",
-"多干了":"多乾了",
-"多干干":"多乾乾",
-"多干擾":"多干擾",
-"多干杯":"多乾杯",
-"多干涉":"多干涉",
-"多干淨":"多乾淨",
-"多干燥":"多乾燥",
-"多干爽":"多乾爽",
-"多干系":"多干系",
-"多干脆":"多乾脆",
-"多干預":"多干預",
-"多海":"多海",
-"多海裡":"多海里",
-"多裡":"多里",
-"多麼":"多麼",
-"多麼丑":"多麼醜",
-"夜光":"夜光",
-"夜光表":"夜光錶",
-"夜游":"夜遊",
-"夠不":"夠不",
-"夠不著":"搆不著",
-"夠准":"夠準",
-"夠怨":"搆怨",
-"夠涂":"夠塗",
-"夠贊":"夠讚",
-"夠鐘":"夠鐘",
-"夠鐘情":"夠鍾情",
-"夠鐘愛":"夠鍾愛",
-"夢回":"夢迴",
-"夢游":"夢遊",
-"夢系":"夢繫",
-"大丑":"大醜",
-"大伙":"大夥",
-"大伙人":"大夥人",
-"大伙兒":"大夥兒",
-"大利":"大利",
-"大利面":"大利麵",
-"大只":"大隻",
-"大小":"大小",
-"大小只":"大小只",
-"大干":"大干",
-"大干特干":"大幹特幹",
-"大斗":"大鬥",
-"大歷":"大曆",
-"大歷史":"大歷史",
-"大潤":"大潤",
-"大潤發":"大潤發",
-"大衝":"大沖",
-"大衝擊":"大衝擊",
-"大贊":"大讚",
-"天佑":"天祐",
-"天干":"天干",
-"天干物燥":"天乾物燥",
-"天後":"天后",
-"天生":"天生",
-"天生干":"天生幹",
-"天翻":"天翻",
-"天翻地復":"天翻地覆",
-"天象":"天象",
-"太不":"太不",
-"太不准":"太不準",
-"太丑":"太醜",
-"太准":"太準",
-"太初":"太初",
-"太初歷":"太初曆",
-"太咸":"太鹹",
-"太太":"太太",
-"太干":"太乾",
-"太後":"太后",
-"太後面":"太後面",
-"太松":"太鬆",
-"太監":"太監",
-"太象":"太像",
-"夫奸":"夫姦",
-"夫役":"伕役",
-"失准":"失準",
-"夾衣":"裌衣",
-"夾襖":"裌襖",
-"奇丑":"奇醜",
-"奇准":"奇準",
-"奇跡":"奇蹟",
-"奏折":"奏摺",
-"奕棋":"弈棋",
-"奧斯":"奧斯",
-"奧斯卡":"奧斯卡",
-"奪斗":"奪鬥",
-"奪杯":"奪盃",
-"奮斗":"奮鬥",
-"奸凶":"姦凶",
-"奸夫":"姦夫",
-"奸婦":"姦婦",
-"奸屍":"姦屍",
-"奸情":"姦情",
-"奸拐":"姦拐",
-"奸殺":"姦殺",
-"奸污":"姦污",
-"奸淫":"姦淫",
-"奸盜":"姦盜",
-"奸通":"姦通",
-"她丑":"她醜",
-"她干":"她幹",
-"她干兒子":"她乾兒子",
-"她干女兒":"她乾女兒",
-"她干擾":"她干擾",
-"她干涉":"她干涉",
-"她干淨":"她乾淨",
-"她干爹":"她乾爹",
-"她干脆":"她乾脆",
-"她干預":"她干預",
-"她象":"她像",
-"好丑":"好醜",
-"好像":"好像",
-"好像征式":"好象徵式",
-"好像征性":"好象徵性",
-"好准":"好準",
-"好咸":"好鹹",
-"好困":"好困",
-"好困吧":"好睏吧",
-"好困啊":"好睏啊",
-"好好":"好好",
-"好好地干":"好好地幹",
-"好好干":"好好幹",
-"好干":"好幹",
-"好斗":"好鬥",
-"好煙":"好菸",
-"好簽":"好籤",
-"好象":"好像",
-"如何":"如何",
-"如何干":"如何幹",
-"如坐":"如坐",
-"如坐針毯":"如坐針氈",
-"如果":"如果",
-"如果象":"如果像",
-"姑負":"辜負",
-"姓名":"姓名",
-"姓名是於":"姓名是于",
-"姓於":"姓于",
-"姓鐘":"姓鍾",
-"委托":"委託",
-"姜就":"姜就",
-"姜就是老":"薑就是老",
-"姜愈":"姜愈",
-"姜愈老":"薑愈老",
-"姜是":"薑是",
-"姜是老":"薑是老",
-"姜末":"薑末",
-"姜汁":"薑汁",
-"姜油":"薑油",
-"姜湯":"薑湯",
-"姜片":"薑片",
-"姜糖":"薑糖",
-"姜絲":"薑絲",
-"姜老":"姜老",
-"姜老辣":"薑老辣",
-"姜花":"薑花",
-"姜茶":"薑茶",
-"姜蔥":"薑蔥",
-"姜還":"姜還",
-"姜還是老":"薑還是老",
-"姜醋":"薑醋",
-"姜餅":"薑餅",
-"姜黃":"薑黃",
-"姣潔":"皎潔",
-"娉亭":"娉婷",
-"婀挪":"婀娜",
-"嫌凶":"嫌兇",
-"嬉游":"嬉遊",
-"嬌揉":"嬌揉",
-"嬌揉造作":"矯揉造作",
-"嬖幸":"嬖倖",
-"嬴余":"嬴餘",
-"子姜":"子薑",
-"子鍾":"子鐘",
-"子雲":"子云",
-"字匯":"字彙",
-"存折":"存摺",
-"孱雜":"羼雜",
-"它象":"它像",
-"宅弟":"宅第",
-"宅紫":"奼紫",
-"守御":"守禦",
-"安全":"安全",
-"宏基":"宏碁",
-"定准":"定準",
-"定制":"定製",
-"定干":"定幹",
-"定干吧":"定幹吧",
-"客游":"客遊",
-"宣傳":"宣傳",
-"宣傳周":"宣傳週",
-"宣布":"宣佈",
-"室內":"室內",
-"宦游":"宦遊",
-"宴游":"宴遊",
-"家丑":"家醜",
-"家伙":"傢伙",
-"家俱":"傢俱",
-"家私":"傢俬",
-"家裡":"家裡",
-"宸游":"宸遊",
-"容范":"容範",
-"寂聊":"寂寥",
-"寄托":"寄託",
-"密布":"密佈",
-"密折":"密摺",
-"密致":"密緻",
-"富余":"富餘",
-"寒倉":"寒傖",
-"寒喧":"寒暄",
-"寒栗":"寒慄",
-"寒而":"寒而",
-"寒而栗":"寒而慄",
-"寡欲":"寡慾",
-"實干":"實幹",
-"實象":"實像",
-"寧舍":"寧捨",
-"寫字":"寫字",
-"寫字台":"寫字檯",
-"寬松":"寬鬆",
-"寬范":"寬範",
-"寸余":"寸餘",
-"封後":"封后",
-"射不":"射不",
-"射不准":"射不準",
-"射准":"射準",
-"射雕":"射鵰",
-"專干":"專幹",
-"尊范":"尊範",
-"尋准":"尋準",
-"尋麻":"蕁麻",
-"對准":"對準",
-"對干":"對幹",
-"對華":"對華",
-"對著":"對著",
-"對象":"對象",
-"對象征":"對象徵",
-"導游":"導遊",
-"小丑":"小丑",
-"小丑丫鬟":"小醜丫鬟",
-"小伙":"小夥",
-"少干":"少幹",
-"少干擾":"少干擾",
-"少干涉":"少干涉",
-"少干系":"少干係",
-"少干預":"少干預",
-"少沖":"少衝",
-"少海":"少海",
-"少海裡":"少海里",
-"少裡":"少里",
-"尚余":"尚餘",
-"尤如":"猶如",
-"尤帶":"猶帶",
-"尤憐":"猶憐",
-"尤疑":"猶疑",
-"尤裡":"尤里",
-"尤豫":"猶豫",
-"就像":"就像",
-"就像征":"就象徵",
-"就吃":"就吃",
-"就吃干":"就吃乾",
-"就干":"就干",
-"就干一":"就幹一",
-"就干一杯":"就乾一杯",
-"就干吧":"就幹吧",
-"就干擾":"就干擾",
-"就干杯":"就乾杯",
-"就干涉":"就干涉",
-"就干淨":"就乾淨",
-"就干爽":"就乾爽",
-"就干脆":"就乾脆",
-"就干預":"就干預",
-"就范":"就範",
-"就象":"就像",
-"尺余":"尺餘",
-"局促":"侷促",
-"局蹙":"跼蹙",
-"局躅":"跼躅",
-"局限":"侷限",
-"居裡":"居里",
-"屋內":"屋內",
-"屋梁":"屋樑",
-"屍位":"尸位",
-"屍祿":"尸祿",
-"展布":"展佈",
-"屢僕":"屢仆",
-"層理":"層理",
-"屬制":"屬製",
-"屯扎":"屯紮",
-"山岳":"山嶽",
-"山涯":"山崖",
-"山羊":"山羊",
-"山羊胡":"山羊鬍",
-"山羊須":"山羊鬚",
-"岳峙":"嶽峙",
-"岳立":"嶽立",
-"峇裡":"峇里",
-"峇釐":"峇里",
-"峰回":"峰迴",
-"島煉":"島鍊",
-"嵇首":"稽首",
-"巡回":"巡迴",
-"巡戈":"巡弋",
-"巡游":"巡遊",
-"工余":"工餘",
-"左沖":"左衝",
-"巧干":"巧幹",
-"巨制":"巨製",
-"巨子":"鉅子",
-"巨細":"鉅細",
-"巨細靡遺":"鉅細靡遺",
-"巨須":"巨鬚",
-"巨鹿":"鉅鹿",
-"己干":"己幹",
-"已余":"已餘",
-"已愈":"已癒",
-"已松":"已鬆",
-"已松下":"已鬆下",
-"已經":"已經",
-"已經斗":"已經鬥",
-"已經象":"已經像",
-"已象":"已像",
-"已象征戰":"已像征戰",
-"已象征服":"已像征服",
-"巴爾":"巴爾",
-"巴爾干":"巴爾幹",
-"巴西":"巴西",
-"巴釐":"峇里",
-"布一":"布一",
-"布一個":"佈一個",
-"布伏":"佈伏",
-"布個":"佈個",
-"布兵":"佈兵",
-"布告":"佈告",
-"布壇":"佈壇",
-"布導":"佈導",
-"布局":"佈局",
-"布崗":"佈崗",
-"布於":"佈於",
-"布施":"佈施",
-"布景":"佈景",
-"布滿":"佈滿",
-"布疑":"布疑",
-"布疑陣":"佈疑陣",
-"布置":"佈置",
-"布署":"佈署",
-"布著":"佈著",
-"布裡":"布里",
-"布裡斯托":"布里斯托",
-"布谷":"布穀",
-"布防":"佈防",
-"布陣":"佈陣",
-"布雷":"佈雷",
-"希拉":"希拉",
-"希拉裡":"希拉里",
-"帕特":"帕特",
-"帕特裡":"帕特里",
-"帝後":"帝后",
-"師云":"師云",
-"師云而雲":"師云而云",
-"師范":"師範",
-"席卷":"席捲",
-"席子":"蓆子",
-"帶征":"帶徵",
-"帶發":"帶髮",
-"帶著":"帶著",
-"常干":"常幹",
-"帽瓔":"帽纓",
-"幡間":"墦間",
-"干一":"幹一",
-"干一壇":"乾一罈",
-"干一壇法":"幹一壇法",
-"干一杯":"乾一杯",
-"干一碗":"乾一碗",
-"干上":"幹上",
-"干下":"干下",
-"干下去":"幹下去",
-"干不":"幹不",
-"干不下":"幹不下",
-"干不干":"幹不幹",
-"干不干杯":"乾不乾杯",
-"干不干淨":"乾不乾淨",
-"干了":"幹了",
-"干事":"幹事",
-"干井":"乾井",
-"干些":"幹些",
-"干什":"幹什",
-"干他":"幹他",
-"干件":"幹件",
-"干任":"干任",
-"干任何":"幹任何",
-"干份":"幹份",
-"干伸":"干伸",
-"干伸舌":"乾伸舌",
-"干似":"乾似",
-"干何":"干何",
-"干何事":"幹何事",
-"干你":"幹你",
-"干你娘":"幹你娘",
-"干個":"幹個",
-"干假":"幹假",
-"干傻":"干傻",
-"干傻事":"幹傻事",
-"干兒":"乾兒",
-"干兩":"干兩",
-"干兩年":"幹兩年",
-"干兩杯":"乾兩杯",
-"干冒":"干冒",
-"干冒煙":"乾冒煙",
-"干冰":"乾冰",
-"干冷":"乾冷",
-"干出":"幹出",
-"干別":"干別",
-"干別的":"幹別的",
-"干到":"幹到",
-"干勁":"幹勁",
-"干化":"乾化",
-"干又":"干又",
-"干又熱":"乾又熱",
-"干吏":"幹吏",
-"干咳":"乾咳",
-"干咽":"乾嚥",
-"干品":"乾品",
-"干員":"幹員",
-"干哥":"乾哥",
-"干哪":"幹哪",
-"干哭":"乾哭",
-"干啞":"乾啞",
-"干啥":"幹啥",
-"干嗎":"幹嗎",
-"干嘔":"乾嘔",
-"干嘛":"幹嘛",
-"干嚎":"乾嚎",
-"干土":"乾土",
-"干坐":"干坐",
-"干坐著":"乾坐著",
-"干坤":"乾坤",
-"干壇":"乾罈",
-"干壇主":"幹壇主",
-"干壇法":"幹壇法",
-"干壞":"干壞",
-"干壞事":"幹壞事",
-"干大":"干大",
-"干大事":"幹大事",
-"干女":"干女",
-"干女兒":"乾女兒",
-"干女友":"幹女友",
-"干女同事":"幹女同事",
-"干女婿":"乾女婿",
-"干女教師":"幹女教師",
-"干她":"幹她",
-"干好":"幹好",
-"干妹":"乾妹",
-"干姊":"乾姊",
-"干姐":"乾姐",
-"干姜":"乾薑",
-"干娘":"乾娘",
-"干媽":"乾媽",
-"干季":"乾季",
-"干孫":"乾孫",
-"干完":"幹完",
-"干就":"干就",
-"干就干":"幹就幹",
-"干屍":"乾屍",
-"干巴":"乾巴",
-"干巴巴":"乾巴巴",
-"干布":"乾布",
-"干干":"乾乾",
-"干干巴巴":"乾乾巴巴",
-"干干淨淨":"乾乾淨淨",
-"干干爽爽":"乾乾爽爽",
-"干干瘦瘦":"乾乾瘦瘦",
-"干干的":"乾乾的",
-"干干脆脆":"乾乾脆脆",
-"干幾":"干幾",
-"干幾件":"幹幾件",
-"干幾宗":"幹幾宗",
-"干幾手":"乾幾手",
-"干幾杯":"乾幾杯",
-"干幾樁":"幹幾樁",
-"干幾番":"幹幾番",
-"干幾碗":"乾幾碗",
-"干幾輩":"幹幾輩",
-"干底":"乾底",
-"干弋":"干戈",
-"干式":"乾式",
-"干弟":"乾弟",
-"干強":"干強",
-"干強盜":"幹強盜",
-"干得":"幹得",
-"干得一杯":"乾得一杯",
-"干得三杯":"乾得三杯",
-"干得兩杯":"乾得兩杯",
-"干得很":"乾得很",
-"干性":"乾性",
-"干慣":"幹慣",
-"干成":"幹成",
-"干我":"幹我",
-"干手":"干手",
-"干手淨腳":"乾手淨腳",
-"干才":"幹才",
-"干打":"干打",
-"干打雷":"乾打雷",
-"干掉":"幹掉",
-"干掉一杯":"乾掉一杯",
-"干掉一瓶":"乾掉一瓶",
-"干掉一碗":"乾掉一碗",
-"干掉杯":"乾掉杯",
-"干掉這杯":"乾掉這杯",
-"干掉這碗":"乾掉這碗",
-"干掉那杯":"乾掉那杯",
-"干掉那碗":"乾掉那碗",
-"干探":"幹探",
-"干撇":"干撇",
-"干撇下":"乾撇下",
-"干擱":"乾擱",
-"干數":"干數",
-"干數杯":"乾數杯",
-"干料":"乾料",
-"干旱":"乾旱",
-"干曬":"乾曬",
-"干材":"幹材",
-"干杯":"乾杯",
-"干果":"乾果",
-"干枝":"乾枝",
-"干枯":"乾枯",
-"干架":"幹架",
-"干柴":"乾柴",
-"干校":"幹校",
-"干梅":"乾梅",
-"干樁":"幹樁",
-"干正":"干正",
-"干正事":"幹正事",
-"干此":"幹此",
-"干此壇":"乾此罈",
-"干此杯":"乾此杯",
-"干毛":"干毛",
-"干毛巾":"乾毛巾",
-"干江":"灨江",
-"干沙":"乾沙",
-"干泥":"乾泥",
-"干洗":"乾洗",
-"干活":"幹活",
-"干流":"幹流",
-"干涸":"乾涸",
-"干淨":"乾淨",
-"干清":"干清",
-"干清宮":"乾清宮",
-"干渴":"乾渴",
-"干澀":"乾澀",
-"干澗":"乾澗",
-"干濕":"乾濕",
-"干濕發":"乾濕髮",
-"干焦":"乾焦",
-"干熬":"乾熬",
-"干熱":"乾熱",
-"干燒":"乾燒",
-"干燥":"乾燥",
-"干爸":"乾爸",
-"干爹":"乾爹",
-"干爽":"乾爽",
-"干犯":"干犯",
-"干犯法":"幹犯法",
-"干球":"幹球",
-"干甚":"幹甚",
-"干生":"干生",
-"干生氣":"乾生氣",
-"干疤":"乾疤",
-"干瘦":"乾瘦",
-"干癟":"乾癟",
-"干癮":"乾癮",
-"干發":"乾髮",
-"干的":"干的",
-"干的好":"幹的好",
-"干盡":"幹盡",
-"干盡一壇":"乾盡一磹",
-"干盡一壺":"乾盡一壺",
-"干盡一杯":"乾盡一杯",
-"干盡一碗":"乾盡一碗",
-"干眼":"乾眼",
-"干瞪":"乾瞪",
-"干硬":"乾硬",
-"干窘":"乾窘",
-"干站":"干站",
-"干站著":"乾站著",
-"干笑":"乾笑",
-"干等":"乾等",
-"干粉":"乾粉",
-"干粗":"干粗",
-"干粗活":"幹粗活",
-"干糧":"乾糧",
-"干系":"干係",
-"干細":"干細",
-"干細胞":"幹細胞",
-"干結":"乾結",
-"干絲":"乾絲",
-"干綱":"干綱",
-"干綱不振":"乾綱不振",
-"干線":"幹線",
-"干練":"幹練",
-"干耗":"乾耗",
-"干聲":"乾聲",
-"干肉":"乾肉",
-"干肉販":"幹肉販",
-"干股":"乾股",
-"干脆":"乾脆",
-"干花":"乾花",
-"干草":"乾草",
-"干菜":"乾菜",
-"干葉":"乾葉",
-"干著":"幹著",
-"干著急":"乾著急",
-"干薪":"乾薪",
-"干號":"乾號",
-"干衣":"乾衣",
-"干裂":"乾裂",
-"干警":"幹警",
-"干貨":"乾貨",
-"干貨物":"幹貨物",
-"干買":"干買",
-"干買賣":"幹買賣",
-"干起":"幹起",
-"干路":"幹路",
-"干辦":"幹辦",
-"干農":"干農",
-"干農活":"幹農活",
-"干透":"乾透",
-"干這":"幹這",
-"干這一杯":"乾這一杯",
-"干這杯":"乾這杯",
-"干過":"幹過",
-"干過一杯":"乾過一杯",
-"干過杯":"乾過杯",
-"干過癮":"乾過癮",
-"干道":"幹道",
-"干那":"幹那",
-"干那一杯":"乾那一杯",
-"干那杯":"乾那杯",
-"干部":"幹部",
-"干酪":"乾酪",
-"干醋":"乾醋",
-"干錯":"幹錯",
-"干陽":"乾陽",
-"干隆":"乾隆",
-"干電":"干電",
-"干電池":"乾電池",
-"干面":"乾麵",
-"干面前":"乾面前",
-"干飯":"乾飯",
-"干麼":"幹麼",
-"干點":"幹點",
-"干！":"幹！",
-"平准":"平準",
-"平台":"平台",
-"年余":"年餘",
-"年准":"年准",
-"年准到":"年準到",
-"年歷":"年曆",
-"年歷史":"年歷史",
-"幸免":"倖免",
-"幸存":"倖存",
-"幸幸":"倖幸",
-"幽郁":"幽鬱",
-"幾乎":"幾乎",
-"幾乎象":"幾乎像",
-"幾個":"幾個",
-"幾出":"幾齣",
-"幾只":"幾隻",
-"幾周":"幾週",
-"幾天":"幾天",
-"幾天后":"幾天後",
-"幾案":"几案",
-"幾海":"幾海",
-"幾海裡":"幾海里",
-"幾淨":"几淨",
-"幾裡":"幾里",
-"庇裡":"庇裡",
-"庇裡牛斯":"庇里牛斯",
-"床席":"床蓆",
-"床第":"床笫",
-"底裡":"底里",
-"店伙":"店夥",
-"店鋪":"店舖",
-"座象":"座像",
-"康愈":"康癒",
-"廚余":"廚餘",
-"廝斗":"廝鬥",
-"廝斗然":"廝陡然",
-"廢後":"廢后",
-"廣布":"廣佈",
-"廣征":"廣徵",
-"延長":"延長",
-"廿余":"廿餘",
-"弄干":"弄乾",
-"弄松":"弄鬆",
-"弄面":"弄面",
-"弄面吃":"弄麵吃",
-"引發":"引發",
-"弘歷":"弘曆",
-"弛騁":"馳騁",
-"弦樂":"絃樂",
-"弦歌":"絃歌",
-"弦琴":"絃琴",
-"弦管":"絃管",
-"弦索":"絃索",
-"弦線":"絃線",
-"弱干":"弱幹",
-"強力":"強力",
-"強力干性":"強力乾性",
-"強力干濕":"強力乾濕",
-"強台":"強颱",
-"強奸":"強姦",
-"強干":"強幹",
-"強干擾":"強干擾",
-"強干杯":"強乾杯",
-"強干涉":"強干涉",
-"強干脆":"強乾脆",
-"強干預":"強干預",
-"強征":"強徵",
-"強御":"強禦",
-"強沖":"強衝",
-"強界":"疆界",
-"彈子":"彈子",
-"彈子台":"彈子檯",
-"彈涂":"彈塗",
-"彈珠":"彈珠",
-"彈珠台":"彈珠檯",
-"彌山":"彌山",
-"彌山遍野":"瀰山遍野",
-"彌彌":"瀰瀰",
-"彌散":"瀰散",
-"彌漫":"瀰漫",
-"彌蒙":"瀰濛",
-"形單":"形單",
-"形單影只":"形單影隻",
-"形影":"形影",
-"形影相吊":"形影相弔",
-"形象":"形象",
-"彩女":"綵女",
-"彩帶":"綵帶",
-"彩排":"綵排",
-"彩杖":"綵杖",
-"彩棚":"綵棚",
-"彩樓":"綵樓",
-"彩燈":"綵燈",
-"彩牌":"綵牌",
-"彩球":"綵球",
-"彩筆":"彩筆",
-"彩筆生":"綵筆生",
-"彩綢":"綵綢",
-"彩線":"綵線",
-"彩緞":"綵緞",
-"彩繒":"綵繒",
-"彩衣":"綵衣",
-"彩裳":"綵裳",
-"彩頭":"綵頭",
-"彩鳳":"綵鳳",
-"彩鸞":"綵鸞",
-"影後":"影后",
-"影象":"影像",
-"彷佛":"彷彿",
-"彷徨":"徬徨",
-"彼彼":"彼彼",
-"彼彼皆是":"比比皆是",
-"往外":"往外",
-"往外斗":"往外鬥",
-"往外沖":"往外衝",
-"往後":"往後",
-"往後沖":"往後衝",
-"征了":"徵了",
-"征人":"徵人",
-"征信":"徵信",
-"征個":"徵個",
-"征候":"徵候",
-"征兆":"徵兆",
-"征入":"徵入",
-"征兵":"徵兵",
-"征募":"徵募",
-"征去":"徵去",
-"征友":"徵友",
-"征取":"徵取",
-"征召":"徵召",
-"征啟":"徵啟",
-"征地":"徵地",
-"征婚":"徵婚",
-"征引":"徵引",
-"征得":"徵得",
-"征意":"征意",
-"征意見":"徵意見",
-"征收":"徵收",
-"征斂":"徵斂",
-"征文":"徵文",
-"征服":"征服",
-"征服兵":"徵服兵",
-"征求":"徵求",
-"征片":"徵片",
-"征狀":"徵狀",
-"征用":"徵用",
-"征稅":"徵稅",
-"征稿":"徵稿",
-"征管":"徵管",
-"征糧":"徵糧",
-"征納":"徵納",
-"征繳":"徵繳",
-"征聘":"徵聘",
-"征詔":"徵詔",
-"征詢":"徵詢",
-"征調":"徵調",
-"征象":"徵象",
-"征財":"徵財",
-"征費":"徵費",
-"征賦":"徵賦",
-"征購":"徵購",
-"征逐":"徵逐",
-"征選":"徵選",
-"征集":"徵集",
-"很丑":"很醜",
-"很准":"很準",
-"很咸":"很鹹",
-"很干":"很乾",
-"很松":"很鬆",
-"很象":"很像",
-"後冠":"后冠",
-"後制":"後製",
-"後土":"后土",
-"後奸":"後姦",
-"後妃":"后妃",
-"後梁":"後樑",
-"後稷":"后稷",
-"後羿":"后羿",
-"後象":"後像",
-"後象征服":"後像征服",
-"徑取":"逕取",
-"徑啟":"逕啟",
-"徑寄":"逕寄",
-"徑庭":"逕庭",
-"徑砸":"逕砸",
-"徑答":"逕答",
-"徑自":"逕自",
-"徒峭":"陡峭",
-"徒托":"徒託",
-"得准":"得準",
-"得准不准":"得準不準",
-"得象":"得像",
-"徜若":"倘若",
-"御侮":"禦侮",
-"御制":"御製",
-"御寇":"禦寇",
-"御寒":"禦寒",
-"御敵":"禦敵",
-"徨恐":"惶恐",
-"復亡":"覆亡",
-"復仞":"復仞",
-"復仞年如":"複仞年如",
-"復以":"復以",
-"復以百萬":"複以百萬",
-"復信":"覆信",
-"復函":"覆函",
-"復分":"復分",
-"復分數":"複分數",
-"復分解":"複分解",
-"復列":"複列",
-"復利":"複利",
-"復制":"複製",
-"復印":"複印",
-"復句":"複句",
-"復合":"復合",
-"復合傳動":"複合傳動",
-"復合包裝":"複合包裝",
-"復合型":"複合型",
-"復合增長":"複合增長",
-"復合字":"複合字",
-"復合式":"複合式",
-"復合性":"複合性",
-"復合技":"複合技",
-"復合材料":"複合材料",
-"復合板":"複合板",
-"復合架":"複合架",
-"復合模":"複合模",
-"復合機":"複合機",
-"復合物":"複合物",
-"復合管":"複合管",
-"復合膜":"複合膜",
-"復合藥":"複合藥",
-"復合裝甲":"複合裝甲",
-"復合詞":"複合詞",
-"復合金屬":"複合金屬",
-"復合體":"複合體",
-"復名":"復名",
-"復命":"覆命",
-"復基":"復基",
-"復基因":"複基因",
-"復壁":"複壁",
-"復姓":"複姓",
-"復字":"復字",
-"復字鍵":"複字鍵",
-"復審":"覆審",
-"復寫":"複寫",
-"復巢":"覆巢",
-"復式":"複式",
-"復復":"覆覆",
-"復敗":"覆敗",
-"復數":"複數",
-"復文":"覆文",
-"復方":"複方",
-"復書":"覆書",
-"復本":"複本",
-"復查":"複查",
-"復核":"覆核",
-"復檢":"複檢",
-"復次":"複次",
-"復比":"複比",
-"復水":"復水",
-"復水難收":"覆水難收",
-"復決":"複決",
-"復沒":"覆沒",
-"復測":"複測",
-"復滅":"覆滅",
-"復疊":"覆疊",
-"復目":"複目",
-"復相":"復相",
-"復相關":"複相關",
-"復眼":"複眼",
-"復種":"複種",
-"復線":"複線",
-"復習":"複習",
-"復舟":"覆舟",
-"復色":"複色",
-"復葉":"複葉",
-"復蓋":"覆蓋",
-"復蘇":"復甦",
-"復診":"複診",
-"復詞":"複詞",
-"復試":"複試",
-"復變":"複變",
-"復賽":"複賽",
-"復車":"復車",
-"復車之戒":"覆車之戒",
-"復轍":"覆轍",
-"復述":"複述",
-"復選":"複選",
-"復錢":"複錢",
-"復閱":"複閱",
-"復雜":"複雜",
-"復電":"覆電",
-"復音":"複音",
-"復韻":"複韻",
-"微松":"微鬆",
-"微軟":"微軟",
-"微須":"微鬚",
-"德裡":"德里",
-"徽菌":"黴菌",
-"心底":"心底",
-"心底裡":"心底裡",
-"心理":"心理",
-"心相":"心相",
-"心相系":"心相繫",
-"心系":"心繫",
-"心系統":"心系統",
-"心細":"心細",
-"心細似發":"心細似髮",
-"心細如發":"心細如髮",
-"心髒":"心臟",
-"必游":"必遊",
-"必須":"必須",
-"必須品":"必需品",
-"忍飢":"忍饑",
-"志哀":"誌哀",
-"志喜":"誌喜",
-"志慶":"誌慶",
-"志異":"誌異",
-"志發":"志發",
-"快且":"快且",
-"快且准":"快且準",
-"快卷":"快捲",
-"快又":"快又",
-"快又准":"快又準",
-"快吃":"快吃",
-"快吃干":"快吃乾",
-"快干":"快干",
-"快干不":"快幹不",
-"快干了":"快乾了",
-"快干事":"快幹事",
-"快干吧":"快幹吧",
-"快干嘛":"快幹嘛",
-"快干完":"快幹完",
-"快干杯":"快乾杯",
-"快干死":"快幹死",
-"快干活":"快幹活",
-"快干裂":"快乾裂",
-"快幹":"快幹",
-"快幹啊":"快幹啊",
-"快快":"快快",
-"快快干":"快快幹",
-"快松":"快松",
-"快松下":"快鬆下",
-"快沖":"快衝",
-"快狠":"快狠",
-"快狠准":"快狠準",
-"快而":"快而",
-"快而准":"快而準",
-"快舍":"快舍",
-"快舍下":"快捨下",
-"念一":"念一",
-"念一念":"唸一唸",
-"念一聲":"唸一聲",
-"念下":"念下",
-"念下去":"唸下去",
-"念了":"念了",
-"念了一聲":"唸了一聲",
-"念佛":"唸佛",
-"念作":"唸作",
-"念來":"念來",
-"念來念去":"唸來唸去",
-"念到":"唸到",
-"念咒":"唸咒",
-"念好":"唸好",
-"念完":"唸完",
-"念得":"唸得",
-"念念":"唸唸",
-"念成":"唸成",
-"念書":"唸書",
-"念法":"唸法",
-"念給":"唸給",
-"念經":"唸經",
-"念著":"唸著",
-"念誦":"唸誦",
-"念起":"念起",
-"念起來":"唸起來",
-"念過":"唸過",
-"念錯":"唸錯",
-"念點":"唸點",
-"忽松":"忽鬆",
-"忽舍":"忽舍",
-"忽舍下":"忽捨下",
-"怎幺":"怎麼",
-"怎麼":"怎麼",
-"怎麼干":"怎麼幹",
-"怎麼斗":"怎麼鬥",
-"怒氣":"怒氣",
-"怒氣沖沖":"怒氣衝衝",
-"怒發":"怒發",
-"怒發沖冠":"怒髮衝冠",
-"怒發沖天":"怒髮衝天",
-"怕丑":"怕醜",
-"怖栗":"怖慄",
-"急卷":"急捲",
-"急征":"急徵",
-"急沖":"急衝",
-"性征":"性徵",
-"性欲":"性慾",
-"恂栗":"恂慄",
-"恢複":"恢復",
-"恥奪":"褫奪",
-"恭逢":"躬逢",
-"息影":"息影",
-"息影後":"息影後",
-"恰象":"恰像",
-"悒郁":"悒鬱",
-"悚栗":"悚慄",
-"悠游":"悠遊",
-"悲戚":"悲慼",
-"悲泯":"悲憫",
-"悲郁":"悲鬱",
-"悶郁":"悶鬱",
-"悸栗":"悸慄",
-"情欲":"情慾",
-"惕栗":"惕慄",
-"惡心":"噁心",
-"惡斗":"惡鬥",
-"想干":"想幹",
-"想干淨":"想乾淨",
-"想干笑":"想乾笑",
-"想干脆":"想乾脆",
-"想斗":"想鬥",
-"想著":"想著",
-"想象":"想像",
-"惴栗":"惴慄",
-"惶栗":"惶慄",
-"惺松":"惺忪",
-"愁郁":"愁鬱",
-"愈卷":"愈捲",
-"愈合":"癒合",
-"愈合乎":"愈合乎",
-"愈合他":"愈合他",
-"愈合作":"愈合作",
-"愈合她":"愈合她",
-"愈合式":"愈合式",
-"愈合得":"愈合得",
-"愈合心":"愈合心",
-"愈合情":"愈合情",
-"愈合意":"愈合意",
-"愈合我":"愈合我",
-"愈合拍":"愈合拍",
-"愈合攏":"愈合攏",
-"愈合於":"愈合於",
-"愈合理":"愈合理",
-"愈合用":"愈合用",
-"愈合身":"愈合身",
-"愈合適":"愈合適",
-"愈斗":"愈斗",
-"愈斗愈":"愈鬥愈",
-"愈瘡":"癒瘡",
-"愈象":"愈像",
-"意函":"意涵",
-"意興":"意興",
-"意興闌姍":"意興闌珊",
-"愚暗":"愚闇",
-"愛屋":"愛屋",
-"愛屋及鳥":"愛屋及烏",
-"愛欲":"愛慾",
-"感覺":"感覺",
-"感覺象":"感覺像",
-"慘栗":"慘慄",
-"慢咽":"慢嚥",
-"慰借":"慰藉",
-"憂郁":"憂鬱",
-"憋腳":"蹩腳",
-"憑借":"憑藉",
-"憑准":"憑準",
-"憑吊":"憑弔",
-"憔瘁":"憔悴",
-"懇托":"懇託",
-"懈松":"懈鬆",
-"懈迨":"懈怠",
-"應征":"應徵",
-"應象":"應像",
-"懊郁":"懊鬱",
-"懍栗":"懍慄",
-"懰栗":"懰慄",
-"懷表":"懷錶",
-"懸梁":"懸樑",
-"懸臂":"懸臂",
-"懸臂梁":"懸臂樑",
-"懼栗":"懼慄",
-"懿范":"懿範",
-"戊戍":"戊戌",
-"戊衛":"戌衛",
-"成長":"成長",
-"我丑":"我醜",
-"我干":"我幹",
-"我干一杯":"我乾一杯",
-"我干兒子":"我乾兒子",
-"我干女兒":"我乾女兒",
-"我干擾":"我干擾",
-"我干杯":"我乾杯",
-"我干涉":"我干涉",
-"我干爹":"我乾爹",
-"我干脆":"我乾脆",
-"我干預":"我干預",
-"我白":"我白",
-"我白干":"我白幹",
-"我象":"我像",
-"戒煙":"戒菸",
-"或咸":"或鹹",
-"或松":"或鬆",
-"或系":"或系",
-"或系之牛":"或繫之牛",
-"或采":"或採",
-"戚嗟":"慼嗟",
-"戚憂":"慼憂",
-"戚戚":"慼慼",
-"戚然":"慼然",
-"戚顏":"慼顏",
-"戰天":"戰天",
-"戰天斗地":"戰天鬥地",
-"戰斗":"戰鬥",
-"戰栗":"顫慄",
-"戲彩":"戲綵",
-"戲扎":"戲紮",
-"戴克":"戴克",
-"戴克裡先":"戴克里先",
-"戴表":"戴錶",
-"所余":"所餘",
-"所制":"所制",
-"所制約":"所制約",
-"所制造":"所製造",
-"所卷":"所捲",
-"所布":"所布",
-"所布之":"所佈之",
-"所布的":"所佈的",
-"所干":"所幹",
-"所扎":"所紮",
-"所系":"所繫",
-"所見":"所見",
-"所見而雲":"所見而云",
-"所謂":"所謂",
-"所謂而雲":"所謂而云",
-"所譏":"所譏",
-"所譏而雲":"所譏而云",
-"所采":"改採",
-"所鐘":"所鍾",
-"所雲":"所云",
-"扁剉":"扁銼",
-"扇動":"搧動",
-"扇耳":"扇耳",
-"扇耳光":"搧耳光",
-"扇開":"搧開",
-"扇風":"搧風",
-"手一":"手一",
-"手一卷":"手一捲",
-"手一松":"手一鬆",
-"手不":"手不",
-"手不松":"手不鬆",
-"手松":"手鬆",
-"手煉":"手鍊",
-"手表":"手錶",
-"手表態":"手表態",
-"手表演":"手表演",
-"手表現":"手表現",
-"手表示":"手表示",
-"手表達":"手表達",
-"手裡":"手裡",
-"手裡余":"手裡餘",
-"手采":"手採",
-"才不":"才不",
-"才不干杯":"才不乾杯",
-"才不干枯":"才不乾枯",
-"才不干涸":"才不乾涸",
-"才不干渴":"才不乾渴",
-"才不干燥":"才不乾燥",
-"才干":"才幹",
-"才干了杯":"才乾了杯",
-"才干干淨":"才乾乾淨",
-"才干旱":"才乾旱",
-"才干杯":"才乾杯",
-"才干枯":"才乾枯",
-"才干涸":"才乾涸",
-"才干淨":"才乾淨",
-"才干爽":"才乾爽",
-"才干耗":"才乾耗",
-"才干脆":"才乾脆",
-"才干著急":"才乾著急",
-"才干透":"才乾透",
-"才干這杯":"才乾這杯",
-"才干那杯":"才乾那杯",
-"才松":"才松",
-"才松下":"才鬆下",
-"才象":"才像",
-"扎上":"紮上",
-"扎下":"紮下",
-"扎了":"紮了",
-"扎住":"紮住",
-"扎囮":"紮囮",
-"扎在":"紮在",
-"扎好":"紮好",
-"扎實":"紮實",
-"扎寨":"紮寨",
-"扎帶":"紮帶",
-"扎彩":"紮彩",
-"扎成":"紮成",
-"扎扎":"紮紮",
-"扎根":"紮根",
-"扎營":"紮營",
-"扎穩":"紮穩",
-"扎結":"紮結",
-"扎緊":"紮緊",
-"扎腳":"紮腳",
-"扎著":"紮著",
-"扎裹":"紮裹",
-"扎詐":"紮詐",
-"扎起":"紮起",
-"扎辮":"紮辮",
-"打不":"打不",
-"打不准":"打不準",
-"打制":"打製",
-"打拼":"打拚",
-"打斗":"打鬥",
-"打谷":"打穀",
-"托了":"託了",
-"托事":"託事",
-"托交":"託交",
-"托人":"託人",
-"托付":"託付",
-"托古":"託古",
-"托名":"託名",
-"托咎":"託咎",
-"托夢":"託夢",
-"托大":"託大",
-"托孤":"託孤",
-"托故":"託故",
-"托疾":"託疾",
-"托病":"託病",
-"托管":"託管",
-"托言":"託言",
-"托詞":"託詞",
-"托買":"託買",
-"托賣":"託賣",
-"托身":"託身",
-"托辭":"託辭",
-"托運":"託運",
-"托過":"託過",
-"扛丸":"摃丸",
-"扛龜":"摃龜",
-"扣准":"扣準",
-"扮象":"扮像",
-"扶余":"扶餘",
-"批復":"批覆",
-"批斗":"批鬥",
-"批注":"批註",
-"扼塞":"阨塞",
-"找不":"找不",
-"找不准":"找不準",
-"找准":"找準",
-"承制":"承製",
-"抉別":"訣別",
-"把你":"把你",
-"把你干":"把我幹",
-"把姜":"把薑",
-"把方":"把方",
-"把方才":"把才纔",
-"抑制":"抑制",
-"抑郁":"抑鬱",
-"抓不":"抓不",
-"抓不准":"抓不準",
-"抓准":"抓準",
-"抓奸":"抓姦",
-"抓斗":"抓鬥",
-"投不":"投不",
-"投不准":"投不準",
-"投准":"投準",
-"抗御":"抗禦",
-"抗斗":"抗鬥",
-"折》":"摺》",
-"折卷":"折捲",
-"折子":"摺子",
-"折扇":"摺扇",
-"折沖":"折衝",
-"折疊":"摺疊",
-"折痕":"摺痕",
-"折紙":"摺紙",
-"折裙":"摺裙",
-"披發":"披髮",
-"抬扛":"抬摃",
-"抱著":"抱著",
-"抵御":"抵禦",
-"抵牾":"牴牾",
-"抵犢":"抵犢",
-"抵犢情深":"舐犢情深",
-"抵礪":"砥礪",
-"抵觸":"牴觸",
-"抹剎":"抹煞",
-"抹台":"抹檯",
-"抹干":"抹乾",
-"抺干":"抺乾",
-"抽傭":"抽佣",
-"抽干":"抽乾",
-"抽煙":"抽菸",
-"抽簽":"抽籤",
-"抽象":"抽象",
-"抿滅":"泯滅",
-"拂須":"拂鬚",
-"拆伙":"拆夥",
-"拆封":"拆封",
-"拆封後":"拆封後",
-"拆斗":"拆鬥",
-"拈須":"拈鬚",
-"拉煉":"拉鍊",
-"拉纖":"拉縴",
-"拉著":"拉著",
-"拉裡":"拉里",
-"拉面":"拉麵",
-"拋准":"拋準",
-"拌面":"拌麵",
-"拍出":"拍出",
-"拍出好戲":"拍齣好戲",
-"拍台":"拍台",
-"拍台拍凳":"拍檯拍凳",
-"拐子":"枴子",
-"拐杖":"枴杖",
-"拓本":"搨本",
-"拓朴":"拓樸",
-"拓朴學":"拓樸學",
-"拓碑":"搨碑",
-"拗別":"拗彆",
-"拘系":"拘繫",
-"拙壯":"茁壯",
-"拚命":"拚命",
-"拚命干":"拚命幹",
-"拚斗":"拚鬥",
-"招魂":"招魂",
-"招魂幡":"招魂旛",
-"拜佔":"拜佔",
-"拜佔庭":"拜占庭",
-"拜托":"拜託",
-"拭干":"拭乾",
-"拮據":"拮据",
-"拼一":"拼一",
-"拼一拼":"拚一拚",
-"拼上":"拼上",
-"拼上一拼":"拚上一拚",
-"拼個":"拚個",
-"拼力":"拚力",
-"拼勁":"拚勁",
-"拼命":"拚命",
-"拼得":"拼得",
-"拼得過":"拚得過",
-"拼搏":"拚搏",
-"拼政":"拚政",
-"拼斗":"拚鬥",
-"拼死":"拚死",
-"拼殺":"拚殺",
-"拼活":"拚活",
-"拼生":"拚生",
-"拼經":"拚經",
-"拼舍":"拚捨",
-"拿不":"拿不",
-"拿不准":"拿不準",
-"拿准":"拿準",
-"拿奸":"拿姦",
-"拿破":"拿破",
-"拿破侖":"拿破崙",
-"挑了":"挑了",
-"挑了只":"挑了隻",
-"挑大":"挑大",
-"挑大梁":"挑大樑",
-"挑斗":"挑鬥",
-"挖松":"挖鬆",
-"振蕩":"振盪",
-"挹郁":"挹鬱",
-"挺准":"挺準",
-"挺而":"鋌而",
-"挺象":"挺像",
-"挽歌":"輓歌",
-"挽發":"挽髮",
-"挽聯":"輓聯",
-"挽詞":"輓詞",
-"挽詩":"輓詩",
-"挽近":"輓近",
-"捂不":"捂不",
-"捂不住":"摀不住",
-"捂住":"摀住",
-"捂蓋":"摀蓋",
-"捆扎":"捆紮",
-"捉不":"捉不",
-"捉不准":"捉不準",
-"捉准":"捉準",
-"捉只":"捉隻",
-"捉奸":"捉姦",
-"捉狹":"促狹",
-"捋胡":"捋鬍",
-"捋虎":"捋虎",
-"捋虎須":"捋虎鬚",
-"捋須":"捋鬚",
-"捍御":"捍禦",
-"捍拒":"扞拒",
-"捍格":"扞格",
-"捏不":"捏不",
-"捏不准":"捏不準",
-"捏准":"捏準",
-"捏制":"捏製",
-"捭合":"捭闔",
-"捶台":"捶台",
-"捶台拍凳":"捶檯拍凳",
-"捻手":"撚手",
-"捻指":"撚指",
-"捻胡":"捻鬍",
-"捻花":"撚花",
-"捻錢":"撚錢",
-"捻須":"撚鬚",
-"捻香":"撚香",
-"掀卷":"掀捲",
-"掃卷":"掃捲",
-"掃描":"掃瞄",
-"掉發":"掉髮",
-"掉發出":"掉發出",
-"掉發現":"掉發現",
-"掉發現象":"掉髮現象",
-"掌握":"掌握",
-"排洩":"排泄",
-"排骨":"排骨",
-"排骨面":"排骨麵",
-"掘松":"掘鬆",
-"掙松":"掙鬆",
-"掛彩":"掛綵",
-"掛歷":"掛曆",
-"掛礙":"罣礙",
-"掠沖":"掠衝",
-"探知":"探知",
-"探知欲":"探知慾",
-"接斗":"接鬥",
-"控制":"控制",
-"控制欲":"控制慾",
-"控制面板":"控制面板",
-"推刨":"推鉋",
-"推情":"推情",
-"推情准理":"推情準理",
-"揉制":"揉製",
-"揉躪":"蹂躪",
-"揉面":"揉麵",
-"描涂":"描塗",
-"換只":"換隻",
-"換簽":"換籤",
-"握不":"握不",
-"握不准":"握不準",
-"握准":"握準",
-"握發":"握髮",
-"握著":"握著",
-"揩台":"揩檯",
-"揩干":"揩乾",
-"揪采":"揪採",
-"揭丑":"揭醜",
-"揮別":"揮別",
-"揮卷":"揮捲",
-"揮桿":"揮杆",
-"揮發":"揮發",
-"揮發油":"揮發油",
-"搏斗":"搏鬥",
-"搜羅":"蒐羅",
-"搜證":"蒐證",
-"搜購":"蒐購",
-"搜集":"蒐集",
-"摒息":"屏息",
-"摸不":"摸不",
-"摸不准":"摸不準",
-"摸准":"摸準",
-"摺雲":"摺云",
-"撞球":"撞球",
-"撞球台":"撞球檯",
-"撣子":"撢子",
-"撫恤":"撫卹",
-"撫須":"撫鬚",
-"撬松":"撬鬆",
-"撲卷":"撲捲",
-"撻虜":"韃虜",
-"擋御":"擋禦",
-"擐系":"擐繫",
-"擔仔":"擔仔",
-"擔仔面":"擔仔麵",
-"擔擔":"擔擔",
-"擔擔面":"擔擔麵",
-"擢發":"擢髮",
-"擢足":"濯足",
-"擦干":"擦乾",
-"擦幹":"擦乾",
-"擰干":"擰乾",
-"擰松":"擰鬆",
-"擺布":"擺佈",
-"擺蕩":"擺盪",
-"攏斷":"壟斷",
-"攝制":"攝製",
-"攝影":"攝影",
-"攝影後":"攝影後",
-"攝服":"懾服",
-"攝象":"攝像",
-"攣生":"孿生",
-"支干":"支幹",
-"支煙":"支菸",
-"支簽":"支籤",
-"收獲":"收穫",
-"收采":"收採",
-"改念":"改唸",
-"改采":"改採",
-"放姜":"放薑",
-"放松":"放鬆",
-"放著":"放著",
-"故布":"故佈",
-"故裡":"故里",
-"效尤":"傚尤",
-"效法":"傚法",
-"敖游":"敖遊",
-"教悔":"教誨",
-"教范":"教範",
-"敢再":"敢再",
-"敢再干杯":"敢再乾杯",
-"敢干":"敢幹",
-"敢念":"敢唸",
-"敢拼":"敢拚",
-"敢斗":"敢鬥",
-"敢松":"敢鬆",
-"散伙":"散夥",
-"散布":"散佈",
-"散署":"佈署",
-"敦厚":"惇厚",
-"敦敦":"惇惇",
-"敦朴":"惇樸",
-"敦睦":"惇睦",
-"敦篤":"惇篤",
-"敦誨":"惇誨",
-"敦謹":"惇謹",
-"整並":"整併",
-"整個":"整個",
-"整出":"整出",
-"整出劇":"整齣劇",
-"整出戲":"整齣戲",
-"整只":"整隻",
-"整周":"整週",
-"敷涂":"敷塗",
-"數只":"數隻",
-"數周":"數週",
-"數天":"數天",
-"數海":"數海",
-"數海裡":"數海里",
-"數裡":"數里",
-"文丑":"文醜",
-"文匯":"文匯",
-"文匯報":"文匯報",
-"文斗":"文鬥",
-"斑爛":"斑斕",
-"斗一":"鬥一",
-"斗一斗":"鬥一鬥",
-"斗上":"鬥上",
-"斗上一斗":"鬥上一鬥",
-"斗下":"斗下",
-"斗下去":"鬥下去",
-"斗不":"鬥不",
-"斗不過":"鬥不過",
-"斗久":"鬥久",
-"斗了":"鬥了",
-"斗了一斗":"鬥了一鬥",
-"斗什":"斗什",
-"斗什麼":"鬥什麼",
-"斗他":"鬥他",
-"斗他一斗":"鬥他一鬥",
-"斗你":"鬥你",
-"斗來":"鬥來",
-"斗個":"鬥個",
-"斗倒":"鬥倒",
-"斗傳":"鬥傳",
-"斗內":"斗內",
-"斗內力":"鬥內力",
-"斗出":"鬥出",
-"斗到":"鬥到",
-"斗劍":"鬥劍",
-"斗力":"鬥力",
-"斗勁":"鬥勁",
-"斗勇":"鬥勇",
-"斗勝":"鬥勝",
-"斗十":"斗十",
-"斗十余":"鬥十餘",
-"斗去":"鬥去",
-"斗口":"鬥口",
-"斗命":"鬥命",
-"斗嘴":"鬥嘴",
-"斗在":"鬥在",
-"斗垮":"鬥垮",
-"斗場":"鬥場",
-"斗士":"鬥士",
-"斗奇":"鬥奇",
-"斗她":"鬥她",
-"斗她一斗":"鬥她一鬥",
-"斗妍":"鬥妍",
-"斗完":"鬥完",
-"斗定":"斗定",
-"斗定力":"鬥定力",
-"斗將":"鬥將",
-"斗就":"斗就",
-"斗就斗":"鬥就鬥",
-"斗弄":"鬥弄",
-"斗得":"鬥得",
-"斗心":"鬥心",
-"斗忍":"鬥忍",
-"斗志":"鬥志",
-"斗快":"鬥快",
-"斗惡":"鬥惡",
-"斗意":"鬥意",
-"斗成":"鬥成",
-"斗我":"鬥我",
-"斗戰":"鬥戰",
-"斗批":"鬥批",
-"斗技":"鬥技",
-"斗招":"鬥招",
-"斗拱":"斗栱",
-"斗拳":"鬥拳",
-"斗掌":"鬥掌",
-"斗擻":"抖擻",
-"斗敗":"鬥敗",
-"斗敵":"鬥敵",
-"斗斗":"鬥鬥",
-"斗智":"鬥智",
-"斗本":"斗本",
-"斗本事":"鬥本事",
-"斗棋":"鬥棋",
-"斗機":"斗機",
-"斗機智":"鬥機智",
-"斗殺":"鬥殺",
-"斗毆":"鬥毆",
-"斗氣":"鬥氣",
-"斗法":"鬥法",
-"斗灼":"斗杓",
-"斗然":"陡然",
-"斗爭":"鬥爭",
-"斗牛":"鬥牛",
-"斗狗":"鬥狗",
-"斗狠":"鬥狠",
-"斗獸":"鬥獸",
-"斗畢":"鬥畢",
-"斗百":"斗百",
-"斗百余":"鬥百餘",
-"斗的":"斗的",
-"斗的過":"鬥的過",
-"斗眼":"鬥眼",
-"斗神":"鬥神",
-"斗私":"鬥私",
-"斗艦":"鬥艦",
-"斗草":"鬥草",
-"斗著":"鬥著",
-"斗藝":"鬥藝",
-"斗蟋":"斗蟋",
-"斗蟋蟀":"鬥蟋蟀",
-"斗角":"鬥角",
-"斗豔":"鬥豔",
-"斗贏":"鬥贏",
-"斗起":"鬥起",
-"斗趣":"鬥趣",
-"斗輸":"鬥輸",
-"斗過":"鬥過",
-"斗酒":"鬥酒",
-"斗雞":"鬥雞",
-"斗飲":"鬥飲",
-"斗鬧":"鬥鬧",
-"斗魂":"鬥魂",
-"斗魚":"鬥魚",
-"斗鵪":"斗鵪",
-"斗鵪鶉":"鬥鵪鶉",
-"斗麗":"鬥麗",
-"斗！":"鬥！",
-"料不":"料不",
-"料不准":"料不準",
-"料准":"料準",
-"斛斗":"觔斗",
-"斜沖":"斜衝",
-"斯裡":"斯里",
-"新征":"新徵",
-"新歷":"新曆",
-"新歷史":"新歷史",
-"新舄":"新潟",
-"斷發":"斷髮",
-"方便":"方便",
-"方便面":"方便麵",
-"於先":"於先",
-"於先生":"于先生",
-"於右":"於右",
-"於右任":"于右任",
-"於嗟":"于嗟",
-"於小":"於小",
-"於小姐":"于小姐",
-"於思":"於思",
-"於思於思":"于思于思",
-"於成":"於成",
-"於成龍":"于成龍",
-"於於":"于于",
-"於歸":"于歸",
-"於歸來":"於歸來",
-"於歸去":"於歸去",
-"於歸回":"於歸回",
-"於歸家":"於歸家",
-"於歸程":"於歸程",
-"於歸路":"於歸路",
-"於歸途":"於歸途",
-"於歸還":"於歸還",
-"於氏":"于氏",
-"於美":"於美",
-"於美人":"于美人",
-"於謙":"于謙",
-"於闐":"于闐",
-"於飛":"於飛",
-"於飛之樂":"于飛之樂",
-"施舍":"施捨",
-"旁征":"旁征",
-"旁征博引":"旁徵博引",
-"旁徨":"徬徨",
-"旁惶":"徬徨",
-"旁注":"旁註",
-"旅游":"旅遊",
-"旋卷":"旋捲",
-"旋松":"旋鬆",
-"旋渦":"漩渦",
-"旗幡":"旗旛",
-"旗桿":"旗杆",
-"既准":"既准",
-"既准且":"既準且",
-"既象":"既像",
-"日余":"日餘",
-"日制":"日製",
-"日制定":"日制定",
-"日制訂":"日制訂",
-"日升":"日昇",
-"日干":"日干",
-"日干夜干":"日幹夜幹",
-"日志":"日誌",
-"日本":"日本",
-"日本制":"日本製",
-"日本制定":"日本制定",
-"日本制訂":"日本制訂",
-"日歷":"日曆",
-"日歷史":"日歷史",
-"日歷年":"日歷年",
-"日歷經":"日歷經",
-"日游":"日遊",
-"早佔":"早佔",
-"早佔勿藥":"早占勿藥",
-"旬余":"旬餘",
-"旱煙":"旱菸",
-"昆侖":"崑崙",
-"昆山":"崑山",
-"昆曲":"崑曲",
-"昆腔":"崑腔",
-"明范":"明範",
-"昏瞑":"昏暝",
-"易斗":"易斗",
-"易斗的":"易鬥的",
-"易松":"易鬆",
-"星歷":"星曆",
-"星歷代":"星歷代",
-"星歷任":"星歷任",
-"星歷史":"星歷史",
-"星歷盡":"星歷盡",
-"星歷經":"星歷經",
-"星歷險":"星歷險",
-"星辰":"星辰",
-"星辰表":"星辰錶",
-"映象":"映像",
-"春卷":"春捲",
-"春游":"春遊",
-"是丑":"是醜",
-"是干":"是干",
-"是干的":"是乾的",
-"是干脆":"是乾脆",
-"是白":"是白",
-"是白干":"是白幹",
-"時干":"時干",
-"時干下":"時幹下",
-"時憲":"時憲",
-"時憲歷":"時憲曆",
-"晉升":"晉陞",
-"普裡":"普里",
-"景致":"景緻",
-"智斗":"智鬥",
-"晾干":"晾乾",
-"暗倫":"闇倫",
-"暗冥":"闇冥",
-"暗劣":"闇劣",
-"暗弱":"闇弱",
-"暗斗":"暗鬥",
-"暗然":"闇然",
-"暗藏":"暗藏",
-"暗藏著":"暗藏著",
-"暗贊":"暗讚",
-"暝目":"瞑目",
-"暢游":"暢遊",
-"暫松":"暫鬆",
-"暴卷":"暴捲",
-"暴沖":"暴衝",
-"暴珍":"暴珍",
-"暴珍天物":"暴殄天物",
-"曬干":"曬乾",
-"曬谷":"曬穀",
-"曲卷":"曲捲",
-"曲薛":"麴薛",
-"曲酒":"麴酒",
-"更丑":"更醜",
-"更准":"更準",
-"更加":"更加",
-"更加丑":"更加醜",
-"更咸":"更鹹",
-"更松":"更鬆",
-"更要":"更要",
-"更要丑":"更要醜",
-"更象":"更像",
-"書台":"書檯",
-"書童":"書僮",
-"書簽":"書籤",
-"最丑":"最醜",
-"最准":"最準",
-"會干":"會幹",
-"會干擾":"會干擾",
-"會干杯":"會乾杯",
-"會干涸":"會乾涸",
-"會干淨":"會乾淨",
-"會干脆":"會乾脆",
-"會干裂":"會乾裂",
-"會斗":"會鬥",
-"會象":"會像",
-"會象棋":"會象棋",
-"月余":"月餘",
-"月歷":"月曆",
-"有余":"有餘",
-"有出":"有出",
-"有出好戲":"有齣好戲",
-"有得":"有得",
-"有得斗":"有得鬥",
-"有感":"有感",
-"有感而雲":"有感而云",
-"有象":"有像",
-"有象征":"有象徵",
-"有關":"有關",
-"有關系":"有關係",
-"有點":"有點",
-"有點丑":"有點醜",
-"服贗":"服膺",
-"朝干":"朝乾",
-"木制":"木製",
-"木梁":"木樑",
-"木簽":"木籤",
-"木納":"木訥",
-"未干":"未乾",
-"未愈":"未癒",
-"未斗":"未鬥",
-"未松":"未松",
-"未松下":"未鬆下",
-"未松口":"未鬆口",
-"未端":"末端",
-"本周":"本週",
-"朱喻":"硃喻",
-"朱批":"硃批",
-"朱砂":"硃砂",
-"朱磯":"硃磯",
-"朱筆":"硃筆",
-"朴修":"朴修",
-"朴修斯":"樸修斯",
-"朴學":"樸學",
-"朴實":"樸實",
-"朴拙":"樸拙",
-"朴玉":"璞玉",
-"朴直":"樸直",
-"朴素":"樸素",
-"朴質":"樸質",
-"朴野":"樸野",
-"李鐘":"李鐘",
-"李鐘郁":"李鍾郁",
-"材干":"材幹",
-"村落":"村落",
-"束發":"束髮",
-"杯干":"杯乾",
-"杯白":"杯白",
-"杯白干":"杯白乾",
-"杯面":"杯麵",
-"東升":"東昇",
-"東周":"東周",
-"東岳":"東嶽",
-"東東":"東東",
-"東東裡面":"東東裡面",
-"東沖":"東衝",
-"東涂":"東塗",
-"東游":"東遊",
-"東湧":"東涌",
-"東裡":"東里",
-"松一":"松一",
-"松一下":"鬆一下",
-"松一些":"鬆一些",
-"松一口氣":"鬆一口氣",
-"松一松":"鬆一鬆",
-"松下":"松下",
-"松下一口":"鬆下一口",
-"松下了":"鬆下了",
-"松下來":"鬆下來",
-"松不":"松不",
-"松不松":"鬆不鬆",
-"松不緊":"鬆不緊",
-"松不開":"鬆不開",
-"松了":"鬆了",
-"松了松":"鬆了鬆",
-"松些":"鬆些",
-"松出":"松出",
-"松出一":"鬆出一",
-"松出口":"鬆出口",
-"松勁":"鬆勁",
-"松動":"鬆動",
-"松化":"鬆化",
-"松去":"鬆去",
-"松口":"鬆口",
-"松土":"鬆土",
-"松垮":"鬆垮",
-"松塌":"鬆塌",
-"松寬":"鬆寬",
-"松干":"松幹",
-"松弛":"鬆弛",
-"松得":"松得",
-"松得多":"鬆得多",
-"松快":"鬆快",
-"松懈":"鬆懈",
-"松手":"鬆手",
-"松掉":"鬆掉",
-"松放":"鬆放",
-"松散":"鬆散",
-"松晃":"鬆晃",
-"松松":"鬆鬆",
-"松柔":"鬆柔",
-"松毛":"松毛",
-"松毛松翼":"鬆毛鬆翼",
-"松氣":"鬆氣",
-"松油":"松油",
-"松油管":"鬆油管",
-"松油門":"鬆油門",
-"松泛":"鬆泛",
-"松泥":"鬆泥",
-"松浮":"鬆浮",
-"松球":"松毬",
-"松甩":"鬆甩",
-"松筋":"鬆筋",
-"松糕":"鬆糕",
-"松綁":"鬆綁",
-"松緊":"鬆緊",
-"松線":"鬆線",
-"松緩":"鬆緩",
-"松繩":"鬆繩",
-"松肘":"鬆肘",
-"松肩":"鬆肩",
-"松脆":"鬆脆",
-"松脫":"鬆脫",
-"松腰":"鬆腰",
-"松蛋":"鬆蛋",
-"松解":"鬆解",
-"松起":"鬆起",
-"松軟":"鬆軟",
-"松通":"鬆通",
-"松開":"鬆開",
-"松閒":"鬆閒",
-"松餅":"鬆餅",
-"松馳":"鬆弛",
-"松骨":"鬆骨",
-"松齒":"鬆齒",
-"板剉":"板銼",
-"果干":"果乾",
-"果系":"果係",
-"果系統":"果系統",
-"枝干":"枝幹",
-"枝芽":"枝枒",
-"枯干":"枯幹",
-"架梁":"架樑",
-"某只":"某隻",
-"染發":"染髮",
-"柔發":"柔髮",
-"柱梁":"柱樑",
-"栓住":"拴住",
-"栗人":"慄人",
-"栗冽":"慄冽",
-"栗栗":"慄慄",
-"栗然":"慄然",
-"栗聲":"慄聲",
-"栗驚":"慄驚",
-"校准":"校準",
-"株象":"株像",
-"核復":"核覆",
-"根煙":"根菸",
-"格斗":"格鬥",
-"格范":"格範",
-"格裡":"格里",
-"格裡歷":"格里曆",
-"桅桿":"桅杆",
-"案幾":"案几",
-"桌面":"桌面",
-"桑干":"桑乾",
-"桿面":"桿麵",
-"梁上":"樑上",
-"梁下":"樑下",
-"梁折":"樑折",
-"梁柱":"樑柱",
-"梁脊":"樑脊",
-"梗直":"鯁直",
-"梯沖":"梯衝",
-"械斗":"械鬥",
-"梳發":"梳髮",
-"棄舍":"棄捨",
-"棉制":"棉製",
-"棕發":"棕髮",
-"棟梁":"棟樑",
-"棵象":"棵像",
-"植發":"植髮",
-"業余":"業餘",
-"極丑":"極醜",
-"極准":"極準",
-"極干":"極乾",
-"極待":"極待",
-"極待救":"亟待救",
-"極待解":"亟待解",
-"極極":"亟亟",
-"極象":"極像",
-"榔頭":"鎯頭",
-"榨取":"搾取",
-"榨干":"榨乾",
-"榨渣":"搾渣",
-"榨甘":"榨甘",
-"榨甘蔗":"搾甘蔗",
-"榴連":"榴槤",
-"構陷":"搆陷",
-"槌打":"搥打",
-"槌背":"搥背",
-"槌胸":"搥胸",
-"槌頭":"鎚頭",
-"槌鼓":"搥鼓",
-"標准":"標準",
-"標志":"標誌",
-"標桿":"標竿",
-"標注":"標註",
-"標簽":"標籤",
-"標致":"標緻",
-"樞鈕":"樞紐",
-"模制":"模製",
-"模范":"模範",
-"樣丑":"樣醜",
-"樣干":"樣幹",
-"樣范":"樣範",
-"樵采":"樵採",
-"樹干":"樹幹",
-"橋梁":"橋樑",
-"橘梗":"桔梗",
-"機器":"機器",
-"機器壓制":"機器壓製",
-"機抒":"機杼",
-"橫卷":"橫捲",
-"橫征":"橫徵",
-"橫梁":"橫樑",
-"橫沖":"橫衝",
-"櫃台":"櫃檯",
-"欄桿":"欄杆",
-"權力":"權力",
-"權力欲":"權力慾",
-"權斗":"權鬥",
-"欠准":"欠準",
-"欲令":"欲令",
-"欲令智昏":"慾令智昏",
-"欲壑":"慾壑",
-"欲女":"慾女",
-"欲孽":"慾孽",
-"欲念":"慾念",
-"欲望":"慾望",
-"欲求":"欲求",
-"欲求不滿":"慾求不滿",
-"欲海":"慾海",
-"欲火":"慾火",
-"欲舍":"欲捨",
-"欲障":"慾障",
-"欷籲":"欷吁",
-"欺蒙":"欺矇",
-"欽崎":"嶔崎",
-"歉咎":"歉疚",
-"歐游":"歐遊",
-"正正":"正正",
-"正正當當":"正正當當",
-"正象":"正像",
-"此丑":"此醜",
-"步步":"步步",
-"步步高升":"步步高陞",
-"武斗":"武鬥",
-"歷數":"曆數",
-"歷日":"曆日",
-"歷書":"曆書",
-"歷月":"曆月",
-"歷法":"曆法",
-"歷落":"曆落",
-"歷象":"曆象",
-"歸並":"歸併",
-"歸並不":"歸並不",
-"歸並且":"歸並且",
-"歸並非":"歸並非",
-"歸依":"歸依",
-"歸依佛門":"皈依佛門",
-"歸依我佛":"皈依我佛",
-"歸朴":"歸璞",
-"歸真":"歸真",
-"歸真反朴":"歸真反璞",
-"歸究":"歸咎",
-"歸裡":"歸里",
-"死斗":"死鬥",
-"殊勳":"殊勛",
-"殘余":"殘餘",
-"殘干":"殘幹",
-"殘廢":"殘廢",
-"殘廢後":"殘廢後",
-"殘肴":"殘餚",
-"殷勤":"慇勤",
-"毀謗":"譭謗",
-"毆斗":"毆鬥",
-"母丑":"母醜",
-"母後":"母后",
-"母范":"母範",
-"每只":"每隻",
-"每周":"每週",
-"比拼":"比拚",
-"比斗":"比鬥",
-"毛制":"毛製",
-"毛姜":"毛薑",
-"毛發":"毛髮",
-"毛裡":"毛裡",
-"毛裡求斯":"毛里求斯",
-"毛裡裘斯":"毛里裘斯",
-"毛須":"毛鬚",
-"毫厘":"毫釐",
-"毫發":"毫髮",
-"民制":"民製",
-"氣一":"氣一",
-"氣一沖":"氣一衝",
-"氣慨":"氣概",
-"氣象":"氣象",
-"氣郁":"氣鬱",
-"氤郁":"氤鬱",
-"水一":"水一",
-"水一沖":"水一衝",
-"水再":"水再",
-"水再沖":"水再衝",
-"水准":"水準",
-"水又":"水又",
-"水又沖":"水又衝",
-"水已":"水已",
-"水已干":"水已乾",
-"水已沖":"水已衝",
-"水干":"水乾",
-"水干掉":"水乾掉",
-"水干盡":"水乾盡",
-"水急":"水急",
-"水急沖":"水急衝",
-"水才":"水才",
-"水才干":"水才乾",
-"水沖":"水沖",
-"水泄":"水泄",
-"水泄不通":"水洩不通",
-"水清":"水清",
-"水清可見":"水清可鑑",
-"水濂":"水簾",
-"水谷":"水谷",
-"水谷之海":"水穀之海",
-"水道":"水道",
-"永志":"永誌",
-"永歷":"永曆",
-"求知":"求知",
-"求知欲":"求知慾",
-"求簽":"求籤",
-"污蔑":"污衊",
-"汨沒":"汩沒",
-"決心":"決心",
-"決心干":"決心幹",
-"決擇":"抉擇",
-"決斗":"決鬥",
-"沈淀":"沈澱",
-"沈郁":"沈鬱",
-"沈陽":"瀋陽",
-"沉淀":"沉澱",
-"沉郁":"沉鬱",
-"沒准":"沒準",
-"沒干":"沒幹",
-"沒干淨":"沒乾淨",
-"沒干頭":"沒幹頭",
-"沒念":"沒唸",
-"沒松":"沒松",
-"沒松下":"沒鬆下",
-"沒松口":"沒鬆口",
-"沒松手":"沒鬆手",
-"沒白":"沒白",
-"沒白干":"沒白幹",
-"沒輒":"沒轍",
-"沖上":"沖上",
-"沖上來":"衝上來",
-"沖下":"衝下",
-"沖不":"沖不",
-"沖不上":"衝不上",
-"沖不下":"衝不下",
-"沖不入":"衝不入",
-"沖不出":"衝不出",
-"沖不破":"衝不破",
-"沖不進":"衝不進",
-"沖不過":"衝不過",
-"沖不開":"衝不開",
-"沖亂":"衝亂",
-"沖了":"沖了",
-"沖了上":"衝了上",
-"沖了下":"衝了下",
-"沖了來":"衝了來",
-"沖了入":"衝了入",
-"沖了出":"衝了出",
-"沖了進":"衝了進",
-"沖了過":"衝了過",
-"沖他":"衝他",
-"沖你":"衝你",
-"沖來":"衝來",
-"沖倒":"衝倒",
-"沖克":"衝克",
-"沖入":"衝入",
-"沖冠":"衝冠",
-"沖出":"衝出",
-"沖到":"衝到",
-"沖刺":"衝刺",
-"沖剋":"衝剋",
-"沖剌":"衝剌",
-"沖前":"衝前",
-"沖劍":"衝劍",
-"沖力":"衝力",
-"沖勁":"衝勁",
-"沖動":"衝動",
-"沖勢":"衝勢",
-"沖去":"衝去",
-"沖口":"衝口",
-"沖向":"衝向",
-"沖向前":"衝向前",
-"沖呀":"衝呀",
-"沖咱":"衝咱",
-"沖啊":"衝啊",
-"沖回":"衝回",
-"沖在":"沖在",
-"沖在下":"衝在下",
-"沖在前":"衝在前",
-"沖在最前":"衝在最前",
-"沖堅":"衝堅",
-"沖墜":"衝墜",
-"沖天":"衝天",
-"沖奔":"衝奔",
-"沖她":"衝她",
-"沖將":"衝將",
-"沖往":"衝往",
-"沖得":"沖得",
-"沖得入":"衝得入",
-"沖得出":"衝得出",
-"沖得進":"衝得進",
-"沖得過":"衝得過",
-"沖心":"衝心",
-"沖我":"衝我",
-"沖拳":"衝拳",
-"沖掠":"衝掠",
-"沖撞":"衝撞",
-"沖撲":"衝撲",
-"沖擊":"衝擊",
-"沖斷":"衝斷",
-"沖殺":"衝殺",
-"沖決":"衝決",
-"沖沖":"沖沖",
-"沖沖沖":"衝衝衝",
-"沖波":"衝波",
-"沖浪":"衝浪",
-"沖激":"衝激",
-"沖然":"衝然",
-"沖犯":"衝犯",
-"沖盹":"沖盹",
-"沖盹兒":"衝盹兒",
-"沖破":"衝破",
-"沖程":"衝程",
-"沖穴":"衝穴",
-"沖突":"衝突",
-"沖脈":"衝脈",
-"沖自":"沖自",
-"沖自己":"衝自己",
-"沖至":"衝至",
-"沖落":"衝落",
-"沖著":"衝著",
-"沖要":"衝要",
-"沖起":"衝起",
-"沖越":"衝越",
-"沖跑":"衝跑",
-"沖軍":"衝軍",
-"沖近":"衝近",
-"沖進":"衝進",
-"沖過":"衝過",
-"沖量":"衝量",
-"沖鋒":"衝鋒",
-"沖鋒槍":"衝鋒槍",
-"沖錯":"衝錯",
-"沖門":"衝門",
-"沖開":"衝開",
-"沖闖":"衝闖",
-"沖陣":"衝陣",
-"沖陷":"衝陷",
-"沖頭":"沖頭",
-"沖頭陣":"衝頭陣",
-"沖風":"衝風",
-"沖馳":"衝馳",
-"沖！":"衝！",
-"河落":"河落",
-"河落海干":"河落海乾",
-"油面":"油麵",
-"治愈":"治癒",
-"治疔":"治療",
-"沾汙":"玷汙",
-"沾污":"玷汙",
-"沾辱":"玷辱",
-"沿著":"沿著",
-"泛游":"泛遊",
-"泛濫":"氾濫",
-"泡制":"泡製",
-"泡面":"泡麵",
-"波裡":"波里",
-"泥涂":"泥塗",
-"泥象":"泥像",
-"注:":"註:",
-"注上":"註上",
-"注冊":"註冊",
-"注失":"註失",
-"注明":"註明",
-"注標":"註標",
-"注生":"注生",
-"注生娘娘":"註生娘娘",
-"注疏":"註疏",
-"注腳":"註腳",
-"注解":"註解",
-"注記":"註記",
-"注譯":"註譯",
-"注釋":"註釋",
-"注銷":"註銷",
-"注：":"註：",
-"泰歷":"泰曆",
-"泱郁":"泱鬱",
-"洗手":"洗手",
-"洗手不干":"洗手不幹",
-"洗發":"洗髮",
-"洗發動機":"洗發動機",
-"洩欲":"洩慾",
-"洩殖":"泄殖",
-"洩沓":"泄沓",
-"洩洩":"泄泄",
-"洪范":"洪範",
-"洲杯":"洲盃",
-"洲游":"洲遊",
-"活干":"活幹",
-"活象":"活像",
-"流卷":"流捲",
-"流干":"流乾",
-"流風":"流風",
-"流風余":"流風餘",
-"浚深":"濬深",
-"浪一":"浪一",
-"浪一沖":"浪一衝",
-"浪卷":"浪捲",
-"浪又":"浪又",
-"浪又沖":"浪又衝",
-"浪沖":"浪衝",
-"浪琴":"浪琴",
-"浪琴表":"浪琴錶",
-"浮松":"浮鬆",
-"浮簽":"浮籤",
-"海函":"海涵",
-"海干":"海乾",
-"浸制":"浸製",
-"涂上":"塗上",
-"涂了":"塗了",
-"涂來":"涂來",
-"涂來涂去":"塗來塗去",
-"涂到":"塗到",
-"涂制":"塗製",
-"涂去":"塗去",
-"涂在":"塗在",
-"涂地":"塗地",
-"涂好":"塗好",
-"涂寫":"塗寫",
-"涂層":"塗層",
-"涂布":"塗布",
-"涂得":"塗得",
-"涂成":"塗成",
-"涂抹":"塗抹",
-"涂掉":"塗掉",
-"涂改":"塗改",
-"涂料":"塗料",
-"涂有":"塗有",
-"涂油":"塗油",
-"涂涂":"塗塗",
-"涂消":"塗消",
-"涂滿":"塗滿",
-"涂漆":"塗漆",
-"涂潭":"塗潭",
-"涂炭":"塗炭",
-"涂竄":"塗竄",
-"涂粉":"塗粉",
-"涂脂":"塗脂",
-"涂膠":"塗膠",
-"涂色":"塗色",
-"涂著":"塗著",
-"涂藥":"塗藥",
-"涂蠟":"塗蠟",
-"涂裝":"塗裝",
-"涂覆":"塗覆",
-"涂說":"塗說",
-"涂過":"塗過",
-"涂銷":"塗銷",
-"涂飾":"塗飾",
-"涂鴉":"塗鴉",
-"涂鴨":"塗鴨",
-"涂黑":"塗黑",
-"涅盤":"涅槃",
-"消費":"消費",
-"消費欲":"消費慾",
-"液晶":"液晶",
-"液晶表":"液晶錶",
-"涼席":"涼蓆",
-"涼面":"涼麵",
-"淀粉":"澱粉",
-"淑范":"淑範",
-"淒息":"棲息",
-"淒戚":"悽慼",
-"淚干":"淚乾",
-"淚難":"淚難",
-"淚難干":"淚難乾",
-"淨余":"淨餘",
-"淨幾":"淨几",
-"淫欲":"淫慾",
-"淬厲":"淬礪",
-"淳於":"淳于",
-"淳朴":"淳樸",
-"混肴":"混淆",
-"清掃":"清掃",
-"清掃夫":"清掃伕",
-"清淅":"清晰",
-"清理":"清理",
-"清道":"清道",
-"清道夫":"清道伕",
-"減征":"減徵",
-"渠沖":"渠衝",
-"測不":"測不",
-"測不准":"測不準",
-"測准":"測準",
-"港制":"港製",
-"游中":"游中",
-"游中國":"遊中國",
-"游亞":"游亞",
-"游亞洲":"遊亞洲",
-"游人":"遊人",
-"游仙":"遊仙",
-"游伴":"遊伴",
-"游俠":"遊俠",
-"游兵":"遊兵",
-"游冶":"遊冶",
-"游刃":"遊刃",
-"游北":"游北",
-"游北美":"遊北美",
-"游台":"游台",
-"游台灣":"遊台灣",
-"游唱":"遊唱",
-"游園":"遊園",
-"游子":"遊子",
-"游學":"遊學",
-"游客":"遊客",
-"游宦":"遊宦",
-"游山":"遊山",
-"游巡":"遊巡",
-"游廊":"遊廊",
-"游必":"游必",
-"游必有方":"遊必有方",
-"游惰":"遊惰",
-"游憩":"遊憩",
-"游戲":"遊戲",
-"游斗":"游鬥",
-"游方":"遊方",
-"游星":"遊星",
-"游樂":"遊樂",
-"游歐":"游歐",
-"游歐洲":"遊歐洲",
-"游歷":"遊歷",
-"游湖":"遊湖",
-"游澳":"游澳",
-"游澳洲":"遊澳洲",
-"游牧":"遊牧",
-"游狎":"遊狎",
-"游獵":"遊獵",
-"游玩":"遊玩",
-"游目":"遊目",
-"游美":"游美",
-"游美洲":"遊美洲",
-"游興":"遊興",
-"游船":"遊船",
-"游艇":"遊艇",
-"游蕩":"遊蕩",
-"游藝":"遊藝",
-"游蜂":"遊蜂",
-"游行":"遊行",
-"游街":"遊街",
-"游覽":"遊覽",
-"游觀":"遊觀",
-"游記":"遊記",
-"游說":"遊說",
-"游豫":"遊豫",
-"游走":"遊走",
-"游蹤":"遊蹤",
-"游軍":"遊軍",
-"游逛":"遊逛",
-"游遍":"遊遍",
-"游錯":"遊錯",
-"游非":"游非",
-"游非洲":"遊非洲",
-"游香":"游香",
-"游香港":"遊香港",
-"游騎":"遊騎",
-"游魂":"遊魂",
-"渾函":"渾涵",
-"湧卷":"湧捲",
-"湮郁":"湮鬱",
-"湯團":"湯糰",
-"湯面":"湯麵",
-"溜達":"溜躂",
-"溜須":"溜須",
-"溜須拍馬":"溜鬚拍馬",
-"溪壑":"谿壑",
-"溶岩":"熔岩",
-"溶接":"熔接",
-"溶爐":"熔爐",
-"溶銷":"熔銷",
-"溶鑄":"熔鑄",
-"滴干":"滴乾",
-"滾卷":"滾捲",
-"滿布":"滿佈",
-"漓水":"灕水",
-"漓江":"灕江",
-"漓湘":"灕湘",
-"漢周":"漢周",
-"漫游":"漫遊",
-"漬已":"漬已",
-"漬已干":"漬已乾",
-"漸愈":"漸癒",
-"漸松":"漸鬆",
-"漿劃":"槳劃",
-"潤發":"潤髮",
-"潦闊":"遼闊",
-"潮卷":"潮捲",
-"澆制":"澆製",
-"激干":"激幹",
-"激斗":"激鬥",
-"激蕩":"激盪",
-"濃發":"濃髮",
-"濃纖":"穠纖",
-"濕發":"濕髮",
-"濛松":"濛鬆",
-"濫征":"濫徵",
-"濫芋":"濫芋",
-"濫芋充數":"濫竽充數",
-"瀝干":"瀝乾",
-"灌制":"灌製",
-"灘涂":"灘塗",
-"火折":"火摺",
-"火拼":"火拚",
-"火簽":"火籤",
-"灰發":"灰髮",
-"灰盡":"灰燼",
-"灰胡":"灰鬍",
-"灰蒙":"灰濛",
-"灰郁":"灰鬱",
-"灸熱":"炙熱",
-"炒干":"炒乾",
-"炒面":"炒麵",
-"炮制":"炮製",
-"炸毀":"炸燬",
-"為准":"為準",
-"為范":"為範",
-"烏丘":"烏坵",
-"烏發":"烏髮",
-"烏裡":"烏里",
-"烏龍":"烏龍",
-"烏龍面":"烏龍麵",
-"烘制":"烘製",
-"烘干":"烘乾",
-"烘熏":"烘燻",
-"烤干":"烤乾",
-"烹制":"烹製",
-"焊接":"銲接",
-"焙干":"焙乾",
-"焚毀":"焚燬",
-"無余":"無餘",
-"無梁":"無樑",
-"無欲":"無慾",
-"無煙":"無煙",
-"無煙室":"無菸室",
-"無煙環境":"無菸環境",
-"無煙餐廳":"無菸餐廳",
-"無釐":"無釐",
-"無釐頭":"無厘頭",
-"焦干":"焦乾",
-"然斗":"然斗",
-"然斗不":"然鬥不",
-"然無":"然無",
-"然無幸":"然無倖",
-"然象":"然像",
-"煉制":"煉製",
-"煉子":"鍊子",
-"煉鎖":"鍊鎖",
-"煎制":"煎製",
-"煙卷":"菸捲",
-"煙商":"菸商",
-"煙客":"菸客",
-"煙槍":"菸槍",
-"煙民":"菸民",
-"煙沒":"湮沒",
-"煙灰":"菸灰",
-"煙焦":"煙焦",
-"煙焦油":"菸焦油",
-"煙熏":"煙燻",
-"煙癮":"菸癮",
-"煙草":"菸草",
-"煙酒":"菸酒",
-"煙頭":"菸頭",
-"煙鬼":"菸鬼",
-"煙齡":"菸齡",
-"照准":"照準",
-"照干":"照幹",
-"照象":"照像",
-"煨干":"煨乾",
-"煩郁":"煩鬱",
-"煮面":"煮麵",
-"熏制":"燻製",
-"熏煙":"燻煙",
-"熏煮":"燻煮",
-"熏熏":"燻燻",
-"熏熟":"燻熟",
-"熏熱":"燻熱",
-"熏肉":"燻肉",
-"熏蒸":"燻蒸",
-"熏豬":"燻豬",
-"熏鴨":"燻鴨",
-"熏鵝":"燻鵝",
-"熏黑":"燻黑",
-"熒光":"螢光",
-"熒屏":"螢屏",
-"熒幕":"螢幕",
-"熒郁":"熒鬱",
-"熔毀":"熔燬",
-"熨干":"熨乾",
-"熬制":"熬製",
-"燈彩":"燈綵",
-"燎發":"燎髮",
-"燒制":"燒製",
-"燒干":"燒乾",
-"燒毀":"燒燬",
-"燕游":"燕遊",
-"燕燕":"燕燕",
-"燕燕於飛":"燕燕于飛",
-"燙發":"燙髮",
-"燙面":"燙麵",
-"營干":"營幹",
-"燴制":"燴製",
-"燴干":"燴乾",
-"燼余":"燼餘",
-"爭斗":"爭鬥",
-"片姜":"片薑",
-"片紙":"片紙",
-"片紙只字":"片紙隻字",
-"片言":"片言",
-"片言只語":"片言隻語",
-"牙簽":"牙籤",
-"牛只":"牛隻",
-"牛柳":"牛柳",
-"牛柳面":"牛柳麵",
-"物志":"物誌",
-"物欲":"物慾",
-"牲干":"牲干",
-"牲干下":"牲幹下",
-"特制":"特製",
-"特征":"特徵",
-"牽一":"牽一",
-"牽一發":"牽一髮",
-"牽系":"牽繫",
-"犁牛":"犛牛",
-"犬只":"犬隻",
-"狂卷":"狂捲",
-"狂征":"狂徵",
-"狂沖":"狂衝",
-"狂贊":"狂讚",
-"狠、":"狠、",
-"狠、准":"狠、準",
-"狠且":"狠且",
-"狠且准":"狠且準",
-"狠又":"狠又",
-"狠又准":"狠又準",
-"狠干":"狠幹",
-"狠斗":"狠鬥",
-"狼借":"狼藉",
-"猛卷":"猛捲",
-"猛干":"猛幹",
-"猛斗":"猛鬥",
-"猛沖":"猛衝",
-"猜准":"猜準",
-"猶太":"猶太",
-"猶太歷":"猶太曆",
-"猷豫":"猶豫",
-"獎杯":"獎盃",
-"獨斗":"獨鬥",
-"獨游":"獨遊",
-"獨辟":"獨闢",
-"獨辟蹊徑":"獨闢蹊徑",
-"獨鐘":"獨鍾",
-"獸奸":"獸姦",
-"獸欲":"獸慾",
-"獻丑":"獻醜",
-"玉佩":"玉珮",
-"玉制":"玉製",
-"王後":"王后",
-"玲朧":"玲瓏",
-"玻裡":"玻裡",
-"玻裡尼西":"玻里尼西",
-"現丑":"現醜",
-"現象":"現象",
-"球台":"球檯",
-"球果":"毬果",
-"球桿":"球杆",
-"球童":"球僮",
-"琅嬛":"瑯嬛",
-"琅當":"琅璫",
-"理發":"理髮",
-"琢制":"琢製",
-"琳琅":"琳瑯",
-"琺琅":"琺瑯",
-"瑞哲":"睿哲",
-"瑞智":"睿智",
-"瑤簽":"瑤籤",
-"璀燦":"璀璨",
-"璨爛":"燦爛",
-"環佩":"環珮",
-"環游":"環遊",
-"瓶白":"瓶白",
-"瓶白干":"瓶白乾",
-"瓷制":"瓷製",
-"甘乃":"甘乃",
-"甘乃迪":"甘迺迪",
-"甘薯":"甘藷",
-"甚丑":"甚醜",
-"甚准":"甚準",
-"甚松":"甚鬆",
-"甚麼":"甚麼",
-"甚麼都干":"甚麼都幹",
-"甜面":"甜面",
-"甜面醬":"甜麵醬",
-"生克":"生剋",
-"生力":"生力",
-"生力面":"生力麵",
-"生姜":"生薑",
-"生得":"生得",
-"生得丑":"生得醜",
-"生理":"生理",
-"生發":"生發",
-"生發水":"生髮水",
-"生發藥":"生髮藥",
-"生長":"生長",
-"甩發":"甩髮",
-"甲胄":"甲冑",
-"男用":"男用",
-"男用表":"男用錶",
-"留發":"留髮",
-"畛懷":"軫懷",
-"畝余":"畝餘",
-"略干":"略乾",
-"略松":"略鬆",
-"畦徑":"畦逕",
-"番薯":"蕃薯",
-"畫象":"畫像",
-"畫餅":"畫餅",
-"畫餅充飢":"畫餅充饑",
-"當一":"當一",
-"當一聲":"噹一聲",
-"當其":"當其",
-"當其沖":"當其衝",
-"當准":"當準",
-"當周":"當週",
-"當周潤發":"當周潤發",
-"當啷":"噹啷",
-"當局":"當局",
-"當年":"當年",
-"當然":"當然",
-"當然羅！":"當然囉！",
-"當當":"噹噹",
-"當當當":"噹噹噹",
-"當的":"當的",
-"當的一聲":"噹的一聲",
-"當的一響":"噹的一響",
-"當郎":"噹啷",
-"疆弩":"疆弩",
-"疆弩之末":"彊弩之末",
-"疏松":"疏鬆",
-"疏浚":"疏濬",
-"疑系":"疑係",
-"疔傷":"療傷",
-"疲困":"疲睏",
-"疾卷":"疾捲",
-"疾沖":"疾衝",
-"病入":"病入",
-"病入膏盲":"病入膏肓",
-"病征":"病徵",
-"病愈":"病癒",
-"病難":"病難",
-"病難愈":"病難癒",
-"症結":"癥結",
-"痊愈":"痊癒",
-"痙孿":"痙攣",
-"發乳":"髮乳",
-"發亂":"發亂",
-"發亂釵橫":"髮亂釵橫",
-"發冠":"髮冠",
-"發卷":"髮捲",
-"發型":"髮型",
-"發夾":"髮夾",
-"發套":"髮套",
-"發如":"發如",
-"發如飛蓬":"髮如飛蓬",
-"發妻":"髮妻",
-"發姐":"髮姐",
-"發布":"發佈",
-"發帶":"髮帶",
-"發干":"發乾",
-"發廊":"髮廊",
-"發式":"髮式",
-"發指":"髮指",
-"發指令":"發指令",
-"發挽":"發挽",
-"發挽雙髻":"髮挽雙髻",
-"發束":"髮束",
-"發根":"髮根",
-"發梢":"髮梢",
-"發梳":"髮梳",
-"發油":"髮油",
-"發皓":"髮皓",
-"發箋":"髮箋",
-"發箍":"髮箍",
-"發簪":"髮簪",
-"發絲":"髮絲",
-"發纓":"髮纓",
-"發膚":"髮膚",
-"發膠":"髮膠",
-"發色":"髮色",
-"發菜":"髮菜",
-"發蠟":"髮蠟",
-"發表":"發表",
-"發表欲":"發表慾",
-"發質":"髮質",
-"發辮":"髮辮",
-"發醅":"醱醅",
-"發針":"髮針",
-"發釵":"髮釵",
-"發長":"髮長",
-"發長音":"發長音",
-"發際":"髮際",
-"發雕":"髮雕",
-"發霜":"髮霜",
-"發面":"發麵",
-"發須":"發須",
-"發須俱":"髮鬚俱",
-"發須已":"髮鬚已",
-"發須斑":"髮鬚斑",
-"發須皆":"髮鬚皆",
-"發須都":"髮鬚都",
-"發飾":"髮飾",
-"發香":"髮香",
-"發髯":"髮髯",
-"發髻":"髮髻",
-"發鬢":"髮鬢",
-"發黴":"發霉",
-"白淅":"白皙",
-"白澈":"白皙",
-"白發":"白髮",
-"白白":"白白",
-"白白干":"白白幹",
-"白術":"白朮",
-"白面":"白面",
-"白面無須":"白面無鬚",
-"白須":"白鬚",
-"百余":"百餘",
-"百余只":"百餘隻",
-"百余裡":"百餘里",
-"百只":"百隻",
-"百天":"百天",
-"百天后":"百天後",
-"百姓":"百姓",
-"百姓於":"百姓於",
-"百拙":"百拙",
-"百拙千丑":"百拙千醜",
-"百海":"百海",
-"百海裡":"百海里",
-"百裡":"百里",
-"的丑":"的醜",
-"的丑相":"的醜相",
-"的亮":"的亮",
-"的亮發":"的亮髮",
-"的發":"的發",
-"的發型":"的髮型",
-"的發式":"的髮式",
-"皆准":"皆準",
-"皆雲":"皆云",
-"皇後":"皇后",
-"皇歷":"皇曆",
-"皓發":"皓髮",
-"皮制":"皮製",
-"皮松":"皮鬆",
-"皺折":"皺摺",
-"盆吊":"盆弔",
-"盈余":"盈餘",
-"盍然":"溘然",
-"盍盍":"溘溘",
-"盍謝":"溘謝",
-"盍逝":"溘逝",
-"盛贊":"盛讚",
-"盜采":"盜採",
-"盡先":"儘先",
-"盡可":"盡可",
-"盡可能":"儘可能",
-"盡夠":"儘夠",
-"盡干":"盡幹",
-"盡教":"儘教",
-"盡早":"儘早",
-"盡是":"儘是",
-"盡有":"盡有",
-"盡有可能":"儘有可能",
-"盡管":"儘管",
-"盡量":"儘量",
-"監制":"監製",
-"盤卷":"盤捲",
-"盤回":"盤迴",
-"盤涅":"槃涅",
-"盤游":"盤遊",
-"目定":"目定",
-"目定口呆":"目瞪口呆",
-"目我":"目我",
-"目我於思":"目我于思",
-"盯准":"盯準",
-"盲干":"盲幹",
-"盲沖":"盲衝",
-"直卷":"直捲",
-"直干":"直幹",
-"直沖":"直衝",
-"直發":"直髮",
-"直發抖":"直發抖",
-"直發毛":"直發毛",
-"直發汗":"直發汗",
-"相並":"相併",
-"相克":"相剋",
-"相准":"相準",
-"相奸":"相姦",
-"相干":"相干",
-"相形":"相形",
-"相形見拙":"相形見絀",
-"相拼":"相拚",
-"相斗":"相鬥",
-"相當":"相當",
-"相當准":"相當準",
-"相象":"相像",
-"相關":"相關",
-"眈誤":"耽誤",
-"看不":"看不",
-"看不准":"看不準",
-"看准":"看準",
-"看出":"看出",
-"看出電影":"看齣電影",
-"看著":"看著",
-"真丑":"真醜",
-"真准":"真準",
-"真凶":"真兇",
-"真知":"真知",
-"真知卓見":"真知灼見",
-"真締":"真諦",
-"真象":"真像",
-"真象科":"真象科",
-"眩耀":"炫耀",
-"眼困":"眼睏",
-"眼框":"眼眶",
-"眼花":"眼花",
-"眼花了亂":"眼花瞭亂",
-"睡眠":"睡眠",
-"睡眠欲":"睡眠慾",
-"瞄不":"瞄不",
-"瞄不准":"瞄不準",
-"瞄准":"瞄準",
-"瞅不":"瞅不",
-"瞅不准":"瞅不準",
-"瞅准":"瞅準",
-"瞌然":"瞌然",
-"瞌然長逝":"溘然長逝",
-"瞎沖":"瞎衝",
-"瞥准":"瞥準",
-"瞧不":"瞧不",
-"瞧不准":"瞧不準",
-"瞧准":"瞧準",
-"瞧著":"瞧著",
-"知丑":"知醜",
-"短小":"短小",
-"短小精干":"短小精悍",
-"短發":"短髮",
-"短發性":"短發性",
-"短須":"短鬚",
-"石制":"石製",
-"石梁":"石樑",
-"石膏":"石膏",
-"石膏象":"石膏像",
-"石英":"石英",
-"石英表":"石英錶",
-"石英鐘":"石英鍾",
-"石象":"石像",
-"砌並":"砌併",
-"砍准":"砍準",
-"研制":"研製",
-"砰當":"砰噹",
-"破壞":"破壞",
-"破壞欲":"破壞慾",
-"硬干":"硬幹",
-"硬拼":"硬拚",
-"硬沖":"硬衝",
-"硬發":"硬髮",
-"碎發":"碎髮",
-"碑志":"碑誌",
-"碗干":"碗乾",
-"碗白":"碗白",
-"碗白干":"碗白乾",
-"碗素":"碗素",
-"碗素面":"碗素麵",
-"碗面":"碗麵",
-"碧咸":"碧咸",
-"確系":"確係",
-"確系著":"確繫著",
-"磁制":"磁製",
-"磋跎":"蹉跎",
-"磨制":"磨製",
-"示范":"示範",
-"神佑":"神祐",
-"神准":"神準",
-"神只":"神祇",
-"神游":"神遊",
-"神祗":"神祇",
-"神經":"神經",
-"神經干":"神經幹",
-"神象":"神像",
-"神跡":"神蹟",
-"祭吊":"祭弔",
-"禁欲":"禁慾",
-"禁煙":"禁菸",
-"禁若":"禁若",
-"禁若寒蟬":"噤若寒蟬",
-"禮贊":"禮讚",
-"禿發":"禿髮",
-"秀發":"秀髮",
-"私下":"私下",
-"私下裡":"私下裡",
-"私斗":"私鬥",
-"私欲":"私慾",
-"秋不":"秋不",
-"秋不干":"秋不乾",
-"秋千":"鞦韆",
-"秋台":"秋颱",
-"秋游":"秋遊",
-"秒表":"秒錶",
-"秘制":"秘製",
-"租貸":"租賃",
-"秤鎚":"秤錘",
-"稀松":"稀鬆",
-"稀裡":"稀里",
-"稍干":"稍干",
-"稍干的":"稍乾的",
-"稍有":"稍有",
-"稍有不准":"稍有不準",
-"稍松":"稍鬆",
-"稟復":"稟覆",
-"種師":"種師",
-"種師中":"种師中",
-"種師道":"种師道",
-"種放":"种放",
-"種發":"種發",
-"種發式":"種髮式",
-"種谷":"種穀",
-"稱贊":"稱讚",
-"稻桿":"稻稈",
-"稻谷":"稻穀",
-"稽康":"嵇康",
-"稽征":"稽徵",
-"積淀":"積澱",
-"積谷":"積穀",
-"積郁":"積鬱",
-"穩且":"穩且",
-"穩且准":"穩且準",
-"穩扎":"穩紮",
-"穩穩":"穩穩",
-"穩穩當當":"穩穩當當",
-"空余":"空餘",
-"空蒙":"空濛",
-"穿著":"穿著",
-"窒欲":"窒慾",
-"窗台":"窗檯",
-"窣悉":"窣窸",
-"窩裡":"窩裡",
-"窩裡斗":"窩裡鬥",
-"窺准":"窺準",
-"竟斗":"竟斗",
-"竟斗不":"竟鬥不",
-"竟象":"竟像",
-"童仆":"僮僕",
-"童婢":"僮婢",
-"競斗":"競鬥",
-"竹制":"竹製",
-"竹席":"竹蓆",
-"竹框":"竹筐",
-"竹桿":"竹竿",
-"竹簽":"竹籤",
-"竹荀":"竹筍",
-"竹藤":"竹籐",
-"竽頭":"芋頭",
-"笆蕉":"芭蕉",
-"笑容":"笑容",
-"笑容可鞠":"笑容可掬",
-"笑訥":"笑納",
-"笑魘":"笑靨",
-"符錄":"符籙",
-"筆禿":"筆禿",
-"筆禿墨干":"筆禿墨乾",
-"等侯":"等候",
-"筋斗":"觔斗",
-"筍干":"筍乾",
-"筍況":"荀況",
-"答復":"答覆",
-"策勳":"策勛",
-"算不":"算不",
-"算不准":"算不準",
-"算丑":"算醜",
-"算准":"算準",
-"管理":"管理",
-"箱函":"箱涵",
-"節余":"節餘",
-"節欲":"節慾",
-"篷松":"篷鬆",
-"篾簞":"篾簟",
-"簡朴":"簡樸",
-"簫條":"蕭條",
-"簽卜":"籤卜",
-"簽文":"籤文",
-"簽條":"籤條",
-"簽爪":"籤爪",
-"簽筒":"籤筒",
-"簽詩":"籤詩",
-"籍口":"藉口",
-"籮卜":"蘿蔔",
-"籲了":"吁了",
-"籲俞":"吁俞",
-"籲咈":"吁咈",
-"籲嗟":"吁嗟",
-"籲嘆":"吁嘆",
-"籲噓":"吁噓",
-"籲氣":"吁氣",
-"籲籲":"吁吁",
-"米余":"米餘",
-"米谷":"米穀",
-"粉制":"粉製",
-"粉干":"粉乾",
-"粗制":"粗製",
-"粗朴":"粗樸",
-"精准":"精準",
-"精制":"精製",
-"精干":"精幹",
-"精致":"精緻",
-"精辟":"精闢",
-"糊口":"餬口",
-"糊涂":"糊塗",
-"糟塌":"糟蹋",
-"糟達":"蹧躂",
-"糟遢":"糟蹋",
-"糸於":"糸於",
-"糸於一線":"繫於一線",
-"系一":"系一",
-"系一片":"係一片",
-"系一番":"係一番",
-"系一種":"係一種",
-"系一線":"繫一線",
-"系上":"繫上",
-"系乎":"繫乎",
-"系了":"繫了",
-"系住":"繫住",
-"系個":"繫個",
-"系囚":"繫囚",
-"系塊":"繫塊",
-"系妥":"繫妥",
-"系帶":"繫帶",
-"系引":"係引",
-"系心":"繫心",
-"系念":"繫念",
-"系懷":"繫懷",
-"系手":"繫手",
-"系指":"係指",
-"系掛":"繫掛",
-"系數":"係數",
-"系於":"繫於",
-"系於一發":"繫於一髮",
-"系條":"繫條",
-"系為":"係為",
-"系爭":"係爭",
-"系牢":"繫牢",
-"系獲":"係獲",
-"系留":"繫留",
-"系發":"系發",
-"系發帶":"繫髮帶",
-"系累":"係累",
-"系絆":"繫絆",
-"系絲":"系絲",
-"系絲帶":"繫絲帶",
-"系緊":"繫緊",
-"系縛":"繫縛",
-"系繩":"繫繩",
-"系纜":"繫纜",
-"系腰":"繫腰",
-"系腳":"繫腳",
-"系膜":"繫膜",
-"系著":"繫著",
-"系裹":"繫裹",
-"系褲":"系褲",
-"系褲子":"繫褲子",
-"系起":"繫起",
-"系踵":"係踵",
-"系辭":"繫辭",
-"系鈴":"繫鈴",
-"系鈴人":"繫鈴人",
-"系鞋":"系鞋",
-"系鞋帶":"繫鞋帶",
-"系頭":"系頭",
-"系頭巾":"繫頭巾",
-"系頸":"繫頸",
-"系頸闕庭":"係頸闕庭",
-"系風":"系風",
-"系風捕景":"係風捕景",
-"系馬":"繫馬",
-"紀念":"紀念",
-"紀念周":"紀念週",
-"紀歷":"紀曆",
-"約斗":"約鬥",
-"約雇":"約僱",
-"紅簽":"紅籤",
-"紅須":"紅須",
-"紅須綠眼":"紅鬚綠眼",
-"紆余":"紆餘",
-"紆回":"紆迴",
-"紆郁":"紆鬱",
-"紈絝":"紈袴",
-"紈絝子弟":"紈褲子弟",
-"紈袴":"紈褲",
-"納征":"納徵",
-"純朴":"純樸",
-"純系":"純係",
-"紙扎":"紙紮",
-"素朴":"素樸",
-"素食":"素食",
-"素食面":"素食麵",
-"紡鎚":"紡錘",
-"索馬":"索馬",
-"索馬利裡":"索馬利里",
-"索馬裡":"索馬里",
-"索馬裡蘭":"索馬里蘭",
-"紫姜":"紫薑",
-"累累":"纍纍",
-"細致":"細緻",
-"結伙":"結夥",
-"結余":"結餘",
-"結彩":"結綵",
-"結扎":"結紮",
-"結發":"結髮",
-"絞干":"絞乾",
-"給他":"給他",
-"給他干糧":"給他乾糧",
-"給你":"給你",
-"給你干糧":"給你乾糧",
-"給她":"給她",
-"給她干糧":"給她乾糧",
-"給我":"給我",
-"給我干淨":"給我乾淨",
-"絲制":"絲製",
-"絲弦":"絲絃",
-"綁扎":"綁紮",
-"經斗":"經斗",
-"經斗數":"經鬥數",
-"經理":"經理",
-"經白":"經白",
-"經白干":"經白幹",
-"綜合":"綜合",
-"綜合征":"綜合徵",
-"綢密":"稠密",
-"維系":"維繫",
-"綰發":"綰髮",
-"網志":"網誌",
-"網游":"網遊",
-"緊卷":"緊捲",
-"緊扎":"緊紮",
-"緒余":"緒餘",
-"緝凶":"緝兇",
-"編發":"編髮",
-"編篡":"編纂",
-"編采":"編採",
-"緩征":"緩徵",
-"緩沖":"緩衝",
-"緲茫":"渺茫",
-"縈回":"縈迴",
-"縝致":"縝緻",
-"縣志":"縣誌",
-"縫制":"縫製",
-"縫扎":"縫紮",
-"縫紝":"縫紉",
-"縮栗":"縮慄",
-"縱橫":"縱橫",
-"縱橫交布":"縱橫交佈",
-"縱欲":"縱慾",
-"縻系":"縻繫",
-"總理":"總理",
-"繁復":"繁複",
-"繞梁":"繞樑",
-"繡象":"繡像",
-"繪制":"繪製",
-"繼續":"繼續",
-"繼續干":"繼續幹",
-"纂位":"篡位",
-"續弦":"續絃",
-"纏卷":"纏捲",
-"纏斗":"纏鬥",
-"纖夫":"縴夫",
-"缺舌":"鴃舌",
-"罔然":"惘然",
-"罔罔":"惘惘",
-"罩准":"罩準",
-"罩卷":"罩捲",
-"罷斗":"罷鬥",
-"羅哩":"囉哩",
-"羅唆":"囉唆",
-"羅嗦":"囉嗦",
-"羅囉":"囉囉",
-"羅囉唆唆":"囉囉唆唆",
-"羅囉嗦嗦":"囉囉嗦嗦",
-"羅羅":"羅羅",
-"羅羅蘇蘇":"囉囉囌囌",
-"羅蘇":"囉囌",
-"羅裡":"羅里",
-"羅裡囉唆":"囉哩囉唆",
-"羅裡囉嗦":"囉哩囉嗦",
-"羅裡羅蘇":"囉哩囉囌",
-"羅難":"罹難",
-"羅音":"囉音",
-"羈拌":"羈絆",
-"美丑":"美醜",
-"美制":"美製",
-"美制定":"美制定",
-"美制訂":"美制訂",
-"美或":"美或",
-"美或丑":"美或醜",
-"美與":"美與",
-"美與丑":"美與醜",
-"羞赦":"羞赧",
-"羡余":"羨餘",
-"群斗":"群鬥",
-"群象":"群像",
-"羹面":"羹麵",
-"翕辟":"翕闢",
-"翦彩":"翦綵",
-"翱游":"翱遊",
-"翻來":"翻來",
-"翻來復去":"翻來覆去",
-"翻卷":"翻捲",
-"翻台":"翻檯",
-"翻松":"翻鬆",
-"翻雲":"翻雲",
-"翻雲復雨":"翻雲覆雨",
-"老丑":"老醜",
-"老且":"老且",
-"老且丑":"老且醜",
-"老姜":"老薑",
-"老板":"老闆",
-"老麼":"老么",
-"考征":"考徵",
-"者系":"者係",
-"者雲":"者云",
-"而云":"而云",
-"而云然雲":"而云然",
-"而云爾雲":"而云爾",
-"而云詩雲":"而云詩",
-"而象":"而像",
-"耐心":"耐心",
-"耐心細緻":"耐心細緻",
-"耕獲":"耕穫",
-"耗干":"耗乾",
-"聊天":"聊天",
-"聊齋":"聊齋",
-"聊齋志異":"聊齋誌異",
-"聖嘆":"聖歎",
-"聖杯":"聖盃",
-"聖象":"聖像",
-"聖跡":"聖蹟",
-"聘雇":"聘僱",
-"聯准":"聯准",
-"聯准局":"聯準局",
-"聯征":"聯徵",
-"聯手":"聯手",
-"聯手斗":"聯手鬥",
-"聯系":"聯繫",
-"聯系統":"聯系統",
-"聽不":"聽不",
-"聽不准":"聽不準",
-"聽著":"聽著",
-"肆業":"肄業",
-"肉干":"肉乾",
-"肉松":"肉鬆",
-"肉欲":"肉慾",
-"肉絲":"肉絲",
-"肉絲面":"肉絲麵",
-"肉面":"肉麵",
-"肖象":"肖像",
-"肝髒":"肝臟",
-"股栗":"股慄",
-"肢体":"肢體",
-"肯干":"肯幹",
-"肯干啊":"肯幹啊",
-"肴撰":"餚饌",
-"肺髒":"肺臟",
-"胃髒":"胃臟",
-"背影":"背影",
-"背影後":"背影後",
-"背系":"背繫",
-"胡子":"鬍子",
-"胡干":"胡幹",
-"胡梢":"鬍梢",
-"胡涂":"胡塗",
-"胡渣":"鬍渣",
-"胡碴":"胡碴",
-"胡碴子":"鬍碴子",
-"胡裡":"胡裡",
-"胡裡胡涂":"糊裡糊塗",
-"胡雪":"胡雪",
-"胡雪岩":"胡雪巖",
-"胡須":"鬍鬚",
-"胡髭":"鬍髭",
-"胡髯":"鬍髯",
-"胡鬢":"鬍鬢",
-"胰髒":"胰臟",
-"能干":"能幹",
-"能干休":"能干休",
-"能干嚎":"能乾嚎",
-"能干巴巴":"能乾巴巴",
-"能干擱":"能乾擱",
-"能干擾":"能干擾",
-"能干政":"能干政",
-"能干杯":"能乾杯",
-"能干涉":"能干涉",
-"能干淨":"能乾淨",
-"能干瞪":"能乾瞪",
-"能干笑":"能乾笑",
-"能干系":"能干系",
-"能干耗":"能乾耗",
-"能干脆":"能乾脆",
-"能干著急":"能乾著急",
-"能干預":"能干預",
-"能舍":"能捨",
-"能象":"能像",
-"能象征":"能象徵",
-"能象征性":"能象徵性",
-"脈沖":"脈衝",
-"脊梁":"脊樑",
-"脣干":"脣乾",
-"脫發":"脫髮",
-"脺髒":"脺臟",
-"脾髒":"脾臟",
-"腎髒":"腎臟",
-"腐余":"腐餘",
-"腑髒":"腑臟",
-"腕表":"腕錶",
-"腦干":"腦幹",
-"腮胡":"腮鬍",
-"腰一":"腰一",
-"腰一卷":"腰一捲",
-"腰干":"腰幹",
-"腰系":"腰繫",
-"腰間":"腰間",
-"腰間系":"腰間繫",
-"腳夫":"腳伕",
-"腳注":"腳註",
-"腳煉":"腳鍊",
-"腸髒":"腸臟",
-"膠制":"膠製",
-"膨松":"膨鬆",
-"膺品":"贗品",
-"膽戰":"膽顫",
-"臂一":"臂一",
-"臂一卷":"臂一捲",
-"臥游":"臥遊",
-"自助":"自助",
-"自助游":"自助遊",
-"自愈":"自癒",
-"自贊":"自讚",
-"自采":"自採",
-"致密":"緻密",
-"舂谷":"舂穀",
-"舄湖":"潟湖",
-"與他":"與他",
-"與他一斗":"與他一鬥",
-"與女":"與女",
-"與女斗":"與女鬥",
-"與她":"與她",
-"與她一斗":"與她一鬥",
-"舊歷":"舊曆",
-"舊歷史":"舊歷史",
-"舊游":"舊遊",
-"舌一":"舌一",
-"舌一卷":"舌一捲",
-"舍下":"舍下",
-"舍下他":"捨下他",
-"舍下你":"捨下你",
-"舍下她":"捨下她",
-"舍下我":"捨下我",
-"舍不":"舍不",
-"舍不得":"捨不得",
-"舍出":"捨出",
-"舍去":"捨去",
-"舍命":"捨命",
-"舍己":"捨己",
-"舍得":"捨得",
-"舍我":"舍我",
-"舍我其誰":"捨我其誰",
-"舍本":"捨本",
-"舍棄":"捨棄",
-"舍死":"捨死",
-"舍生":"捨生",
-"舍短":"捨短",
-"舍舊":"舍舊",
-"舍舊迎新":"捨舊迎新",
-"舍身":"捨身",
-"舍車":"舍車",
-"舍車保帥":"捨車保帥",
-"舍近":"捨近",
-"舒卷":"舒捲",
-"舒寧":"舒寧",
-"舒寧咸":"舒寧咸",
-"舔干":"舔乾",
-"舞卷":"舞捲",
-"舞娘":"舞孃",
-"舞後":"舞后",
-"舟揖":"舟楫",
-"航海":"航海",
-"航海歷":"航海曆",
-"般准":"般準",
-"舶船":"泊船",
-"船只":"船隻",
-"船夫":"船伕",
-"船梁":"船樑",
-"船漿":"船槳",
-"船纖":"船縴",
-"船蓬":"船篷",
-"艦只":"艦隻",
-"艨沖":"艨衝",
-"良游":"良遊",
-"色欲":"色慾",
-"花卷":"花捲",
-"花發":"花發",
-"花發老":"花髮老",
-"花雕":"花彫",
-"苛征":"苛徵",
-"若干":"若干",
-"若象":"若像",
-"苦干":"苦幹",
-"苦斗":"苦鬥",
-"苦瓜":"苦瓜",
-"苦瓜干":"苦瓜乾",
-"英九":"英九",
-"英寸":"英吋",
-"英尺":"英呎",
-"英裡":"英里",
-"范不":"范不",
-"范不著":"範不著",
-"范例":"範例",
-"范圍":"範圍",
-"范式":"範式",
-"范文":"範文",
-"范本":"範本",
-"范疇":"範疇",
-"范行":"範行",
-"范險":"範險",
-"茅蘆":"茅廬",
-"茶余":"茶餘",
-"茶已":"茶已",
-"茶已干":"茶已乾",
-"茶干":"茶干",
-"茶干掉":"荼乾掉",
-"茶干盡":"茶乾盡",
-"茶幾":"茶几",
-"茶毒":"荼毒",
-"茶鹵":"茶滷",
-"草席":"草蓆",
-"草管":"草管",
-"草管人命":"草菅人命",
-"荊朴":"荊璞",
-"荷裡":"荷裡",
-"荷裡活":"荷里活",
-"莒光":"莒光",
-"莒光周":"莒光週",
-"莖干":"莖幹",
-"莫明":"莫明",
-"莫明其妙":"莫名其妙",
-"莫裡":"莫里",
-"菜干":"菜乾",
-"菜肴":"菜餚",
-"華裡":"華裡",
-"華裡沙":"華里沙",
-"萬余":"萬餘",
-"萬余只":"萬餘隻",
-"萬余裡":"萬餘里",
-"萬只":"萬隻",
-"萬只怕":"萬只怕",
-"萬歷":"萬曆",
-"萬海":"萬海",
-"萬海裡":"萬海里",
-"萬裡":"萬里",
-"落發":"落髮",
-"葉裡":"葉裡",
-"葉裡溫":"葉里溫",
-"著干":"著幹",
-"葡萄":"葡萄",
-"葡萄干":"葡萄乾",
-"蒙住":"矇住",
-"蒙叟":"矇叟",
-"蒙懂":"懞懂",
-"蒙敝":"矇蔽",
-"蒙昧":"矇昧",
-"蒙朧":"矇矓",
-"蒙沖":"蒙衝",
-"蒙混":"矇混",
-"蒙瞀":"矇瞀",
-"蒙瞢":"矇瞢",
-"蒙瞽":"矇瞽",
-"蒙矓":"矇矓",
-"蒙蒙":"濛濛",
-"蒙蒙朧":"朦朦朧",
-"蒙蒙朧朧":"矇矇矓矓",
-"蒙蒙矓":"矇矇矓",
-"蒙蔽":"矇蔽",
-"蒙霧":"濛霧",
-"蒙騙":"矇騙",
-"蒲席":"蒲蓆",
-"蒸干":"蒸乾",
-"蒼術":"蒼朮",
-"蒼郁":"蒼鬱",
-"蓄發":"蓄髮",
-"蓄須":"蓄鬚",
-"蓊郁":"蓊鬱",
-"蓬松":"蓬鬆",
-"蓬發":"蓬髮",
-"蓬蓬":"蓬蓬",
-"蓬蓬松松":"蓬蓬鬆鬆",
-"蔣干":"蔣幹",
-"蔥郁":"蔥鬱",
-"蔭郁":"蔭鬱",
-"蕩口":"盪口",
-"蕩氣":"蕩氣",
-"蕩氣回腸":"蕩氣迴腸",
-"蕩氣回陽":"蕩氣迴陽",
-"蕩秋":"蕩秋",
-"蕩秋千":"盪鞦韆",
-"蕩舟":"盪舟",
-"蕩開":"盪開",
-"薄幸":"薄倖",
-"薦飢":"薦饑",
-"藤制":"藤製",
-"藤條":"籐條",
-"藤椅":"籐椅",
-"藥制":"藥製",
-"藥簽":"藥籤",
-"藥而":"藥而",
-"藥而愈":"藥而癒",
-"蘇裡":"蘇里",
-"蘇醒":"甦醒",
-"蘊借":"蘊藉",
-"蘊函":"蘊涵",
-"蘊釀":"醞釀",
-"蘿卜":"蘿蔔",
-"蘿卜干":"蘿蔔乾",
-"虎咽":"虎嚥",
-"虎斗":"虎鬥",
-"虎須":"虎鬚",
-"處理":"處理",
-"處高":"處高",
-"處高務遠":"處高鶩遠",
-"虛象":"虛像",
-"號志":"號誌",
-"虯須":"虯鬚",
-"蛋卷":"蛋捲",
-"蛩音":"跫音",
-"蝨目":"虱目",
-"蟹螫":"蟹螯",
-"蠟象":"蠟像",
-"蠻准":"蠻準",
-"蠻干":"蠻幹",
-"蠻干淨":"蠻乾淨",
-"蠻干燥":"蠻乾燥",
-"蠻干爽":"蠻乾爽",
-"血已":"血已",
-"血已干":"血已乾",
-"血才":"血才",
-"血才干":"血才乾",
-"血斗":"血鬥",
-"行事":"行事",
-"行事歷":"行事曆",
-"行凶":"行兇",
-"行屍":"行尸",
-"行李":"行李",
-"行李卷":"行李捲",
-"行游":"行遊",
-"衣物":"衣物",
-"衣物已干":"衣物已乾",
-"衣物漸干":"衣物漸乾",
-"衣衫":"衣衫",
-"衣衫已干":"衣衫已乾",
-"衣驅":"衣袪",
-"表帶":"錶帶",
-"表店":"錶店",
-"表廠":"錶廠",
-"表征":"表徵",
-"表殼":"錶殼",
-"表演":"表演",
-"表演欲":"表演慾",
-"表煉":"錶鏈",
-"表范":"表範",
-"表鏈":"錶鏈",
-"衫畢":"衫畢",
-"衫畢挺":"衫筆挺",
-"袖一":"袖一",
-"袖一卷":"袖一捲",
-"被並":"被併",
-"被卷":"被捲",
-"被干":"被幹",
-"被干擾":"被干擾",
-"被干涉":"被干涉",
-"被干預":"被干預",
-"被復":"被覆",
-"被發":"被發",
-"被發左衽":"被髮左衽",
-"被褥":"被縟",
-"裁並":"裁併",
-"裁制":"裁製",
-"裊雄":"梟雄",
-"裕余":"裕餘",
-"裙擺":"裙襬",
-"補注":"補註",
-"裝璜":"裝潢",
-"裡亞":"里亞",
-"裡仁":"里仁",
-"裡余":"裡餘",
-"裡士":"裡士",
-"裡士滿":"里士滿",
-"裡夫":"里夫",
-"裡奧":"里奧",
-"裡尼":"里尼",
-"裡布":"里布",
-"裡拉":"里拉",
-"裡斯":"裡斯",
-"裡斯本":"里斯本",
-"裡昂":"里昂",
-"裡根":"里根",
-"裡民":"里民",
-"裡爾":"里爾",
-"裡科":"里科",
-"裡程":"里程",
-"裡約":"里約",
-"裡納":"里納",
-"裡美":"里美",
-"裡肌":"里肌",
-"裡蘭":"里蘭",
-"裡象":"裡像",
-"裡路":"里路",
-"裡里":"里奇",
-"裡長":"里長",
-"裡面":"裡面",
-"裡面沖":"裡面衝",
-"裡面沖水":"裡面沖水",
-"裡面沖洗":"裡面沖洗",
-"裡面沖涼":"裡面沖涼",
-"裡面沖澡":"裡面沖澡",
-"裡面沖身":"裡面沖身",
-"裹屍":"裹尸",
-"裹扎":"裹紮",
-"褐發":"褐髮",
-"褻讀":"褻瀆",
-"襁保":"襁褓",
-"襤縷":"襤褸",
-"襲卷":"襲捲",
-"西周":"西周",
-"西岳":"西嶽",
-"西後":"西后",
-"西昆":"西崑",
-"西歷":"西曆",
-"西沖":"西衝",
-"西游":"西遊",
-"西裝":"西裝",
-"西裝畢挺":"西裝筆挺",
-"西裡":"西里",
-"要丑":"要丑",
-"要丑上":"要醜上",
-"要干":"要干",
-"要干了":"要乾了",
-"要干什麼":"要幹什麼",
-"要干嘛":"要幹嘛",
-"要干杯":"要乾杯",
-"要斗":"要鬥",
-"要衝":"要衝",
-"要衝廁":"要沖廁",
-"要衝水":"要沖水",
-"要衝涼":"要沖涼",
-"要象":"要像",
-"規划":"規劃",
-"規模":"規模",
-"規范":"規範",
-"覓准":"覓準",
-"親自":"親自",
-"親自干":"親自幹",
-"覬幸":"覬倖",
-"覷准":"覷準",
-"觀斗":"觀鬥",
-"觀采":"觀採",
-"角斗":"角鬥",
-"角落":"角落",
-"解簽":"解籤",
-"解簽約":"解簽約",
-"解象":"解像",
-"解雇":"解僱",
-"觴子":"殤子",
-"觸須":"觸鬚",
-"訂制":"訂製",
-"計准":"計準",
-"計劃":"計畫",
-"評擊":"抨擊",
-"評注":"評註",
-"詞余":"詞餘",
-"詞匯":"詞彙",
-"詠嘆":"詠歎",
-"試卷":"試卷",
-"詩余":"詩餘",
-"詩雲":"詩云",
-"話雲":"話云",
-"該像":"該像",
-"該像征":"該象徵",
-"該象":"該像",
-"詳征":"詳徵",
-"詳注":"詳註",
-"誇克":"夸克",
-"誇克星":"夸克星",
-"誇父":"夸父",
-"誇爾":"夸爾",
-"誇特":"夸特",
-"誇脫":"夸脫",
-"誇贊":"誇讚",
-"認丑":"認醜",
-"認准":"認準",
-"誘奸":"誘姦",
-"誘欲":"誘慾",
-"誘發":"誘發",
-"誘發型":"誘發型",
-"語匯":"語彙",
-"語雲":"語云",
-"誠征":"誠徵",
-"誠朴":"誠樸",
-"誣蔑":"誣衊",
-"誤沖":"誤衝",
-"說不":"說不",
-"說不准":"說不準",
-"說干":"說干",
-"說干便干":"說幹便幹",
-"說干就干":"說幹就幹",
-"說得":"說得",
-"說得個准":"說得個準",
-"說的":"說的",
-"說的准":"說的準",
-"說簽":"說籤",
-"誰干":"誰幹",
-"誰干淨":"誰乾淨",
-"誰干脆":"誰乾脆",
-"誰是":"誰是",
-"誰是於":"誰是于",
-"課余":"課餘",
-"課征":"課徵",
-"調准":"調準",
-"調制":"調製",
-"調制解調":"調制解調",
-"調松":"調鬆",
-"請托":"請託",
-"諸余":"諸餘",
-"諸候":"諸侯",
-"諺雲":"諺云",
-"謀干":"謀幹",
-"謬贊":"謬讚",
-"謷丑":"謷醜",
-"譁啦":"嘩啦",
-"譁喇":"嘩喇",
-"譁的":"譁的",
-"譁的一聲":"嘩的一聲",
-"譁譁":"嘩嘩",
-"警剔":"警惕",
-"譯注":"譯註",
-"護發":"護髮",
-"讀取":"讀取",
-"讀取頭":"讀取頭",
-"讀寫":"讀寫",
-"讀寫頭":"讀寫頭",
-"變丑":"變醜",
-"變征":"變徵",
-"變松":"變鬆",
-"讖悔":"懺悔",
-"谷倉":"穀倉",
-"谷場":"穀場",
-"谷子":"穀子",
-"谷殼":"穀殼",
-"谷物":"穀物",
-"谷皮":"穀皮",
-"谷神":"穀神",
-"谷稻":"穀稻",
-"谷米":"穀米",
-"谷粉":"穀粉",
-"谷粒":"穀粒",
-"谷粟":"穀粟",
-"谷艙":"穀艙",
-"谷苗":"穀苗",
-"谷草":"穀草",
-"谷賤":"穀賤",
-"谷道":"穀道",
-"谷雨":"穀雨",
-"谷類":"穀類",
-"谷麥":"穀麥",
-"豁上":"豁上",
-"豁上干":"豁上幹",
-"豁出":"豁出",
-"豁出命干":"豁出命幹",
-"豁命":"豁命",
-"豁命干":"豁命幹",
-"豁壑":"谿壑",
-"豁死":"豁死",
-"豁死干":"豁死幹",
-"豁起":"豁起",
-"豁起來干":"豁起來幹",
-"豆制":"豆製",
-"豆腐":"豆腐",
-"豆腐干":"豆腐乾",
-"豆蔻":"荳蔻",
-"豆鼓":"豆豉",
-"豐姿":"丰姿",
-"豐度":"丰度",
-"豐情":"丰情",
-"豐神":"丰神",
-"豐諛":"豐腴",
-"豐韻":"丰韻",
-"豔後":"豔后",
-"象一":"像一",
-"象三":"像三",
-"象不":"象不",
-"象不像":"像不像",
-"象之":"象之",
-"象之前":"像之前",
-"象二":"像二",
-"象些":"像些",
-"象人":"像人",
-"象什":"像什",
-"象今":"像今",
-"象他":"像他",
-"象以":"象以",
-"象以前":"像以前",
-"象以往":"像以往",
-"象似":"像似",
-"象你":"像你",
-"象個":"像個",
-"象兩":"像兩",
-"象其":"像其",
-"象別":"像別",
-"象剛":"像剛",
-"象嗎":"像嗎",
-"象在":"像在",
-"象夢":"像夢",
-"象她":"像她",
-"象它":"像它",
-"象已":"像已",
-"象帶":"像帶",
-"象平":"象平",
-"象平時":"像平時",
-"象往":"像往",
-"象征":"象徵",
-"象征戰":"像征戰",
-"象征服":"像征服",
-"象您":"像您",
-"象我":"像我",
-"象所":"象所",
-"象所有":"像所有",
-"象昨":"像昨",
-"象是":"像是",
-"象有":"像有",
-"象極":"像極",
-"象樣":"像樣",
-"象水":"像水",
-"象永":"像永",
-"象片":"像片",
-"象牙":"象牙",
-"象牙齒":"像牙齒",
-"象現":"像現",
-"象甚":"像甚",
-"象當":"像當",
-"象瘋":"像瘋",
-"象素":"像素",
-"象自":"象自",
-"象自己":"像自己",
-"象要":"像要",
-"象話":"像話",
-"象誰":"像誰",
-"象贊":"像贊",
-"象這":"像這",
-"象過":"象過",
-"象過去":"像過去",
-"象那":"像那",
-"象銀":"像銀",
-"象雨":"像雨",
-"象風":"像風",
-"象鬼":"像鬼",
-"象麼":"像麼",
-"豪華":"豪華",
-"豪華游":"豪華遊",
-"豬只":"豬隻",
-"豬肝":"豬肝",
-"豬肝面":"豬肝麵",
-"豬腳":"豬腳",
-"豬腳面":"豬腳麵",
-"豬舌":"豬舌",
-"豬舌面":"豬舌麵",
-"貌丑":"貌醜",
-"貝胄":"貝冑",
-"貝裡":"貝里",
-"財欲":"財慾",
-"貪欲":"貪慾",
-"貴干":"貴幹",
-"買凶":"買兇",
-"買姜":"買薑",
-"買煙":"買菸",
-"貼志":"貼誌",
-"貼著":"貼著",
-"賑飢":"賑饑",
-"賜復":"賜覆",
-"賣奸":"賣姦",
-"賣姜":"賣薑",
-"質朴":"質樸",
-"賭台":"賭檯",
-"賭台灣":"賭台灣",
-"賭斗":"賭鬥",
-"賸余":"賸餘",
-"購並":"購併",
-"購並不":"購並不",
-"購並且":"購並且",
-"購並非":"購並非",
-"購買":"購買",
-"購買欲":"購買慾",
-"贊一":"贊一",
-"贊一句":"讚一句",
-"贊一聲":"讚一聲",
-"贊一贊":"讚一讚",
-"贊不":"贊不",
-"贊不絕口":"讚不絕口",
-"贊了":"讚了",
-"贊他":"讚他",
-"贊仰":"讚仰",
-"贊你":"讚你",
-"贊佩":"讚佩",
-"贊個":"贊個",
-"贊個不":"讚個不",
-"贊兩":"贊兩",
-"贊兩句":"讚兩句",
-"贊口":"贊口",
-"贊口不":"讚口不",
-"贊嘆":"讚歎",
-"贊她":"讚她",
-"贊好":"讚好",
-"贊妙":"讚妙",
-"贊得":"讚得",
-"贊您":"讚您",
-"贊我":"讚我",
-"贊捧":"讚捧",
-"贊揚":"讚揚",
-"贊樂":"讚樂",
-"贊歌":"讚歌",
-"贊的":"讚的",
-"贊美":"讚美",
-"贊羡":"讚羨",
-"贊羨":"讚羨",
-"贊聲":"讚聲",
-"贊自":"贊自",
-"贊自己":"讚自己",
-"贊著":"讚著",
-"贊許":"讚許",
-"贊詞":"讚詞",
-"贊詠":"讚詠",
-"贊誦":"讚誦",
-"贊譽":"讚譽",
-"贊賞":"讚賞",
-"贊辭":"讚辭",
-"贊這":"讚這",
-"贊過":"讚過",
-"贊道":"讚道",
-"贊那":"讚那",
-"贊頌":"讚頌",
-"贊：":"讚：",
-"贏余":"贏餘",
-"赤坎":"赤崁",
-"赤發":"赤髮",
-"起來":"起來",
-"起來象":"起來像",
-"起哄":"起鬨",
-"起干":"起幹",
-"超征":"超徵",
-"越丑":"越醜",
-"越准":"越準",
-"越咸":"越鹹",
-"越斗":"越鬥",
-"越象":"越像",
-"趕制":"趕製",
-"趨之":"趨之",
-"趨之若務":"趨之若鶩",
-"跌沖":"跌衝",
-"跖骨":"蹠骨",
-"跟他":"跟他",
-"跟他一斗":"跟他一鬥",
-"跟他斗":"跟他鬥",
-"跟你":"跟你",
-"跟你斗":"跟你鬥",
-"跟女":"跟女",
-"跟女斗":"跟女鬥",
-"跟她":"跟她",
-"跟她斗":"跟她鬥",
-"跟我":"跟我",
-"跟我斗":"跟我鬥",
-"跟斗":"觔斗",
-"跟蹌":"踉蹌",
-"路簽":"路籤",
-"路裡":"路里",
-"跳只":"跳只",
-"跳只舞":"跳隻舞",
-"跳梁":"跳樑",
-"跳表":"跳錶",
-"踅門":"踅門",
-"踅門了戶":"踅門瞭戶",
-"踐道":"棧道",
-"踢踏":"踢踏",
-"踢踏舞":"踢躂舞",
-"蹉陀":"蹉跎",
-"蹣頇":"顢頇",
-"躁郁":"躁鬱",
-"身分":"身份",
-"身系":"身繫",
-"躬敬":"恭敬",
-"軀干":"軀幹",
-"車夫":"車伕",
-"車游":"車遊",
-"軌范":"軌範",
-"軒辟":"軒闢",
-"軔性":"韌性",
-"軟發":"軟髮",
-"軫域":"畛域",
-"較丑":"較醜",
-"較像":"較像",
-"較像征":"較象徵",
-"較准":"較準",
-"較斗":"較鬥",
-"較松":"較鬆",
-"較象":"較像",
-"輕台":"輕颱",
-"輕挑":"輕佻",
-"輕松":"輕鬆",
-"輕篾":"輕蔑",
-"輕輕":"輕輕",
-"輕輕松松":"輕輕鬆鬆",
-"輩子":"輩子",
-"輩子干":"輩子幹",
-"輪回":"輪迴",
-"輪奸":"輪姦",
-"輪流":"輪流",
-"輪流干":"輪流幹",
-"輯凶":"輯兇",
-"轉寰":"轉圜",
-"轉戾":"轉戾",
-"轉戾點":"轉捩點",
-"轉托":"轉託",
-"轎夫":"轎伕",
-"辟佛":"闢佛",
-"辟作":"闢作",
-"辟劃":"闢劃",
-"辟土":"闢土",
-"辟地":"闢地",
-"辟室":"闢室",
-"辟建":"闢建",
-"辟為":"闢為",
-"辟田":"闢田",
-"辟築":"闢築",
-"辟謠":"闢謠",
-"辟謬":"闢謬",
-"辟谷":"辟榖",
-"辟蹊":"辟蹊",
-"辟蹊徑":"闢蹊徑",
-"辟透":"闢透",
-"辟邪":"闢邪",
-"辟門":"辟門",
-"辟門戶":"闢門戶",
-"辟闔":"闢闔",
-"辣面":"辣麵",
-"辨奸":"辨奸",
-"辨奸論":"辨姦論",
-"辭匯":"辭彙",
-"辮發":"辮髮",
-"農歷":"農曆",
-"農民":"農民",
-"農民歷":"農民曆",
-"農獲":"農穫",
-"迂回":"迂迴",
-"迎斗":"迎鬥",
-"迥廊":"迴廊",
-"迥避":"迴避",
-"迨盡":"殆盡",
-"迷奸":"迷姦",
-"迷朦":"迷濛",
-"迷罔":"迷惘",
-"迷蒙":"迷濛",
-"迷迷":"迷迷",
-"迷迷蒙蒙":"迷迷濛濛",
-"迻邐":"迆邐",
-"追凶":"追兇",
-"追朔":"追溯",
-"追根":"追根",
-"追根究抵":"追根究柢",
-"追謚":"追諡",
-"透辟":"透闢",
-"這伙":"這伙",
-"這伙人":"這夥人",
-"這出":"這出",
-"這出劇":"這齣劇",
-"這出好戲":"這齣好戲",
-"這出戲":"這齣戲",
-"這出電影":"這齣電影",
-"這般":"這般",
-"這般准":"這般準",
-"這般干法":"這般幹法",
-"這裡":"這裡",
-"這麼":"這麼",
-"這麼丑":"這麼醜",
-"這麼像征":"這麼象徵",
-"這麼准":"這麼準",
-"這麼干":"這麼幹",
-"通奸":"通姦",
-"通心":"通心",
-"通心面":"通心麵",
-"通浚":"通濬",
-"通諜":"通牒",
-"通輯":"通緝",
-"逞凶":"逞兇",
-"速干":"速乾",
-"速食":"速食",
-"速食面":"速食麵",
-"造象":"造像",
-"連傑":"連杰",
-"連卷":"連捲",
-"連抉":"連袂",
-"連斗":"連鬥",
-"連系":"連繫",
-"連系統":"連系統",
-"連絡":"聯絡",
-"逼奸":"逼姦",
-"逾年":"踰年",
-"逾矩":"踰矩",
-"逾越":"踰越",
-"逾閒":"踰閒",
-"遇著":"遇著",
-"運籌":"運籌",
-"運籌惟幄":"運籌帷幄",
-"遍布":"遍佈",
-"過松":"過鬆",
-"遐疵":"瑕疵",
-"道坎":"道檻",
-"道裡":"道里",
-"道贊":"道讚",
-"遞回":"遞迴",
-"遠揚":"遠颺",
-"遠游":"遠遊",
-"遨游":"遨遊",
-"遮丑":"遮醜",
-"選准":"選準",
-"遺撼":"遺憾",
-"遺范":"遺範",
-"遺象":"遺像",
-"遺跡":"遺蹟",
-"遼亮":"嘹喨",
-"遼沈":"遼瀋",
-"邀天":"邀天",
-"邀天之幸":"邀天之倖",
-"邀斗":"邀鬥",
-"邂怠":"懈怠",
-"還丑":"還醜",
-"還要":"還要",
-"還要丑":"還要醜",
-"還象":"還像",
-"還象征服":"還像征服",
-"邊斗":"邊斗",
-"邊斗邊":"邊鬥邊",
-"邋塌":"邋遢",
-"邋蹋":"邋遢",
-"那伙":"那伙",
-"那伙人":"那夥人",
-"那出":"那出",
-"那出劇":"那齣劇",
-"那出好戲":"那齣好戲",
-"那出戲":"那齣戲",
-"那出電影":"那齣電影",
-"那只":"那隻",
-"那只不過":"那只不過",
-"那只佔":"那只佔",
-"那只剩":"那只剩",
-"那只加":"那只加",
-"那只在":"那只在",
-"那只增":"那只增",
-"那只多了":"那只多了",
-"那只夠":"那只夠",
-"那只好":"那只好",
-"那只差":"那只差",
-"那只是":"那只是",
-"那只會":"那只會",
-"那只有":"那只有",
-"那只減":"那只減",
-"那只缺":"那只缺",
-"那只能":"那只能",
-"那只要":"那只要",
-"那只需":"那只需",
-"那般":"那般",
-"那般准":"那般準",
-"那裡":"那裡",
-"那麼":"那麼",
-"那麼丑":"那麼醜",
-"那麼准":"那麼準",
-"那麼干":"那麼幹",
-"邪欲":"邪慾",
-"郁卒":"鬱卒",
-"郁堙":"鬱堙",
-"郁塞":"鬱塞",
-"郁壘":"鬱壘",
-"郁巍":"鬱巍",
-"郁律":"鬱律",
-"郁怒":"鬱怒",
-"郁悒":"鬱悒",
-"郁悶":"鬱悶",
-"郁憤":"鬱憤",
-"郁抑":"鬱抑",
-"郁挹":"鬱挹",
-"郁於":"鬱於",
-"郁暗":"鬱暗",
-"郁林":"鬱林",
-"郁沈":"鬱沈",
-"郁沉":"鬱沉",
-"郁泱":"鬱泱",
-"郁火":"鬱火",
-"郁熱":"鬱熱",
-"郁燠":"鬱燠",
-"郁症":"鬱症",
-"郁痛":"鬱痛",
-"郁積":"鬱積",
-"郁紆":"鬱紆",
-"郁結":"鬱結",
-"郁蓊":"鬱蓊",
-"郁血":"鬱血",
-"郁邑":"鬱邑",
-"郁郁":"鬱鬱",
-"郁金":"郁金",
-"郁金香":"鬱金香",
-"郁閉":"鬱閉",
-"郁黑":"鬱黑",
-"郊游":"郊遊",
-"郎鐺":"啷鐺",
-"部落":"部落",
-"郭朴":"郭璞",
-"都肯":"都肯",
-"都肯干":"都肯幹",
-"都舍":"都舍",
-"都舍下":"都捨下",
-"都象":"都像",
-"鄉裡":"鄉里",
-"鄉願":"鄉愿",
-"鄉願意":"鄉願意",
-"鄭凱":"鄭凱",
-"鄭凱云":"鄭凱云",
-"鄰裡":"鄰里",
-"配制":"配製",
-"配膳":"配膳",
-"配膳台":"配膳檯",
-"酒一":"酒一",
-"酒一壇":"酒一罈",
-"酒壇":"酒罈",
-"酒已":"酒已",
-"酒已干":"酒已乾",
-"酒干":"酒干",
-"酒干了":"酒乾了",
-"酒干掉":"酒乾掉",
-"酒干盡":"酒乾盡",
-"酒曲":"酒麴",
-"酒肴":"酒餚",
-"酣斗":"酣鬥",
-"酥松":"酥鬆",
-"酬傭":"酬佣",
-"酸疼":"痠疼",
-"酸痛":"痠痛",
-"酸軟":"痠軟",
-"醃制":"醃製",
-"醃臢":"腌臢",
-"醇朴":"醇樸",
-"醋壇":"醋罈",
-"醬面":"醬麵",
-"醮墨":"蘸墨",
-"釀制":"釀製",
-"采伐":"採伐",
-"采供":"採供",
-"采信":"採信",
-"采光":"採光",
-"采卵":"採卵",
-"采取":"採取",
-"采寫":"採寫",
-"采拾":"採拾",
-"采挖":"採挖",
-"采掘":"採掘",
-"采摘":"採摘",
-"采摭":"採摭",
-"采擇":"採擇",
-"采擷":"採擷",
-"采收":"採收",
-"采桑":"採桑",
-"采梅":"採梅",
-"采樣":"採樣",
-"采樵":"採樵",
-"采樹":"採樹",
-"采檢":"採檢",
-"采水":"採水",
-"采油":"採油",
-"采煤":"採煤",
-"采獲":"採獲",
-"采獵":"採獵",
-"采珠":"採珠",
-"采生":"採生",
-"采用":"採用",
-"采石":"採石",
-"采硫":"採硫",
-"采礦":"採礦",
-"采種":"採種",
-"采稿":"採稿",
-"采粉":"採粉",
-"采精":"採精",
-"采納":"採納",
-"采編":"採編",
-"采花":"採花",
-"采芹":"採芹",
-"采茶":"採茶",
-"采草":"採草",
-"采蓮":"採蓮",
-"采薇":"採薇",
-"采藥":"採藥",
-"采蜜":"採蜜",
-"采血":"採血",
-"采行":"採行",
-"采補":"採補",
-"采計":"採計",
-"采訪":"採訪",
-"采認":"採認",
-"采證":"採證",
-"采買":"採買",
-"采購":"採購",
-"采辦":"採辦",
-"采選":"採選",
-"采金":"採金",
-"采錄":"採錄",
-"采集":"採集",
-"采風":"采風",
-"采風力":"採風力",
-"采食":"採食",
-"采餌":"採餌",
-"采驗":"採驗",
-"重復":"重複",
-"重托":"重託",
-"重涂":"重塗",
-"重游":"重遊",
-"重簷":"重檐",
-"野游":"野遊",
-"金侖":"金崙",
-"金杯":"金盃",
-"金橘":"金桔",
-"金象":"金象",
-"金象獎":"金像獎",
-"釘槌":"釘鎚",
-"針炙":"針灸",
-"針黹":"鍼黹",
-"釣游":"釣遊",
-"釵環":"釵鐶",
-"鉗制":"箝制",
-"鉗口":"箝口",
-"鉛鎚":"鉛錘",
-"鉸煉":"鉸鍊",
-"銀制":"銀製",
-"銀須":"銀鬚",
-"銅制":"銅製",
-"銅象":"銅像",
-"銜土":"啣土",
-"銜泥":"啣泥",
-"銜環":"啣環",
-"銜草":"啣草",
-"銷溶":"銷鎔",
-"鋁制":"鋁製",
-"鋪位":"舖位",
-"鋼制":"鋼製",
-"鋼梁":"鋼樑",
-"錄制":"錄製",
-"錄影":"錄影",
-"錄影後":"錄影後",
-"錄象":"錄像",
-"錙衣":"緇衣",
-"錢鐘":"錢鐘",
-"錢鐘書":"錢鍾書",
-"錦鏢":"錦標",
-"錫制":"錫製",
-"錫坤":"錫堃",
-"鍛煉":"鍛鍊",
-"鍛羽":"鎩羽",
-"鍛鏈":"鍛鍊",
-"鍾山":"鍾山",
-"鎖煉":"鎖鍊",
-"鎖碎":"瑣碎",
-"鎚煉":"錘煉",
-"鎧胄":"鎧冑",
-"鏟刈":"剷刈",
-"鏟去":"剷去",
-"鏟平":"剷平",
-"鏟滅":"剷滅",
-"鏟起":"剷起",
-"鏟除":"剷除",
-"鏡象":"鏡像",
-"鐘不":"鍾不",
-"鐘情":"鍾情",
-"鐘意":"鍾意",
-"鐘愛":"鍾愛",
-"鐘欣":"鐘欣",
-"鐘欣桐":"鍾欣桐",
-"鐘無":"鐘無",
-"鐘無豔":"鍾無艷",
-"鐘表":"鐘錶",
-"鐘表態":"鍾表態",
-"鐘表明":"鍾表明",
-"鐘表決":"鍾表決",
-"鐘表演":"鐘表演",
-"鐘表現":"鐘表現",
-"鐘表示":"鍾表示",
-"鐘表露":"鍾表露",
-"鐘表面":"鐘錶面",
-"鐘馗":"鍾馗",
-"鐵制":"鐵製",
-"鐵槌":"鐵鎚",
-"鐵煉":"鐵鍊",
-"鐵練":"鐵鍊",
-"鑲制":"鑲製",
-"鑲崁":"鑲嵌",
-"長台":"長檯",
-"長得":"長得",
-"長得丑":"長得醜",
-"長發":"長發",
-"長發布":"長發布",
-"長發邨":"長發邨",
-"長籲":"長吁",
-"長象":"長像",
-"長須":"長鬚",
-"門坎":"門檻",
-"門斗":"門鬥",
-"門梁":"門樑",
-"閃鑠":"閃爍",
-"開征":"開徵",
-"開浚":"開濬",
-"開發":"開發",
-"開辟":"開闢",
-"開采":"開採",
-"開鋪":"開舖",
-"閒余":"閒餘",
-"閒游":"閒遊",
-"閒遐":"閒暇",
-"閒靜":"閑靜",
-"間不":"間不",
-"間不容發":"間不容髮",
-"閨范":"閨範",
-"閫范":"閫範",
-"闔家":"閤家",
-"闔府":"閤府",
-"闔府上下":"閤府上下",
-"關注":"關注",
-"關系":"關係",
-"關系統":"關系統",
-"防台":"防颱",
-"防台灣":"防台灣",
-"防御":"防禦",
-"防水":"防水",
-"防水表":"防水錶",
-"防礙":"妨礙",
-"防范":"防範",
-"防飢":"防饑",
-"阿腴":"阿諛",
-"阿裡":"阿里",
-"陀手":"舵手",
-"陀表":"陀錶",
-"附注":"附註",
-"限制":"限制",
-"陡松":"陡鬆",
-"陪吊":"陪弔",
-"陪著":"陪著",
-"陰干":"陰乾",
-"陰歷":"陰曆",
-"陰郁":"陰鬱",
-"陷井":"陷阱",
-"陽春":"陽春",
-"陽春面":"陽春麵",
-"陽歷":"陽曆",
-"隆准":"隆準",
-"隔周":"隔週",
-"雄赳":"雄赳",
-"雄赳赳":"雄糾糾",
-"雅致":"雅緻",
-"雅范":"雅範",
-"集團":"集團",
-"集游":"集游",
-"集游法":"集遊法",
-"雇不":"雇不",
-"雇不起":"僱不起",
-"雇主":"僱主",
-"雇人":"僱人",
-"雇傭":"僱傭",
-"雇員":"僱員",
-"雇役":"僱役",
-"雇有":"僱有",
-"雇用":"僱用",
-"雇舟":"僱舟",
-"雇船":"僱船",
-"雇請":"僱請",
-"雇車":"僱車",
-"雇農":"僱農",
-"雇馬":"僱馬",
-"雕喪":"彫喪",
-"雕梁":"雕樑",
-"雕殘":"彫殘",
-"雕落":"彫落",
-"雕蟲":"彫蟲",
-"雕謝":"彫謝",
-"雕象":"雕像",
-"雕零":"彫零",
-"雖然":"雖然",
-"雖然丑":"雖然醜",
-"雙並":"雙併",
-"雙周":"雙週",
-"雙斗":"雙鬥",
-"雜志":"雜誌",
-"雞只":"雞隻",
-"雞奸":"雞姦",
-"雞絲":"雞絲",
-"雞絲面":"雞絲麵",
-"雞腿":"雞腿",
-"雞腿面":"雞腿麵",
-"難咽":"難嚥",
-"難干":"難乾",
-"難斗":"難斗",
-"難斗的":"難鬥的",
-"難舍":"難捨",
-"雲吞":"雲吞",
-"雲吞面":"雲吞麵",
-"雲宵":"雲霄",
-"雲游":"雲遊",
-"雲爾":"云爾",
-"雲雲":"云云",
-"雲：":"云：",
-"電子":"電子",
-"電子表":"電子錶",
-"電子表單":"電子表單",
-"電子表情":"電子表情",
-"電子表格":"電子表格",
-"電子表示":"電子表示",
-"電子鐘":"電子鐘",
-"電影":"電影",
-"電影後":"電影後",
-"電桿":"電杆",
-"電線":"電線",
-"電線桿":"電線杆",
-"震松":"震鬆",
-"震栗":"震慄",
-"震蕩":"震盪",
-"露丑":"露醜",
-"靈俐":"伶俐",
-"靈幡":"靈旛",
-"靈欲":"靈慾",
-"青棵":"青稞",
-"青笞":"青苔",
-"非常":"非常",
-"非常不准":"非常不準",
-"非常丑":"非常醜",
-"非常准":"非常準",
-"非常象":"非常像",
-"非涂":"非涂",
-"非涂不可":"非塗不可",
-"非象":"非像",
-"非象征服":"非像征服",
-"面丑":"面醜",
-"面價":"麵價",
-"面價值":"面價值",
-"面團":"麵糰",
-"面店":"麵店",
-"面店鋪":"麵店舖",
-"面廠":"麵廠",
-"面攤":"麵攤",
-"面攤牌":"面攤牌",
-"面攤開":"面攤開",
-"面杖":"麵杖",
-"面湯":"麵湯",
-"面灰":"麵灰",
-"面疙":"面疙",
-"面疙瘩":"麵疙瘩",
-"面白":"面白",
-"面白無須":"面白無鬚",
-"面碗":"麵碗",
-"面筋":"麵筋",
-"面粉":"麵粉",
-"面糊":"麵糊",
-"面線":"麵線",
-"面茶":"麵茶",
-"面食":"麵食",
-"面食面":"麵食麵",
-"面餃":"麵餃",
-"面餅":"麵餅",
-"面館":"麵館",
-"面饃":"麵饃",
-"面龜":"麵龜",
-"靦典":"靦腆",
-"鞣制":"鞣製",
-"韃伐":"撻伐",
-"韋席":"韋蓆",
-"韋後":"韋后",
-"韓制":"韓製",
-"音不":"音不",
-"音不准":"音不準",
-"音准":"音準",
-"音象":"音像",
-"響當":"響噹",
-"響當當":"響噹噹",
-"頁簽":"頁籤",
-"頁面":"頁面",
-"頂凶":"頂兇",
-"頂頂":"頂頂",
-"頂頂大名":"鼎鼎大名",
-"項煉":"項鍊",
-"順發":"順發",
-"順發凝露":"順髮凝露",
-"順發水":"順髮水",
-"順發液":"順髮液",
-"順發素":"順髮素",
-"順發霧":"順髮霧",
-"順發露":"順髮露",
-"順順":"順順",
-"順順當當":"順順當當",
-"須叟":"須臾",
-"須子":"鬚子",
-"須根":"鬚根",
-"須毛":"鬚毛",
-"須求":"需求",
-"須發":"鬚髮",
-"須發展":"須發展",
-"須發文":"須發文",
-"須發表":"須發表",
-"須眉":"鬚眉",
-"須采":"須採",
-"須須":"鬚鬚",
-"須髯":"鬚髯",
-"須鬢":"鬚鬢",
-"須鯨":"鬚鯨",
-"頌贊":"頌讚",
-"預制":"預製",
-"預布":"預佈",
-"頗准":"頗準",
-"頗象":"頗像",
-"領袖":"領袖",
-"領袖欲":"領袖慾",
-"頜須":"頜鬚",
-"頭亮":"頭亮",
-"頭亮發":"頭亮髮",
-"頭扎":"頭紮",
-"頭散":"頭散",
-"頭散發":"頭散髮",
-"頭發":"頭髮",
-"頭發了狂":"頭發了狂",
-"頭發了瘋":"頭發了瘋",
-"頭發光":"頭發光",
-"頭發出":"頭發出",
-"頭發力":"頭發力",
-"頭發功":"頭發功",
-"頭發呆":"頭發呆",
-"頭發售":"頭發售",
-"頭發炎":"頭發炎",
-"頭發狂":"頭發狂",
-"頭發現":"頭發現",
-"頭發生":"頭發生",
-"頭發瘋":"頭發瘋",
-"頭發育":"頭發育",
-"頭發麻":"頭發麻",
-"頭發黑":"頭發黑",
-"頭發黴":"頭發霉",
-"頭花":"頭花",
-"頭花發":"頭花髮",
-"頰須":"頰鬚",
-"頸系":"頸繫",
-"頹圯":"頹圮",
-"題簽":"題籤",
-"願干":"願幹",
-"願干一杯":"願乾一杯",
-"願干坤":"願乾坤",
-"願干旱":"願乾旱",
-"願干杯":"願乾杯",
-"願干涸":"願乾涸",
-"願干淨":"願乾淨",
-"願干燥":"願乾燥",
-"願干糧":"願乾糧",
-"願干脆":"願乾脆",
-"願干這杯":"願乾這杯",
-"願干那杯":"願乾那杯",
-"願意":"願意",
-"願意干":"願意幹",
-"願意干坤":"願意乾坤",
-"顛僕":"顛仆",
-"顛復":"顛覆",
-"顛顛":"顛顛",
-"顛顛僕僕":"顛顛仆仆",
-"顧恤":"顧卹",
-"顫斗":"顫抖",
-"顫栗":"顫慄",
-"顯象":"顯像",
-"風刮":"風颳",
-"風卷":"風捲",
-"風卷殘雲":"風捲殘雲",
-"風土":"風土",
-"風土志":"風土誌",
-"風姿":"風姿",
-"風姿卓約":"風姿綽約",
-"風干":"風乾",
-"風物":"風物",
-"風物志":"風物誌",
-"風范":"風範",
-"飄游":"飄遊",
-"飆卷":"飆捲",
-"飛升":"飛昇",
-"飛卷":"飛捲",
-"食欲":"食慾",
-"食面":"食麵",
-"飢年":"饑年",
-"飢荒":"饑荒",
-"飢饉":"饑饉",
-"飩面":"飩麵",
-"飯團":"飯糰",
-"飲一":"飲一",
-"飲一壇":"飲一罈",
-"飲三":"飲三",
-"飲三壇":"飲三罈",
-"飲兩":"飲兩",
-"飲兩壇":"飲兩罈",
-"飲干":"飲乾",
-"飲白":"飲白",
-"飲白干":"飲白乾",
-"飲而":"飲而",
-"飲而干":"飲而乾",
-"餅干":"餅乾",
-"養發":"養髮",
-"餐台":"餐檯",
-"餛飩":"餛飩",
-"餛飩面":"餛飩麵",
-"饋贈":"餽贈",
-"首只":"首隻",
-"首周":"首週",
-"首發":"首發",
-"香港":"香港",
-"香港游":"香港遊",
-"香煙":"香菸",
-"馬僵":"馬韁",
-"馬夫":"馬伕",
-"馬夫人":"馬夫人",
-"馬斯":"馬斯",
-"馬斯特裡":"馬斯特里",
-"馬表":"馬錶",
-"馬裡":"馬里",
-"馬裡蘭":"馬里蘭",
-"駐扎":"駐紮",
-"騁凶":"騁兇",
-"騰格":"騰格",
-"騰格裡":"騰格里",
-"騷首":"騷首",
-"騷首弄姿":"搔首弄姿",
-"驚栗":"驚慄",
-"驚螫":"驚蟄",
-"驚贊":"驚讚",
-"驟松":"驟鬆",
-"骨壇":"骨罈",
-"骨干":"骨幹",
-"骨松":"骨鬆",
-"骨灰":"骨灰",
-"骨灰壇":"骨灰罈",
-"骼膊":"胳膊",
-"髒器":"臟器",
-"髒腑":"臟腑",
-"體征":"體徵",
-"體范":"體範",
-"高升":"高昇",
-"高干":"高幹",
-"高干脆":"高乾脆",
-"髭須":"髭鬚",
-"髯發":"髯髮",
-"髯胡":"髯鬍",
-"髯須":"髯鬚",
-"鬈發":"鬈髮",
-"鬢發":"鬢髮",
-"鬢須":"鬢鬚",
-"鬼谷":"鬼谷",
-"鬼谷子":"鬼谷子",
-"魁悟":"魁梧",
-"魂游":"魂遊",
-"魏征":"魏徵",
-"魚干":"魚乾",
-"魚干擾":"魚干擾",
-"魚松":"魚鬆",
-"鮮于":"鮮于",
-"鯨須":"鯨鬚",
-"鳳佔":"鳳占",
-"鳳凰":"鳳凰",
-"鳳凰於飛":"鳳凰于飛",
-"鳳斗":"鳳鬥",
-"鳳皇":"鳳皇",
-"鳳皇於蜚":"鳳皇于蜚",
-"鴻准":"鴻準",
-"鴻范":"鴻範",
-"鵝准":"鵝準",
-"鶴發":"鶴髮",
-"鸕慈":"鸕鶿",
-"鹵制":"滷製",
-"鹵味":"滷味",
-"鹵水":"滷水",
-"鹵汁":"滷汁",
-"鹵牛":"鹵牛",
-"鹵牛肉":"滷牛肉",
-"鹵肉":"滷肉",
-"鹵菜":"滷菜",
-"鹵蛋":"滷蛋",
-"鹵醬":"滷醬",
-"鹵雞":"滷雞",
-"鹵面":"滷麵",
-"鹵鵝":"滷鵝",
-"鹽余":"鹽餘",
-"鹽鹵":"鹽滷",
-"麥面":"麥麵",
-"麻病":"痲病",
-"麻瘋":"痲瘋",
-"麼丑":"麼醜",
-"麼像":"麼像",
-"麼像征":"麼象徵",
-"麼兒":"么兒",
-"麼准":"麼准",
-"麼准的":"麼準的",
-"麼女":"么女",
-"麼妹":"么妹",
-"麼子":"么子",
-"麼干":"麼幹",
-"麼弟":"么弟",
-"麼象":"麼像",
-"黃歷":"黃曆",
-"黃發":"黃髮",
-"黃發垂髫":"黃髮垂髫",
-"黃金":"黃金",
-"黃金周":"黃金週",
-"黑亮":"黑亮",
-"黑亮發":"黑亮髮",
-"黑發":"黑髮",
-"黑發暗":"黑發暗",
-"黑須":"黑鬚",
-"點干":"點幹",
-"點干吧":"點幹吧",
-"點松":"點鬆",
-"點象":"點像",
-"黴味":"霉味",
-"黴天":"霉天",
-"黴氣":"霉氣",
-"黴濕":"霉濕",
-"黴爛":"霉爛",
-"黴運":"霉運",
-"黴雨":"霉雨",
-"黴頭":"霉頭",
-"鼎邊":"鼎邊",
-"鼎邊剉":"鼎邊銼",
-"鼻准":"鼻準",
-"鼻梁":"鼻樑",
-"齊沖":"齊衝",
-"齧齒":"囓齒",
-"龍卷":"龍卷",
-"龍卷風":"龍捲風",
-"龍眼":"龍眼",
-"龍眼干":"龍眼乾",
-"龍鐘":"龍鍾",
-"龍須":"龍鬚",
-"！恩":"！嗯",
-"！象":"！像",
-"０余":"０餘",
-"０海":"０海",
-"０海裡":"０海里",
-"０裡":"０里",
-"披肩發":"披肩髮"
-};//);
