--- lib/tongwen/tongwen_table_ss2t.js
@@ -1,31 +1,0 @@
-var symbolS2T = {
-	"·" :"‧",
-	"―":"─",
-	"‖":"∥",
-	"‘":"『",
-	"’":"』",
-	"“":"「",
-	"”":"」",
-	"″":"〞",
-	"∏":"Π",
-	"∑":"Σ",
-	"∧":"︿",
-	"∨":"﹀",
-	"∶":"︰",
-	"≈":"≒",
-	"≤":"≦",
-	"≥":"≧",
-	"━":"─",
-	"┃":"│",
-	"┏":"┌",
-	"┓":"┐",
-	"┗":"└",
-	"┛":"┘",
-	"┣":"├",
-	"┫":"┤",
-	"┳":"┬",
-	"┻":"┴",
-	"╋":"┼",
-	"〖":"【",
-	"〗":"】"
-};
