--- lib/tongwen/tongwen_table_st2s.js
@@ -1,16 +1,0 @@
-var symbolT2S = {
-	"¯" :"ˉ",
-	"‥":"¨" ,
-	"‧":"·" ,
-	"‵":"｀",
-	"≒":"≈" ,
-	"≦":"≤" ,
-	"≧":"≥" ,
-	"╱":"／",
-	"╲":"＼",
-	"╴":"＿",
-	"「":"“",
-	"」":"”",
-	"『":"‘",
-	"』":"’"
-};
