+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "new-tongwentang-for-firefox",
+  "version": "1.4.0",
+  "description": "Firefox Add-on, converts web page between Simplified and Traditional Chinese.",
+  "main": "index.js",
+  "scripts": {
+    "build": "webpack",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/softcup/New-Tongwentang-for-Firefox"
+  },
+  "author": "softcup",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/softcup/New-Tongwentang-for-Firefox/issues"
+  },
+  "homepage": "https://github.com/softcup/New-Tongwentang-for-Firefox#readme",
+  "devDependencies": {
+    "copy-webpack-plugin": "^4.0.1",
+    "webpack": "^3.1.0"
+  }
+}
