+++ src/README.md
@@ -1,0 +1,15 @@
+# New-Tongwentang
+Firefox Add-on, converts web page between Simplified and Traditional Chinese.
+
+# License
+Released under [MIT/X11 License](https://opensource.org/licenses/mit-license.php)
+
+# Todos
+- modular
+  - es6
+  - webpack
+  - typescript
+- UI rework
+  - react / angular / vue
+- reactive
+- integrate sync with third party
