+++ src/background.js
