+++ src/lib/tongwen/tongwen_core.js
