+++ src/lib/tongwen/tongwen_table_pt2s.js
@@ -1,0 +1,90 @@
+//TongWen.addT2STable(
+var default_PT2S_table = {
+"专着":"专著",
+"亲着":"亲著",
+"卓着":"卓著",
+"原着":"原著",
+"发着":"发著",
+"合着":"合著",
+"名着":"名著",
+"吐着":"吐著",
+"土着":"土著",
+"大着":"大著",
+"巨着":"巨著",
+"干丝":"乾丝",
+"干乾":"康乾",
+"干元":"乾元",
+"干卦":"乾卦",
+"干县":"乾县",
+"干图":"乾图",
+"干圖":"乾图",
+"干坤":"乾坤",
+"干宅":"乾宅",
+"干断":"乾断",
+"干旦":"乾旦",
+"干曜":"乾曜",
+"干清":"干清",
+"干清宮":"乾清宫",
+"干盛":"干盛",
+"干盛世":"乾盛世",
+"干着":"干着",
+"干着急":"乾着急",
+"干紅":"乾红",
+"干綱":"乾纲",
+"干红":"乾紅",
+"干纲":"乾纲",
+"干象":"乾象",
+"干陵":"乾陵",
+"干隆":"乾隆",
+"張法":"張法",
+"張法干":"张法乾",
+"彰明":"彰明",
+"彰明较着":"彰明较著",
+"所着":"所著",
+"所着手":"所着手",
+"批着":"批著",
+"撰着":"撰著",
+"新着":"新著",
+"旧着":"旧著",
+"昭着":"昭著",
+"显着":"显著",
+"显着性":"显著性",
+"洋泾":"洋泾",
+"洋泾滨":"洋泾浜",
+"着书":"著书",
+"着作":"著作",
+"着名":"著名",
+"着庸":"著庸",
+"着录":"著录",
+"着明":"著明",
+"着有":"著有",
+"着毋":"着毋",
+"着毋庸议":"著毋庸议",
+"着称":"著称",
+"着籍":"著籍",
+"着绩":"著绩",
+"着者":"著者",
+"着述":"著述",
+"着迷":"着迷",
+"纨裤":"纨绔",
+"纨裤子弟":"纨绔子弟",
+"绮襦":"绮襦",
+"绮襦纨裤":"绮襦纨绔",
+"编着":"编著",
+"萧干":"萧乾",
+"蕭干":"萧乾",
+"裤下":"裤下",
+"裤下之辱":"绔下之辱",
+"见微":"见微",
+"见微知着":"见微知著",
+"计画":"计划",
+"论着":"论著",
+"译着":"译著",
+"象征":"象征",
+"象征着":"象徵着",
+"遗着":"遗著",
+"郭子":"郭子",
+"郭子干":"郭子乾",
+"颇着":"颇著",
+"鲇鱼":"鲶鱼"
+};//);
