+++ src/lib/tongwen/tongwen_table_ss2t.js
@@ -1,0 +1,31 @@
+var symbolS2T = {
+	"·" :"‧",
+	"―":"─",
+	"‖":"∥",
+	"‘":"『",
+	"’":"』",
+	"“":"「",
+	"”":"」",
+	"″":"〞",
+	"∏":"Π",
+	"∑":"Σ",
+	"∧":"︿",
+	"∨":"﹀",
+	"∶":"︰",
+	"≈":"≒",
+	"≤":"≦",
+	"≥":"≧",
+	"━":"─",
+	"┃":"│",
+	"┏":"┌",
+	"┓":"┐",
+	"┗":"└",
+	"┛":"┘",
+	"┣":"├",
+	"┫":"┤",
+	"┳":"┬",
+	"┻":"┴",
+	"╋":"┼",
+	"〖":"【",
+	"〗":"】"
+};
