+++ src/lib/tongwen/tongwen_table_st2s.js
@@ -1,0 +1,16 @@
+var symbolT2S = {
+	"¯" :"ˉ",
+	"‥":"¨" ,
+	"‧":"·" ,
+	"‵":"｀",
+	"≒":"≈" ,
+	"≦":"≤" ,
+	"≧":"≥" ,
+	"╱":"／",
+	"╲":"＼",
+	"╴":"＿",
+	"「":"“",
+	"」":"”",
+	"『":"‘",
+	"』":"’"
+};
